### 2020-05-23 #2243 @whedon
[PRE REVIEW]: PypeIt: The Python Spectroscopic Data Reduction Pipeline 
**Submitting author:** @profxj (<a href="http://orcid.org/0000-0002-7738-6875">J. Xavier Prochaska</a>)
**Repository:** <a href="https://github.com/pypeit/PypeIt" target ="_blank">https://github.com/pypeit/PypeIt</a>
**Version:** v1.0.3
