### 2020-05-23 #2243 @arfon
@profxj - thanks for your submission to JOSS. Does this submission accompany a AAS paper?
