### 2020-05-23 #2243 @profxj
It does not.

We may do an AAS Research Note, which is not peer reviewed.

Thanks

X
