### 2020-05-24 f18543c
minor logic
