### 2020-05-24 aab315e
Addressing my own comments to this PR, since someone refused to make changes.
