+++ pypeit/core/extract.py
@@ -2138,7 +2138,7 @@ def ech_objfind(image, ivar, slitmask, slit_left, slit_righ, order_vec, maskslit
     sobjs_trim = specobjs.SpecObjs()
     # objids are 1 based so that we can easily asign the negative to negative objects
     iobj_keep = 1
-    # TODO JFH: Fix this ugly and dangerous hack that was added to accomodate hand apertures 
+    # TODO JFH: Fix this ugly and dangerous hack that was added to accomodate hand apertures
     hand_frac = [-1000] if hand_extract_dict is None else [int(np.round(ispat*1000)) for ispat in f_spats]
     for iobj in range(nobj):
         if (SNR_arr[:,iobj].max() > max_snr) or (np.sum(SNR_arr[:,iobj] > min_snr) >= nabove_min_snr) \
