### 2020-05-24 f60cb6e
add MMT MMIRS spectrograph
