+++ pypeit/spectrographs/mmt_mmirs.py
@@ -1,0 +1,271 @@
+"""
+Module for MMT MMIRS
+"""
+import glob
+import numpy as np
+from astropy.time import Time
+from astropy.io import fits
+
+from pypeit import msgs
+from pypeit import telescopes
+from pypeit.core import parse
+from pypeit.core import framematch
+from pypeit.par import pypeitpar
+from pypeit.images import detector_container
+from pypeit.spectrographs import spectrograph
+from pkg_resources import resource_filename
+
+
+class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
+    """
+    Child to handle MMT/MMIRS specific code
+
+
+    """
+    ndet = 1
+
+    def __init__(self):
+        # Get it started
+        super(MMTMMIRSSpectrograph, self).__init__()
+        self.spectrograph = 'mmt_mmirs'
+        self.telescope = telescopes.MMTTelescopePar()
+
+    @staticmethod
+    def default_pypeit_par():
+        """
+        Set default parameters for the reductions.
+        """
+        par = pypeitpar.PypeItPar()
+        return par
+
+    def init_meta(self):
+        """
+        Generate the meta data dict
+        Note that the children can add to this
+
+        Returns:
+            self.meta: dict (generated in place)
+
+        """
+        meta = {}
+        # Required (core)
+        meta['ra'] = dict(ext=1, card='RA')
+        meta['dec'] = dict(ext=1, card='DEC')
+        meta['target'] = dict(ext=1, card='OBJECT')
+        meta['decker'] = dict(ext=1, card='APERTURE')
+        meta['dichroic'] = dict(ext=1, card='FILTER')
+        meta['binning'] = dict(ext=1, card=None, default='1,1')
+
+        meta['mjd'] = dict(ext=0, card=None, compound=True)
+        meta['exptime'] = dict(ext=1, card='EXPTIME')
+        meta['airmass'] = dict(ext=1, card='AIRMASS')
+        # Extras for config and frametyping
+        meta['dispname'] = dict(ext=1, card='DISPERSE')
+        meta['idname'] = dict(ext=1, card='IMAGETYP')
+
+        # Ingest
+        self.meta = meta
+
+    def compound_meta(self, headarr, meta_key):
+        if meta_key == 'mjd':
+            time = headarr[1]['DATE-OBS']
+            ttime = Time(time, format='isot')
+            return ttime.mjd
+        else:
+            msgs.error("Not ready for this compound meta")
+
+    def get_detector_par(self, hdu, det):
+        """
+        Return a DectectorContainer for the current image
+
+        Args:
+            hdu (`astropy.io.fits.HDUList`):
+                HDUList of the image of interest.
+                Ought to be the raw file, or else..
+            det (int):
+
+        Returns:
+            :class:`pypeit.images.detector_container.DetectorContainer`:
+
+        """
+        # Detector 1
+        detector_dict = dict(
+            binning='1,1',
+            det             = 1,
+            dataext         = 1,
+            specaxis        = 0,
+            specflip        = False,
+            spatflip        = False,
+            platescale      = 0.2012,
+            darkcurr        = 0.01,
+            saturation      = 700000., #155400.,
+            nonlinear       = 1.0,
+            mincounts       = -1e10,
+            numamplifiers   = 1,
+            gain            = np.atleast_1d(0.95),
+            ronoise         = np.atleast_1d(3.140000),
+            datasec         = np.atleast_1d('[:,:]'),
+            oscansec        = np.atleast_1d('[:,:]')
+            )
+        return detector_container.DetectorContainer(**detector_dict)
+
+    def default_pypeit_par(self):
+        """
[…]
