+++ pypeit/defs.py
@@ -10,4 +10,4 @@ pypeit_spectrographs = ['keck_deimos', 'keck_lris_blue', 'keck_lris_red', 'keck_
             'gemini_gmos_north_e2v', 'gemini_gmos_north_ham',
             'magellan_fire', 'magellan_fire_long', 'magellan_mage',
             'lbt_mods1r', 'lbt_mods1b', 'lbt_mods2r', 'lbt_mods2b', 'lbt_luci1', 'lbt_luci2',
-            'mmt_binospec', 'mdm_osmos_mdm4k', 'not_alfosc']
+            'mmt_binospec', 'mmt_mmirs', 'mdm_osmos_mdm4k', 'not_alfosc']
