+++ pypeit/scripts/view_fits.py
@@ -38,6 +38,7 @@ def main(args):
     from pypeit.spectrographs import keck_deimos
     from pypeit.spectrographs import gemini_gmos
     from pypeit.spectrographs import mmt_binospec
+    from pypeit.spectrographs import mmt_mmirs
     from pypeit import msgs
     from pypeit import ginga
 
@@ -72,6 +73,11 @@ def main(args):
         #
         gen_bino = mmt_binospec.MMTBINOSPECSpectrograph()
         img = gen_bino.get_rawimage(args.file, args.det)[1]
+    # RAW_MMIRS
+    elif args.spectrograph == 'mmt_mmirs':
+        #
+        gen_mmirs = mmt_mmirs.MMTMMIRSSpectrograph()
+        img = gen_mmirs.get_rawimage(args.file, args.det)[1]
     else:
         hdu = fits.open(args.file)
         img = hdu[args.exten].data
