### 2020-05-24 8c6f166
Bugfix in robust_polyfit_djs, when arrays are not 1d code crashes if inmask not specified.
