+++ pypeit/core/fitting.py
@@ -1033,3 +1033,147 @@ def scale_minmax(x, minx=None, maxx=None):
     return xv
 
 
+# TODO: JFH This routine should be put in the bspline module as a utility function and the bspline class should be renamed to Bspline
+# but I'm confused by the __init__ in that directory.
+def iterfit(xdata, ydata, invvar=None, inmask=None, upper=5, lower=5, x2=None,
+            maxiter=10, nord=4, bkpt=None, fullbkpt=None, kwargs_bspline={}, kwargs_reject={}):
+        """Iteratively fit a b-spline set to data, with rejection. This is a utility function that allows
+        the bspline to use via a direct function call.
+
+        Parameters
+        ----------
+        xdata : :class:`numpy.ndarray`
+            Independent variable.
+        ydata : :class:`numpy.ndarray`
+            Dependent variable.
+        invvar : :class:`numpy.ndarray`
+            Inverse variance of `ydata`.  If not set, it will be calculated based
+            on the standard deviation.
+        upper : :class:`int` or :class:`float`
+            Upper rejection threshold in units of sigma, defaults to 5 sigma.
+        lower : :class:`int` or :class:`float`
+            Lower rejection threshold in units of sigma, defaults to 5 sigma.
+        x2 : :class:`numpy.ndarray`, optional
+            Orthogonal dependent variable for 2d fits.
+        maxiter : :class:`int`, optional
+            Maximum number of rejection iterations, default 10.  Set this to
+            zero to disable rejection.
+
+        Returns
+        -------
+        :func:`tuple`
+            A tuple containing the fitted bspline object and an output mask.
+        """
+        # from .math import djs_reject
+        nx = xdata.size
+        if ydata.size != nx:
+            raise ValueError('Dimensions of xdata and ydata do not agree.')
+        if invvar is not None:
+            if invvar.size != nx:
+                raise ValueError('Dimensions of xdata and invvar do not agree.')
+        else:
+            #
+            # This correction to the variance makes it the same
+            # as IDL's variance()
+            #
+            var = ydata.var() * (float(nx) / float(nx - 1))
+            if var == 0:
+                var = 1.0
+            invvar = np.ones(ydata.shape, dtype=ydata.dtype) / var
+
+        if inmask is None:
+            inmask = invvar > 0.0
+
+        if x2 is not None:
+            if x2.size != nx:
+                raise ValueError('Dimensions of xdata and x2 do not agree.')
+        yfit = np.zeros(ydata.shape)
+        if invvar.size == 1:
+            outmask = True
+        else:
+            outmask = np.ones(invvar.shape, dtype='bool')
+        xsort = xdata.argsort()
+        maskwork = (outmask & inmask & (invvar > 0.0))[xsort]
+        if 'oldset' in kwargs_bspline:
+            sset = kwargs_bspline['oldset']
+            sset.mask = True
+            sset.coeff = 0
+        else:
+            if not maskwork.any():
+                raise ValueError('No valid data points.')
+                # return (None,None)
+            # JFH comment this out for now
+            #        if 'fullbkpt' in kwargs:
+            #            fullbkpt = kwargs['fullbkpt']
+            else:
+                sset = bspline.bspline(xdata[xsort[maskwork]], nord=nord, bkpt=bkpt, fullbkpt=fullbkpt, **kwargs_bspline)
+                if maskwork.sum() < sset.nord:
+                    print('Number of good data points fewer than nord.')
+                    return (sset, outmask)
+                if x2 is not None:
+                    if 'xmin' in kwargs_bspline:
+                        xmin = kwargs_bspline['xmin']
+                    else:
+                        xmin = x2.min()
+                    if 'xmax' in kwargs_bspline:
+                        xmax = kwargs_bspline['xmax']
[…]
