+++ pypeit/core/flexure.py
@@ -245,15 +245,23 @@ def spec_flex_shift(obj_skyspec, arx_skyspec, mxshft=20):
     # Deal with underlying continuum
     msgs.work("Consider taking median first [5 pixel]")
     everyn = obj_skyspec.npix // 20
-    bspline_par = dict(everyn=everyn)
-    # TODO - Replace with the other bspline fitter as this no longer works!
-    pypeitFit_obj = fitting.robust_fit(obj_skyspec.wavelength.value, obj_skyspec.flux.value, 3,
-                                    function='bspline', lower=3., upper=3., bspline_par=bspline_par)
-    obj_sky_cont = pypeitFit_obj.val(obj_skyspec.wavelength.value)
+    #bspline_par = dict(everyn=everyn)
+    #pypeitFit_obj = fitting.robust_fit(obj_skyspec.wavelength.value, obj_skyspec.flux.value, 3,
+    #                                function='bspline', lower=3., upper=3., bspline_par=bspline_par)
+    #obj_sky_cont = pypeitFit_obj.val(obj_skyspec.wavelength.value)
+    pypeitFit_obj, _ = fitting.iterfit(obj_skyspec.wavelength.value, obj_skyspec.flux.value,
+                                       nord = 3,  kwargs_bspline={'everyn': everyn}, kwargs_reject={'groupbadpix':True,'maxrej':1},
+                                       maxiter = 15, upper = 3.0, lower = 3.0)
+    obj_sky_cont = pypeitFit_obj.value(obj_skyspec.wavelength.value)
+
     obj_sky_flux = obj_skyspec.flux.value - obj_sky_cont
-    pypeitFit_sky= fitting.robust_fit(arx_skyspec.wavelength.value, arx_skyspec.flux.value, 3,
-                                        function='bspline', lower=3., upper=3., bspline_par=bspline_par)
-    arx_sky_cont = pypeitFit_sky.val(arx_skyspec.wavelength.value)
+    pypeitFit_sky, _ = fitting.iterfit(arx_skyspec.wavelength.value, arx_skyspec.flux.value,
+                                       nord = 3,  kwargs_bspline={'everyn': everyn}, kwargs_reject={'groupbadpix':True,'maxrej':1},
+                                       maxiter = 15, upper = 3.0, lower = 3.0)
+    arx_sky_cont = pypeitFit_sky.value(arx_skyspec.wavelength.value)
+    #pypeitFit_sky= fitting.robust_fit(arx_skyspec.wavelength.value, arx_skyspec.flux.value, 3,
+    #                                    function='bspline', lower=3., upper=3., bspline_par=bspline_par)
+    #arx_sky_cont = pypeitFit_sky.val(arx_skyspec.wavelength.value)
     arx_sky_flux = arx_skyspec.flux.value - arx_sky_cont
 
     # Consider sharpness filtering (e.g. LowRedux)
