### 2020-05-25 3ed7be8
Implementing some of my comments and fixing the massive bug that was introduced.
