### 2020-05-25 3ddb290
Fixing the massive number of bugs introduced by the wvcalib_refactor PR.
