+++ pypeit/spectrographs/keck_mosfire.py
@@ -67,7 +67,7 @@ class KeckMOSFIRESpectrograph(spectrograph.Spectrograph):
         par['rdx']['spectrograph'] = 'keck_mosfire'
         # Wavelengths
         # 1D wavelength solution
-        par['calibrations']['wavelengths']['rms_threshold'] = 0.20 #0.20  # Might be grating dependent..
+        par['calibrations']['wavelengths']['rms_threshold'] = 0.30 #0.20  # Might be grating dependent..
         par['calibrations']['wavelengths']['sigdetect']=5.0
         par['calibrations']['wavelengths']['fwhm']= 5.0
         par['calibrations']['wavelengths']['n_final']= 4
