+++ pypeit/flatfield.py
@@ -222,10 +222,8 @@ class FlatField(object):
     For the primary methods, see :func:`run`.
 
     Args:
-        rawpixflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`):
+        rawflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`):
             Processed, combined set of pixelflat images
-        rawillumflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`, None):
-            Processed, combined set of illumflat images
         spectrograph (:class:`pypeit.spectrographs.spectrograph.Spectrograph`):
             The `Spectrograph` instance that sets the instrument used to
             take the observations.  See usage by
@@ -237,10 +235,16 @@ class FlatField(object):
             The current slit traces.
         wavetilts (:class:`pypeit.wavetilts.WaveTilts`):
             The current wavelength tilt traces; see
+        spat_illum_only (boolean):
+            Only perform the spatial illumination calculation, and ignore
+            the 2D bspline fit. This should only be set to true if you
+            want the spatial illumination profile only. If you want to
+            simultaneously generate a pixel flat and a spatial
+            illumination profile from the same input, this should be
+            False (which is the default).
 
     Attributes:
-        rawpixflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`):
-        rawillumflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`, None):
+        rawflatimg (:class:`pypeit.images.pypeitimage.PypeItImage`):
         mspixelflat (`numpy.ndarray`_):
             Normalized flat
         msillumflat (`numpy.ndarray`_):
@@ -256,7 +260,7 @@ class FlatField(object):
     master_type = 'Flat'
 
 
-    def __init__(self, rawpixflatimg, rawillumflatimg, spectrograph, flatpar, slits, wavetilts):
+    def __init__(self, rawflatimg, spectrograph, flatpar, slits, wavetilts, spat_illum_only=False):
 
         # Defaults
         self.spectrograph = spectrograph
@@ -271,12 +275,12 @@ class FlatField(object):
         self.wavetilts.is_synced(self.slits)
 
         # Attributes unique to this Object
-        self.rawpixflatimg = rawpixflatimg      # Un-normalized pixel flat as a PypeItImage
-        self.rawillumflatimg = rawillumflatimg      # Un-normalized illumination flat as a PypeItImage
+        self.rawflatimg = rawflatimg      # Un-normalized pixel flat as a PypeItImage
         self.mspixelflat = None     # Normalized pixel flat
         self.msillumflat = None     # Illumination flat
         self.flat_model = None      # Model flat
         self.list_of_spat_bsplines = None
+        self.spat_illum_only = spat_illum_only
 
         # Completed steps
         self.steps = []
@@ -362,14 +366,8 @@ class FlatField(object):
         #self.build_pixflat()
 
         # Fit it
-        # NOTE: Tilts do not change and self.slits is updated
-        # internally.
-        if self.rawillumflatimg is not None:
-            msgs.info("Generating illumination profile from master illumflat")
-            # Generate the illumination profile first
-            self.fit(illumflat=True, debug=debug)
-        # Generate the pixel flat (and illumination flat, if not already made)
-        self.fit(debug=debug)
+        # NOTE: Tilts do not change and self.slits is updated internally.
+        self.fit(illumflat=self.spat_illum_only, debug=debug)
 
         if show:
             # Global skysub is the first step in a new extraction so clear the channels here
@@ -383,7 +381,7 @@ class FlatField(object):
                 bpmflats[bpm] = self.slits.bitmask.turn_on(bpmflats[bpm], flag)
 
         # Return
[…]
