### 2020-05-25 d9fe543
MOSFIRE Ql is alive.
