+++ pypeit/scripts/ql_keck_mosfire.py
@@ -9,16 +9,39 @@ This script runs PypeIt on a pair of NIRES images (A-B)
 import argparse
 
 from pypeit import msgs
+import os
+import sys
+import numpy as np
+import copy
+from astropy.io import fits
+from IPython import embed
 
+from pypeit import pypeit
+from pypeit import par, msgs
+from pypeit import pypeitsetup
+from pypeit import wavecalib
+from pypeit import wavetilts
+from pypeit import spec2dobj
+from pypeit.core import framematch
+from pypeit import specobjs
+from pypeit.images import buildimage
+from pypeit import slittrace
+from pypeit.spectrographs.util import load_spectrograph
+from pypeit import reduce
+from pypeit import calibrations
 import warnings
 
 def parser(options=None):
 
-    parser = argparse.ArgumentParser(description='Script to run PypeIt on a pair of NIRES files (A-B)')
+    parser = argparse.ArgumentParser(description='Script to run PypeIt on a pair of MOSFIRE files (A-B)')
     parser.add_argument('full_rawpath', type=str, help='Full path to the raw files')
     parser.add_argument('fileA', type=str, help='A frame')
     parser.add_argument('fileB', type=str, help='B frame')
     parser.add_argument('-b', '--box_radius', type=float, help='Set the radius for the boxcar extraction')
+    parser.add_argument("--redux_path", type=str, default=os.getcwd(),
+                        help="Location where reduction outputs should be stored.")
+    parser.add_argument("--show", default=False, action="store_true",
+                        help="Show the reduction steps. Equivalent to the -s option when running pypeit.")
 
     if options is None:
         pargs = parser.parse_args()
@@ -26,24 +49,202 @@ def parser(options=None):
         pargs = parser.parse_args(options)
     return pargs
 
+def config_lines(pargs):
+
+    # Config the run
+    cfg_lines = ['[rdx]']
+    cfg_lines += ['    spectrograph = {0}'.format('keck_mosfire')]
+    cfg_lines += ['    redux_path = {0}'.format(pargs.redux_path)]
+    cfg_lines += ['    scidir = Science_QL']
+    # Calibrations
+    cfg_lines += ['[baseprocess]']
+    cfg_lines += ['    use_pixelflat = False']
+    cfg_lines += ['    use_illumflat = False']
+    cfg_lines += ['[calibrations]']
+    cfg_lines += ['    [[wavelengths]]']
+    cfg_lines += ['        frame = observed']
+    cfg_lines += ['[scienceframe]']
+    cfg_lines += ['    [[process]]']
+    cfg_lines += ['        mask_cr = False']
+    cfg_lines += ['[reduce]']
+    cfg_lines += ['    [[extraction]]']
+    cfg_lines += ['        skip_optimal = True']
+    if pargs.box_radius is not None: # Boxcar radius
+        cfg_lines += ['        boxcar_radius = {0}'.format(pargs.box_radius)]
+    cfg_lines += ['    [[findobj]]']
+    cfg_lines += ['        skip_second_find = True']
+
+    return cfg_lines
+
+
+def save_exposure(fitstbl, frame, spectrograph, science_path, par, caliBrate, all_spec2d, all_specobjs):
+    """
+    Save the outputs from extraction for a given exposure
+
+    Args:
+        frame (:obj:`int`):
+            0-indexed row in the metadata table with the frame
+            that has been reduced.
+        all_spec2d(:class:`pypeit.spec2dobj.AllSpec2DObj`):
+        sci_dict (:obj:`dict`):
+            Dictionary containing the primary outputs of
+            extraction
+        basename (:obj:`str`):
[…]
