+++ pypeit/spectrographs/mmt_mmirs.py
@@ -3,8 +3,10 @@ Module for MMT MMIRS
 """
 import glob
 import numpy as np
+from scipy.signal import savgol_filter
 from astropy.time import Time
 from astropy.io import fits
+from astropy.stats import sigma_clipped_stats
 
 from pypeit import msgs
 from pypeit import telescopes
@@ -103,7 +105,7 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
             mincounts       = -1e10,
             numamplifiers   = 1,
             gain            = np.atleast_1d(0.95),
-            ronoise         = np.atleast_1d(3.140000),
+            ronoise         = np.atleast_1d(3.14),
             datasec         = np.atleast_1d('[:,:]'),
             oscansec        = np.atleast_1d('[:,:]')
             )
@@ -124,7 +126,7 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
         #par['calibrations']['pixelflatframe']['process']['use_darkimage'] = True
         #par['calibrations']['illumflatframe']['process']['use_darkimage'] = True
         #par['scienceframe']['process']['use_darkimage'] = True
-        #par['scienceframe']['process']['use_illumflat'] = True
+        par['scienceframe']['process']['use_illumflat'] = True
 
         # Wavelengths
         # 1D wavelength solution with arc lines
@@ -151,7 +153,15 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
         par['calibrations']['darkframe']['exprng'] = [30, None]
         par['scienceframe']['exprng'] = [30, None]
 
-        # Scienceimage parameters
+        # dark
+        par['calibrations']['darkframe']['process']['apply_gain'] = True
+
+        # cosmic ray rejection
+        par['scienceframe']['process']['sigclip'] = 5.0
+        par['scienceframe']['process']['objlim'] = 2.0
+        par['scienceframe']['process']['grow'] = 0.5
+
+        # Science reduction
         par['reduce']['findobj']['sig_thresh'] = 5.0
         par['reduce']['skysub']['sky_sigrej'] = 5.0
         par['reduce']['findobj']['find_trim_edge'] = [5,5]
@@ -205,7 +215,7 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
         if ftype in ['pinhole', 'bias']:
             # No pinhole or bias frames
             return np.zeros(len(fitstbl), dtype=bool)
-        if ftype in ['pixelflat', 'trace']:
+        if ftype in ['pixelflat', 'trace', 'illumflat']:
             return good_exp & (fitstbl['idname'] == 'flat')
         if ftype == 'standard':
             return good_exp & (fitstbl['idname'] == 'object')
@@ -217,6 +227,40 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
             return good_exp & (fitstbl['idname'] == 'dark')
         msgs.warn('Cannot determine if frames are of type {0}.'.format(ftype))
         return np.zeros(len(fitstbl), dtype=bool)
+
+    def bpm(self, filename, det, shape=None, msbias=None):
+        """ Generate a BPM
+
+        Parameters
+        ----------
+        det : int, REQUIRED
+        **null_kwargs:
+           Captured and never used
+
+        Returns
+        -------
+        badpix : ndarray
+
+        """
+        # Get the empty bpm: force is always True
+        bpm_img = self.empty_bpm(filename, det, shape=shape)
+
+        # Fill in bad pixels if a master bias frame is provided
+        if msbias is not None:
[…]
