+++ pypeit/scripts/ql_keck_mosfire.py
@@ -22,6 +22,7 @@ from pypeit import pypeitsetup
 from pypeit import wavecalib
 from pypeit import wavetilts
 from pypeit import spec2dobj
+from pypeit import coadd2d
 from pypeit.core import framematch
 from pypeit import specobjs
 from pypeit.images import buildimage
@@ -38,6 +39,8 @@ def parser(options=None):
     parser.add_argument('fileA', type=str, help='A frame')
     parser.add_argument('fileB', type=str, help='B frame')
     parser.add_argument('-b', '--box_radius', type=float, help='Set the radius for the boxcar extraction')
+    parser.add_argument('--samp_fact', default=1.0, type=float,
+                        help="Make the wavelength grid finer (samp_fact > 1.0) or coarser (samp_fact < 1.0) by this sampling factor")
     parser.add_argument("--redux_path", type=str, default=os.getcwd(),
                         help="Location where reduction outputs should be stored.")
     parser.add_argument("--show", default=False, action="store_true",
@@ -122,6 +125,8 @@ def save_exposure(fitstbl, frame, spectrograph, science_path, par, caliBrate, al
         # Info
         outfiletxt = os.path.join(science_path, 'spec1d_{:s}.txt'.format(basename))
         all_specobjs.write_info(outfiletxt, spectrograph.pypeline)
+    else:
+        outfile1d = None
 
     # 2D spectra
     outfile2d = os.path.join(science_path, 'spec2d_{:s}.fits'.format(basename))
@@ -133,16 +138,35 @@ def save_exposure(fitstbl, frame, spectrograph, science_path, par, caliBrate, al
                                            subheader=subheader)
     # Write
     all_spec2d.write_to_fits(outfile2d, pri_hdr=pri_hdr, update_det=par['rdx']['detnum'])
+    return outfile2d, outfile1d
+
+def parse_dither_pattern(file_list, ext):
+
+    nfiles = len(file_list)
+    offset_arcsec = np.zeros(nfiles)
+    dither_pattern = []
+    dither_id = []
+    for ifile, file in enumerate(file_list):
+        hdr = fits.getheader(file, ext)
+        dither_pattern.append(hdr['PATTERN'])
+        dither_id.append(hdr['FRAMEID'])
+        offset_arcsec[ifile] = hdr['YOFFSET']
+
+    return dither_pattern, dither_id, offset_arcsec
 
 
 
 def main(pargs):
 
     # Build the fitstable since we currently need it for output. This should not be the case!
-    data_files = [os.path.join(pargs.full_rawpath, pargs.fileA), os.path.join(pargs.full_rawpath,pargs.fileB)]
-    ps = pypeitsetup.PypeItSetup([data_files[0]], path='./', spectrograph_name='keck_mosfire')
+    A_files = [os.path.join(pargs.full_rawpath, pargs.fileA)]
+    B_files = [os.path.join(pargs.full_rawpath, pargs.fileB)]
+    data_files = A_files + B_files
+    ps = pypeitsetup.PypeItSetup(A_files, path='./', spectrograph_name='keck_mosfire')
     ps.build_fitstbl()
     fitstbl = ps.fitstbl
+
+
 
     # Read in the spectrograph, config the parset
     spectrograph = load_spectrograph('keck_mosfire')
@@ -221,7 +245,6 @@ def main(pargs):
     for sobj in sobjs:
         sobj.DETECTOR = sciImg.detector
 
-
     # Construct the Spec2DObj
     spec2DObj = spec2dobj.Spec2DObj(det=det,
                                     sciimg=sciImg.image,
[…]
