### 2020-05-26 2c58fe1
Fixed error in MOSFIRE plate scale.
