### 2020-05-26 d30bbd1
add arxiv wave solution for MMIRS HK+zJ
