+++ pypeit/coadd2d.py
@@ -43,7 +43,7 @@ class CoAdd2D(object):
     # Superclass factory method generates the subclass instance
     @classmethod
     def get_instance(cls, spec2dfiles, spectrograph, par, det=1, offsets=None, weights='auto',
-                     sn_smooth_npix=None, ir_redux=False, master_dir=None, show=False,
+                     sn_smooth_npix=None, ir_redux=False, show=False,
                      show_peaks=False, debug_offsets=False, debug=False, **kwargs_wave):
         """
         Instantiate the CoAdd2d subclass appropriate for the provided spectrograph.
@@ -63,19 +63,19 @@ class CoAdd2D(object):
         return next(c for c in cls.__subclasses__() 
                     if c.__name__ == (spectrograph.pypeline + 'CoAdd2D'))(
                         spec2dfiles, spectrograph, par, det=det, offsets=offsets, weights=weights,
-                        sn_smooth_npix=sn_smooth_npix, ir_redux=ir_redux, master_dir=master_dir,
+                        sn_smooth_npix=sn_smooth_npix, ir_redux=ir_redux,
                         show=show, show_peaks=show_peaks, debug_offsets=debug_offsets, debug=debug,
                         **kwargs_wave)
 
-    def __init__(self, spec2d_files, spectrograph, par, det=1, offsets=None, weights='auto',
-                 sn_smooth_npix=None, ir_redux=False, master_dir=None, show=False,
+    def __init__(self, spec2d, spectrograph, par, det=1, offsets=None, weights='auto',
+                 sn_smooth_npix=None, ir_redux=False, show=False,
                  show_peaks=False, debug_offsets=False, debug=False, **kwargs_wave):
         """
 
         Parameters
         ----------
             spec2d_files (list):
-               List of spec2d files
+               List of spec2d files or a list of Spec2dObj objects
             spectrograph (:class:`pypeit.spectrographs.spectrograph.Spectrograph`):
                 The instrument used to collect the data to be reduced.
             par (:class:`pypeit.par.parset.ParSet`):
@@ -96,9 +96,6 @@ class CoAdd2D(object):
             ir_redux (bool): optional, default=False
                 Is this an near-IR reduction, True=yes. This parameter is passed to pypeit.reduce for determining the
                 reduction steps.
-            master_dir (:obj:`str`, optional):
-                Directory for the coadding master frames. If None,
-                set by ``os.getcwd()``.
             show (bool): optional, default=False
                 Show results to ginga
             show_peaks (bool): optional, default=False
@@ -119,14 +116,14 @@ class CoAdd2D(object):
         if offsets is not None and 'auto' in weights:
             msgs.error("Automatic weights cannot be computed for input offsets. "
                        "Set weights='uniform' or input an array of weights with shape (nexp,)")
-        self.spec2d_files = spec2d_files
+        self.spec2d = spec2d
         self.spectrograph = spectrograph
         self.par = par
         self.det = det
         self.offsets = offsets
         self.weights = weights
         self.ir_redux = ir_redux
-        self.master_dir = os.getcwd() if master_dir is None else master_dir
+        #self.master_dir = os.getcwd() if master_dir is None else master_dir
         self.show = show
         self.show_peaks = show_peaks
         self.debug_offsets = debug_offsets
@@ -140,7 +137,7 @@ class CoAdd2D(object):
 
 
         # Load the stack_dict
[…]
