+++ pypeit/scripts/ql_keck_mosfire.py
@@ -35,15 +35,31 @@ from pypeit.core.parse import get_dnum
 from astropy.stats import sigma_clipped_stats
 import warnings
 
+
+
+# A trick from stackoverflow to allow multi-line output in the help:
+#https://stackoverflow.com/questions/3853722/python-argparse-how-to-insert-newline-in-the-help-text
+class SmartFormatter(argparse.HelpFormatter):
+
+    def _split_lines(self, text, width):
+        if text.startswith('R|'):
+            return text[2:].splitlines()
+        # this is the RawTextHelpFormatter._split_lines
+        return argparse.HelpFormatter._split_lines(self, text, width)
+
 def parser(options=None):
 
-    parser = argparse.ArgumentParser(description='Script to run PypeIt on a pair of MOSFIRE files (A-B)')
+    parser = argparse.ArgumentParser(description='Script to run PypeIt on a pair of MOSFIRE files (A-B)', formatter_class=SmartFormatter)
     parser.add_argument('full_rawpath', type=str, help='Full path to the raw files')
     parser.add_argument('-A','--Afiles', type=str, nargs='+', help='list of frames at dither position A, i.e. -A A1.fits A2.fits')
     parser.add_argument('-B','--Bfiles', type=str, nargs='+', help='list of frames at dither position B  i.e. -B B1.fits B2.fits')
     parser.add_argument('--samp_fact', default=1.0, type=float,
                         help="Make the wavelength grid finer (samp_fact > 1.0) or coarser (samp_fact < 1.0) by this sampling factor")
     parser.add_argument('--box_radius', type=float, help='Set the radius for the boxcar extraction')
+    parser.add_argument('--offset', type=float, default=None,
+                        help='R|Override the automatic offsets determined from the headers. Offset is in pixels.\n'
+                        'This option is useful if a standard dither pattern was not executed.\n'
+                        'The offset convention is such that a negative offset will move the (negative) B image to the left')
     parser.add_argument("--redux_path", type=str, default=os.getcwd(),
                         help="Location where reduction outputs should be stored.")
     parser.add_argument('--embed', default=False, help='Upon completion embed in ipython shell',
@@ -188,8 +204,8 @@ def main(pargs):
     slit_masterframe_name = os.path.join(master_dir, 'MasterSlits_A_3_01.fits.gz')
     tilts_masterframe_name = os.path.join(master_dir, 'MasterTilts_A_2_01.fits')
     wvcalib_masterframe_name = os.path.join(master_dir, 'MasterWaveCalib_A_2_01.fits')
-    std_outfile =os.path.join('/Users/joe/Dropbox/PypeIt_Redux/MOSFIRE/Nov19/quicklook/Science/',
-                              'spec1d_m191118_0064-GD71_MOSFIRE_2019Nov18T104704.507.fits')
+    std_outfile = os.path.join('/Users/joe/Dropbox/PypeIt_Redux/MOSFIRE/Nov19/quicklook/Science/',
+                               'spec1d_m191118_0064-GD71_MOSFIRE_2019Nov18T104704.507.fits')
 
     # Read in the msbpm
     det = 1 # MOSFIRE has a single detector
@@ -284,7 +300,28 @@ def main(pargs):
     # Parse the offset information out of the headers. TODO in the future get this out of fitstable
     dither_pattern_A, dither_id_A, offset_arcsec_A = parse_dither_pattern(A_files, spectrograph.primary_hdrext)
     dither_pattern_B, dither_id_B, offset_arcsec_B = parse_dither_pattern(B_files, spectrograph.primary_hdrext)
[…]
