### 2020-05-26 1a26f4d
implement new data container
