+++ pypeit/flatfield.py
@@ -48,19 +48,26 @@ class FlatImages(datamodel.DataContainer):
     master_file_format = 'fits'
 
     datamodel = {
-        'procflat':  dict(otype=np.ndarray, atype=np.floating, desc='Processed, combined flats'),
-        'pixelflat': dict(otype=np.ndarray, atype=np.floating, desc='Pixel normalized flat'),
-        'flat_model': dict(otype=np.ndarray, atype=np.floating, desc='Model flat'),
         'PYP_SPEC': dict(otype=str, desc='PypeIt spectrograph name'),
-        'bpmflats': dict(otype=np.ndarray, atype=np.integer,
-                         desc='Mirrors SlitTraceSet mask for the Flat-specific flags'),
-        'spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline,
-                              desc='B-spline models for Illumination flat'),
         'spat_id': dict(otype=np.ndarray, atype=np.integer, desc='Slit spat_id '),
+        'pixelflat_raw': dict(otype=np.ndarray, atype=np.floating, desc='Processed, combined pixel flats'),
+        'pixelflat_norm': dict(otype=np.ndarray, atype=np.floating, desc='Normalized pixel flat'),
+        'pixelflat_model': dict(otype=np.ndarray, atype=np.floating, desc='Model flat'),
+        'pixelflat_spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline,
+                                        desc='B-spline models for pixel flat'),
+        'pixelflat_bpm': dict(otype=np.ndarray, atype=np.integer,
+                              desc='Mirrors SlitTraceSet mask for the Flat-specific flags'),
+        'illumflat_raw': dict(otype=np.ndarray, atype=np.floating, desc='Processed, combined illum flats'),
+        'illumflat_spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline,
+                                        desc='B-spline models for illum flat'),
+        'illumflat_bpm': dict(otype=np.ndarray, atype=np.integer,
+                              desc='Mirrors SlitTraceSet mask for the Flat-specific flags'),
     }
 
-    def __init__(self, procflat=None, pixelflat=None, bpmflats=None,
-                 flat_model=None, spat_bsplines=None, PYP_SPEC=None, spat_id=None):
+    def __init__(self, PYP_SPEC=None, spat_id=None,
+                 illumflat_raw=None, illumflat_spat_bsplines=None, illumflat_bpm=None,
+                 pixelflat_raw=None, pixelflat_norm=None, pixelflat_bpm=None,
+                 pixelflat_model=None, pixelflat_spat_bsplines=None):
         # Parse
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         d = dict([(k,values[k]) for k in args[1:]])
@@ -123,6 +130,82 @@ class FlatImages(datamodel.DataContainer):
                 d[0][key] = self[key]
         # Return
         return d
+
+    def merge_with(self, flatimgs):
+        """
+        Merge flatimgs into the current :class:`pypeit.flatfield.FlatImages` class.
+
+        Parameters
+        ----------
+        flatimgs : :class:`pypeit.flatfield.FlatImages`
+        """
+        for key in flatimgs.keys():
+            # Skip None
+            if flatimgs[key] is None:
+                continue
+            else:
+                self[key] = flatimgs[key]
+
+    @property
+    def procflat(self, frametype='pixel'):
+        if frametype == 'pixel':
+            return self.pixelflat_raw
+        elif frametype == 'illum':
+            return self.illumflat_raw
+        else:
+            # Return the default option, but warn of the invalid type
[…]
