### 2020-05-26 7c07b4b
arxiv wavelength solution
