### 2020-05-27 47e945f
Killer quicklook. Some changes throughout.
