### 2020-05-27 a946dd6
tweaking some parameters
