### 2020-05-27 e1e1e8d
Fixing data typing for MOSFIRE.
