### 2020-05-27 4ff252a
minor fix on arc QA plot
