### 2020-05-27 d221061
back to merging and validation
