### 2020-05-27 ec06954
doc fixes
