+++ doc/api/pypeit.images.rst
@@ -5,6 +5,7 @@ Submodules
 ----------
 
 .. toctree::
+   :maxdepth: 4
 
    pypeit.images.buildimage
    pypeit.images.combineimage
