### 2020-05-27 358c012
Bump version: 1.0.4dev → 1.0.4
