+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.0.4dev
+current_version = 1.0.4
 commit = False
 tag = False
 tag_name = {new_version}
