+++ pypeit/flatfield.py
@@ -63,19 +63,24 @@ class FlatImages(datamodel.DataContainer):
         'illumflat_bpm': dict(otype=np.ndarray, atype=np.integer,
                               desc='Mirrors SlitTraceSet mask for the Flat-specific flags'),
         'PYP_SPEC': dict(otype=str, desc='PypeIt spectrograph name'),
-        'spat_id': dict(otype=np.ndarray, atype=np.integer, desc='Slit spat_id '),
+        'spat_id': dict(otype=np.ndarray, atype=np.integer, desc='Slit spat_id'),
+        'procflat': dict(otype=np.ndarray, atype=np.floating, desc='testing'),
+        'pixelflat': dict(otype=np.ndarray, atype=np.floating, desc='testing'),
+        'flat_model': dict(otype=np.ndarray, atype=np.floating, desc='testing'),
+        'bpmflats': dict(otype=np.ndarray, atype=np.integer, desc='testing'),
+        'spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline, desc = 'testing'),
     }
 
     def __init__(self, pixelflat_raw=None, pixelflat_norm=None, pixelflat_bpm=None,
                  pixelflat_model=None, pixelflat_spat_bsplines=None, pixelflat_spec_illum=None,
                  illumflat_raw=None, illumflat_spat_bsplines=None, illumflat_bpm=None,
-                 PYP_SPEC=None, spat_id=None):
+                 PYP_SPEC=None, spat_id=None,
+                 procflat=None, pixelflat=None, flat_model=None, bpmflats=None, spat_bsplines=None):
         # Parse
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         d = dict([(k,values[k]) for k in args[1:]])
         # Setup the DataContainer
         datamodel.DataContainer.__init__(self, d=d)
-        self.frametype = None
 
     def _init_internals(self):
         # Master stuff
@@ -130,15 +135,15 @@ class FlatImages(datamodel.DataContainer):
             elif key == 'pixelflat_spat_bsplines':
                 for ss, spat_bspl in enumerate(self[key]):
                     # Naming
-                    spat_bspl.hdu_prefix = 'SPAT_ID-{}_'.format(self.spat_id[ss])
-                    dkey = 'pixbspl-{}'.format(self.spat_id[ss])
+                    spat_bspl.hdu_prefix = 'PIXELFLAT_SPAT_ID-{}_'.format(self.spat_id[ss])
+                    dkey = 'bspline-{}'.format(self.spat_id[ss])
                     # Save
                     d.append({dkey: spat_bspl})
             elif key == 'illumflat_spat_bsplines':
                 for ss, spat_bspl in enumerate(self[key]):
                     # Naming
-                    spat_bspl.hdu_prefix = 'SPAT_ID-{}_'.format(self.spat_id[ss])
-                    dkey = 'illbspl-{}'.format(self.spat_id[ss])
+                    spat_bspl.hdu_prefix = 'ILLUMFLAT_SPAT_ID-{}_'.format(self.spat_id[ss])
+                    dkey = 'bspline-{}'.format(self.spat_id[ss])
                     # Save
                     d.append({dkey: spat_bspl})
             else: # Add to header of the primary image
@@ -170,69 +175,27 @@ class FlatImages(datamodel.DataContainer):
     #     datamodel.DataContainer.__init__(self, d=d)
     #     #return FlatImages(**self)
 
-    def set_frametype(self, frametype="pixel"):
-        self.frametype = frametype
-
     @property
     def procflat(self):
-        if self.frametype is None or self.frametype == 'pixel':
-            return self.pixelflat_raw
-        elif self.frametype == 'illum':
-            return self.illumflat_raw
-        else:
-            # Return the default option, but warn of the invalid type
[…]
