### 2020-05-27 396e974
Bump version: 1.0.4 → 1.0.5dev
