+++ pypeit/images/rawimage.py
@@ -235,10 +235,10 @@ class RawImage(object):
         # Apply the relative spectral illumination
         spec_illum = 1.0
         if self.par['use_specillum']:
-            if flatimages is None or flatimages.spec_illum is None:
+            if flatimages is None or flatimages.pixelflat_spec_illum is None:
                 msgs.error("Spectral illumination correction desired but not generated/provided.")
             else:
-                spec_illum = flatimages.spec_illum.copy()
+                spec_illum = flatimages.pixelflat_spec_illum.copy()
 
         # Flat field -- We cannot do illumination flat without a pixel flat (yet)
         if self.par['use_pixelflat'] or self.par['use_illumflat']:
