### 2020-05-28 8b216ae
Fixing data typing for MOSFIRE.
