### 2020-05-28 40ff2ca
Fixing data typing for MOSFIRE.
