### 2020-05-29 0e0dc0f
Unelegant hack for crashes for bad headers.
