### 2020-06-06 #2243 @kyleniemeyer
Hi @harpolea, are you able to edit this submission, as your first editorial assignment?
