### 2020-06-06 6cfa642
Fixing data typing for MOSFIRE.
