### 2020-06-07 d022702
patterns implemented
