+++ pypeit/core/procimg.py
@@ -5,6 +5,7 @@
 """
 import numpy as np
 from scipy import signal, ndimage
+from scipy.optimize import curve_fit
 from IPython import embed
 
 from pypeit import msgs
@@ -332,8 +333,15 @@ def subtract_overscan(rawframe, datasec_img, oscansec_img,
             Frame from which to subtract overscan
         numamplifiers (int):
             Number of amplifiers for this detector.
-        datasec_img (np.ndarray):
-        oscansec_img (np.ndarray):
+        datasec_img (:obj:`numpy.ndarray`):
+            An array the same shape as rawframe that identifies
+            the pixels associated with the data on each amplifier.
+            0 for not data, 1 for amplifier 1, 2 for amplifier 2, etc.
+        oscansec_img (:obj:`numpy.ndarray`):
+            An array the same shape as rawframe that identifies
+            the pixels associated with the overscan region on each
+            amplifier.
+            0 for not data, 1 for amplifier 1, 2 for amplifier 2, etc.
         method (:obj:`str`, optional):
             The method used to fit the overscan region.  Options are
             polynomial, savgol, median.
@@ -387,6 +395,164 @@ def subtract_overscan(rawframe, datasec_img, oscansec_img,
         no_overscan[tuple(data_slice)] -= (ossub[:, None] if compress_axis == 1 else ossub[None, :])
 
     return no_overscan
+
+
+def subtract_pattern(rawframe, datasec_img, oscansec_img, frequency=None, axis=1):
+    """
+    Subtract a sinusoidal pattern from the input rawframe. The algorithm
+    calculates the frequency of the signal, generates a model, and subtracts
+    this signal from the data. This sinusoidal pattern noise was first
+    identified in KCWI, but the source of this pattern noise is not
+    currently known.
+
+    Args:
+        rawframe (:obj:`numpy.ndarray`):
+            Frame from which to subtract overscan
+        numamplifiers (int):
+            Number of amplifiers for this detector.
+        datasec_img (:obj:`numpy.ndarray`):
+            An array the same shape as rawframe that identifies
+            the pixels associated with the data on each amplifier.
+            0 for not data, 1 for amplifier 1, 2 for amplifier 2, etc.
+        oscansec_img (:obj:`numpy.ndarray`):
+            An array the same shape as rawframe that identifies
+            the pixels associated with the overscan region on each
+            amplifier.
+            0 for not data, 1 for amplifier 1, 2 for amplifier 2, etc.
+        frequency (float, list, optional):
+            The frequency (or list of frequencies - one for each amplifier)
+            of the sinusoidal pattern. If None, the frequency of each amplifier
+            will be determined from the overscan region.
+        axis (int):
+            Which axis should the pattern subtraction be applied?
+
+    Returns:
+        :obj:`numpy.ndarray`: The input frame with the pattern subtracted
+    """
+    # TODO :: REMOVE THIS EMBED BEFORE MERGE!!!
+    embed()
+
+    # Copy the data so that the subtraction is not done in place
+    frame_orig = rawframe.copy()
+    outframe = rawframe.copy()
+    tmp_oscan = oscansec_img.copy()
+    tmp_data = datasec_img.copy()
+    if axis == 0:
+        frame_orig = rawframe.copy().T
+        outframe = rawframe.copy().T
+        tmp_oscan = oscansec_img.copy().T
[…]
