### 2020-06-08 #2243 @harpolea
Hi @profxj, thanks for your patience in getting this review started!

Do you have any suggestions for potential reviewers? This [list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
