### 2020-06-08 #2243 @profxj
Of the ones that have listed Astronomy, I suggest:

ygrange
benjaminrose
steven-murray
github.com/Gabriel-p  -- Seems the best choice
astrom-tom  -- also very good
mwcraig
zpace

On Mon, Jun 8, 2020 at 4:44 AM Alice Harpole <notifications@github.com>
wrote:

> Hi @profxj <https://github.com/profxj>, thanks for your patience in
> getting this review started!
>
> Do you have any suggestions for potential reviewers? This list of people
> <https://bit.ly/joss-reviewers> have already agreed to review for JOSS
> and may be suitable for this submission (please start at the bottom of the
> list).
>
