### 2020-06-08 #2243 @astrom-tom
Hi @harpolea , it is gonna be difficult for me this week but I can try next week if needed. Let me know!
