### 2020-06-08 #2243 @harpolea
Hi @astrom-tom, we aim for reviews to be completed within 2-4 weeks, so as long as you feel that you can complete your review within that time, that should be ok!
