### 2020-06-08 #2308 @whedon
[REVIEW]: PypeIt: The Python Spectroscopic Data Reduction Pipeline 
**Submitting author:** @profxj (<a href="http://orcid.org/0000-0002-7738-6875">J. Xavier Prochaska</a>)
**Repository:** <a href="https://github.com/pypeit/PypeIt" target ="_blank">https://github.com/pypeit/PypeIt</a>
**Version:** v1.3
