### 2020-06-10 pypeit/PypeIt#949 @Gabriel-p
[JOSS REVIEW] Installation questions
Hi there I'm Gabriel one of the reviewers for the [Joss article](https://github.com/openjournals/joss-reviews/issues/2308). Some questions/comments:

1. In the [docs](https://pypeit.readthedocs.io/en/latest/installing.html) it is stated as a requirement:

> PyQT5 version 5 (needed for linetools)

Could you specify which version is required?

2.  `extension_helpers` should be `extension-helpers`.

3. By this:

> Use conda’s pip install to install the following:
>
> linetools – version 0.2 or later

do you mean "*use pip install inside the conda environment*"? What if the user is not using `conda`?

4. The packages in the docs do not match entirely those in the [requirements.txt](https://github.com/pypeit/PypeIt/blob/master/pypeit/requirements.txt) file:

* `packaging`: present in `requirements.txt`, not in the docs. Optional? For developers?
* `pytest`: present in `requirements.txt`, not in the docs. For developers? If so, is `h5py` also for developers as stated in that file?

5. There a re **a lot** of dependencies. Could you provide a single `conda` command to install those listed at the top of the docs, for convenience?
