+++ pypeit/flatfield.py
@@ -152,30 +152,6 @@ class FlatImages(datamodel.DataContainer):
         # Return
         return d
 
-    # def merge_with(self, flatimgs):
-    #     """
-    #     Merge flatimgs into the current :class:`pypeit.flatfield.FlatImages` class.
-    #
-    #     Parameters
-    #     ----------
-    #     flatimgs : :class:`pypeit.flatfield.FlatImages`
-    #
-    #     Parameters
-    #     ----------
-    #     :class:`pypeit.flatfield.FlatImages` : A new instance of the FlatImages class with merged properties.
-    #     """
-    #     for key in flatimgs.keys():
-    #         # Skip None
-    #         if flatimgs[key] is None:
-    #             continue
-    #         else:
-    #             self[key] = flatimgs[key]
-    #     # Regenerate the data container
-    #     d = dict([(k, self[k]) for k in self.keys()])
-    #     # Setup the DataContainer
-    #     datamodel.DataContainer.__init__(self, d=d)
-    #     #return FlatImages(**self)
-
     @property
     def procflat(self):
         return self.pixelflat_raw
@@ -242,8 +218,10 @@ class FlatImages(datamodel.DataContainer):
         # Now the bsplines
         list_of_pixbsplines, list_of_illbsplines = [], []
         pixspat_ids, illspat_ids = [], []
+        has_pixel, has_illum = False, False
         for ihdu in hdu:
             if 'PIXELFLAT' in ihdu.name and 'BSPLINE' in ihdu.name:
+                has_pixel = True
                 ibspl = bspline.bspline.from_hdu(ihdu)
                 if ibspl.version != bspline.bspline.version:
                     msgs.warn("Your bspline is out of date!!")
@@ -253,6 +231,7 @@ class FlatImages(datamodel.DataContainer):
                 i1 = ihdu.name.find('_BSP')
                 pixspat_ids.append(int(ihdu.name[i0+3:i1]))
             if 'ILLUMFLAT' in ihdu.name and 'BSPLINE' in ihdu.name:
+                has_illum = True
                 ibspl = bspline.bspline.from_hdu(ihdu)
                 if ibspl.version != bspline.bspline.version:
                     msgs.warn("Your bspline is out of date!!")
@@ -262,9 +241,9 @@ class FlatImages(datamodel.DataContainer):
                 i1 = ihdu.name.find('_BSP')
                 illspat_ids.append(int(ihdu.name[i0 + 3:i1]))
         # Check
-        if pixspat_ids != _d['spat_id'].tolist():
+        if pixspat_ids != _d['spat_id'].tolist() and has_pixel:
             msgs.error("Bad parsing of pixelflat BSPLINE spat_id in the MasterFlat")
-        if illspat_ids != _d['spat_id'].tolist():
+        if illspat_ids != _d['spat_id'].tolist() and has_illum:
             msgs.error("Bad parsing of illumflat BSPLINE spat_id in the MasterFlat")
         # Finish
         _d['pixelflat_spat_bsplines'] = np.asarray(list_of_pixbsplines)
@@ -1379,3 +1358,57 @@ def show_flats(image_list, wcs_match=True, slits=None):
         if clear:
             clear = False
 
+
+def merge(pixelflat, illumflat):
+    """
+    Merge secondary into primary, and return a merged :class:`pypeit.flatfield.FlatImages` class.
+    If an element exists in both primary and secondary, the primary value is taken
+
+    Parameters
+    ----------
+    pixelflat : :class:`pypeit.flatfield.FlatImages`
+        The pixelflat instance of FlatImages
[…]
