+++ pypeit/core/procimg.py
@@ -495,16 +495,30 @@ def subtract_pattern(rawframe, datasec_img, oscansec_img, frequency=None, axis=1
         val = np.zeros(overscan.shape[0])
         # Get the best estimate of the amplitude
         for ii in range(overscan.shape[0]):
-            popt, pcov = curve_fit(cosfunc, xdata, overscan[ii, :], p0=[amps[ii], use_fr, phss[ii]],
-                                   bounds=([-np.inf, use_fr * 0.99999999, -np.inf], [+np.inf, use_fr * 1.00000001, +np.inf]))
+            try:
+                popt, pcov = curve_fit(cosfunc, xdata, overscan[ii, :], p0=[amps[ii], use_fr, phss[ii]],
+                                       bounds=([-np.inf, use_fr * 0.99999999, -np.inf], [+np.inf, use_fr * 1.00000001, +np.inf]))
+            except ValueError:
+                msgs.warn("Input data invalid for pattern subtraction of row {0:d}/{1:d}".format(ii + 1, overscan.shape[0]))
+                continue
+            except RuntimeError:
+                msgs.warn("Pattern subtraction fit failed for row {0:d}/{1:d}".format(ii + 1, overscan.shape[0]))
+                continue
             val[ii] = popt[0]
             model_pattern[ii, :] = cosfunc(xdata_all, *popt)
         use_amp = np.median(val)
         # Get the best estimate of the phase, and generate a model
         for ii in range(overscan.shape[0]):
-            popt, pcov = curve_fit(cosfunc, xdata, overscan[ii, :], p0=[use_amp, use_fr, phss[ii]],
-                                   bounds=([use_amp * 0.99999999, use_fr * 0.99999999, -np.inf],
-                                           [use_amp * 1.00000001, use_fr * 1.00000001, +np.inf]))
+            try:
+                popt, pcov = curve_fit(cosfunc, xdata, overscan[ii, :], p0=[use_amp, use_fr, phss[ii]],
+                                       bounds=([use_amp * 0.99999999, use_fr * 0.99999999, -np.inf],
+                                               [use_amp * 1.00000001, use_fr * 1.00000001, +np.inf]))
+            except ValueError:
+                msgs.warn("Input data invalid for pattern subtraction of row {0:d}/{1:d}".format(ii + 1, overscan.shape[0]))
+                continue
+            except RuntimeError:
+                msgs.warn("Pattern subtraction fit failed for row {0:d}/{1:d}".format(ii + 1, overscan.shape[0]))
+                continue
             model_pattern[ii, :] = cosfunc(xdata_all, *popt)
         outframe[tuple(osd_slice)] -= model_pattern
 
@@ -534,6 +548,8 @@ def pattern_frequency(frame, axis=1):
     Args:
         frame (:obj:`numpy.ndarray`):
             2D array to measure the pattern frequency
+        axis (int, optional):
+            Which axis should the pattern frequency be measured?
 
     Returns:
         :obj:`float`: The frequency of the sinusoidal pattern.
@@ -563,10 +579,6 @@ def pattern_frequency(frame, axis=1):
     phss = np.arctan2(amp.imag, amp.real)[idx]
     frqs = idx[1]
 
-    # This does a reasonable job, but needs to be much better - a fit or CC would be best to get subpixel sampling.
-    # posn = np.linspace(0.0, 1.0, arr.shape[1])[np.newaxis, :]
-    # tmpsig = amps[:, np.newaxis] * np.cos(2.0 * np.pi * frqs[:, np.newaxis] * posn + phss[:, np.newaxis])
-
[…]
