+++ pypeit/tests/test_calibrations.py
@@ -33,7 +33,7 @@ def fitstbl():
         return fitstbl
 
     fitstbl = dummy_fitstbl(directory=os.path.join(os.getenv('PYPEIT_DEV'), 'RAW_DATA',
-                                                           'shane_kast_blue', '600_4310_d55'))
+                                                   'shane_kast_blue', '600_4310_d55'))
     # Set the Bias to known
     fitstbl['framebit'][0] = fitstbl.type_bitmask.turn_off(fitstbl['framebit'][0], flag='bias')
     fitstbl['filename'][1] = 'b1.fits.gz'
