### 2020-06-12 ebdfdb5
skyreg update
