+++ pypeit/images/buildimage.py
@@ -162,7 +162,8 @@ def buildimage_fromlist(spectrograph, det, frame_par, file_list,
                                    flatimages=flatimages,
                                    sigma_clip=sigma_clip,
                                    sigrej=sigrej, maxiters=maxiters,
-                                   ignore_saturation=ignore_saturation, slits=slits)
+                                   ignore_saturation=ignore_saturation, slits=slits,
+                                   combine=frame_par['process']['combine'])
     #
     # Decorate according to the type of calibration
     #   Primarily for handling MasterFrames
