+++ pypeit/images/combineimage.py
@@ -54,7 +54,7 @@ class CombineImage(object):
             msgs.error('Combineimage requires a list of files to instantiate')
 
     def run(self, bias=None, flatimages=None, ignore_saturation=False, sigma_clip=True,
-            bpm=None, sigrej=None, maxiters=5, slits=None, dark=None):
+            bpm=None, sigrej=None, maxiters=5, slits=None, dark=None, combine='weightmean'):
         """
         Generate a PypeItImage from a list of images
 
@@ -77,6 +77,9 @@ class CombineImage(object):
                 If True, turn off the saturation flag in the individual images before stacking
                 This avoids having such values set to 0 which for certain images (e.g. flat calibrations)
                 can have unintended consequences.
+            combine (str):
+                Method to combine images
+                Allowed options are 'weightmean', 'median'
 
         Returns:
             :class:`pypeit.images.pypeitimage.PypeItImage`:
@@ -149,16 +152,24 @@ class CombineImage(object):
         img_list = [img_stack]
         var_stack = utils.inverse(ivar_stack)
         var_list = [var_stack, rn2img_stack]
-        img_list_out, var_list_out, outmask, nused = combine.weighted_combine(
-            weights, img_list, var_list, (mask_stack == 0),
-            sigma_clip=sigma_clip, sigma_clip_stack=img_stack, sigrej=sigrej, maxiters=maxiters)
+        if combine == 'weightmean':
+            img_list_out, var_list_out, gpm, nused = combine.weighted_combine(
+                weights, img_list, var_list, (mask_stack == 0),
+                sigma_clip=sigma_clip, sigma_clip_stack=img_stack, sigrej=sigrej, maxiters=maxiters)
+        elif combine == 'median':
+            img_list_out = [np.median(img_stack, axis=0)]
+            var_list_out = [np.median(var_stack, axis=0)]
+            var_list_out += [np.median(rn2img_stack, axis=0)]
+            gpm = np.ones_like(img_list_out[0], dtype='bool')
+        else:
+            msgs.error("Bad choice for combine.  Allowed options are 'median', 'weightmean'.")
 
         # Build the last one
         final_pypeitImage = pypeitimage.PypeItImage(img_list_out[0],
                                                     ivar=utils.inverse(var_list_out[0]),
                                                     bpm=pypeitImage.bpm,
                                                     rn2img=var_list_out[1],
-                                                    crmask=np.invert(outmask),
+                                                    crmask=np.invert(gpm),
                                                     detector=pypeitImage.detector,
                                                     PYP_SPEC=pypeitImage.PYP_SPEC)
         # Internals
