### 2020-06-15 51f0b54
I have adjusted these files to allow for the manual addition of objects. Generally I've used the object finding gui to determine the parameter set, but this isn't necessary to add manual extraction parameters. I've relied on access to par['reduce]['extraction']['manual'] in metadata.py from the pypeit file to make this functional and I'm a little uncertain if this is accessed in the same way as in v0.13.1 (which is where I've tested this code), but the code in extract.py should work once the manual extraction parameters are accessed from the pypeit file. I've also made it possible to manually 
[…]
