+++ pypeit/pypeit.py
@@ -653,6 +653,7 @@ class PypeIt(object):
         # Prep for manual extraction (if requested)
         manual_extract_dict = self.fitstbl.get_manual_extract(frames, det)
 
+        msgs.warn("Check that manual extract dict is passed: {}".format(manual_extract_dict))
         skymodel, objmodel, ivarmodel, outmask, sobjs, waveImg, tilts = self.redux.run(
             std_trace=std_trace, manual_extract_dict=manual_extract_dict, show_peaks=self.show,
             basename=self.basename, ra=self.fitstbl["ra"][frames[0]], dec=self.fitstbl["dec"][frames[0]],
