+++ pypeit/flatfield.py
@@ -30,9 +30,171 @@ from pypeit.core import basis
 from pypeit import slittrace
 
 
+class Flats:
+    """
+    A utility container to extract the desired properties from FlatImages
+    """
+
+    def __init__(self, flatimages):
+        self.flatimages = flatimages
+
+    @property
+    def shape(self):
+        if self.flatimages.pixelflat_raw is not None:
+            return self.flatimages.pixelflat_raw.shape
+        elif self.flatimages.illumflat_raw is not None:
+            return self.flatimages.illumflat_raw.shape
+        else:
+            msgs.error("Shape of FlatImages could not be determined")
+
+    @property
+    def procflat(self, frametype='pixel'):
+        if frametype == 'illum':
+            return self.flatimages.illumflat_raw
+        else:
+            return self.flatimages.pixelflat_raw
+
+    @property
+    def pixelflat(self):
+        return self.flatimages.pixelflat_norm
+
+    @property
+    def flat_model(self):
+        return self.flatimages.pixelflat_model
+
+    @property
+    def bpmflats(self, frametype='pixel'):
+        # Check if both BPMs are none
+        if self.flatimages.pixelflat_bpm is None and self.flatimages.illumflat_bpm is None:
+            msgs.warn("FlatImages contains no BPM - trying to generate one")
+            return np.zeros(self.shape, dtype=np.int)
+        # Now return the requested case, checking for None
+        if frametype == 'illum':
+            if self.flatimages.illumflat_bpm is not None:
+                return self.flatimages.illumflat_bpm
+            else:
+                msgs.warn("illumflat has no BPM - using the pixelflat BPM")
+                return self.flatimages.pixelflat_bpm
+        else:
+            if self.flatimages.pixelflat_bpm is not None:
+                return self.flatimages.pixelflat_bpm
+            else:
+                msgs.warn("illumflat has no BPM - using the illumflat BPM")
+                return self.flatimages.illumflat_bpm
+
+    @property
+    def spat_bsplines(self, frametype='illum'):
+        # Check if both spat bsplines are none
+        if self.flatimages.pixelflat_spat_bsplines is None and self.flatimages.illumflat_spat_bsplines is None:
+            msgs.error("FlatImages contains no spatial bspline fit")
+        # Now return the requested case, checking for None
+        if frametype == 'illum':
+            if self.flatimages.illumflat_spat_bsplines is not None:
+                return self.flatimages.illumflat_spat_bsplines
+            else:
+                msgs.warn("illumflat has no spatial bspline fit - using the pixelflat")
+                return self.flatimages.pixelflat_spat_bsplines
+        else:
+            if self.flatimages.pixelflat_spat_bsplines is not None:
+                return self.flatimages.pixelflat_spat_bsplines
+            else:
+                msgs.warn("pixelflat has no spatial bspline fit - using the illumflat")
+                return self.flatimages.illumflat_spat_bsplines
+
+    def fit2illumflat(self, slits, frametype='illum', initial=False, flexure_shift=None):
+        """
+
+        Args:
+            slits (:class:`pypeit.slittrace.SlitTraceSet`):
+            frametype (str):
+                Should the pixel or illum flat spatial profile be generated? (options include: 'illum' or 'pixel').
+                The default is to use 'illum' unless frametype='pixel'.
+            initial (bool, optional):
+            flexure_shift (float, optional):
+
+        Returns:
+
+        """
+        illumflat = np.ones_like(self.pixelflat)
+        # Load spatial bsplines
+        spat_bsplines = self.spat_bsplines(frametype=frametype)
+
+        # Loop
+        for slit_idx in range(slits.nslits):
+            # Skip masked
[…]
