+++ pypeit/flatfield.py
@@ -35,7 +35,9 @@ class Flats:
     A utility container to extract the desired properties from FlatImages
     """
 
-    def __init__(self, flatimages):
+    def __init__(self, flatimages=None):
+        if flatimages is None:
+            flatimages = FlatImages()
         self.flatimages = flatimages
 
     @property
@@ -47,23 +49,13 @@ class Flats:
         else:
             msgs.error("Shape of FlatImages could not be determined")
 
-    @property
-    def procflat(self, frametype='pixel'):
+    def get_procflat(self, frametype='pixel'):
         if frametype == 'illum':
             return self.flatimages.illumflat_raw
         else:
             return self.flatimages.pixelflat_raw
 
-    @property
-    def pixelflat(self):
-        return self.flatimages.pixelflat_norm
-
-    @property
-    def flat_model(self):
-        return self.flatimages.pixelflat_model
-
-    @property
-    def bpmflats(self, frametype='pixel'):
+    def get_bpmflats(self, frametype='pixel'):
         # Check if both BPMs are none
         if self.flatimages.pixelflat_bpm is None and self.flatimages.illumflat_bpm is None:
             msgs.warn("FlatImages contains no BPM - trying to generate one")
@@ -79,11 +71,30 @@ class Flats:
             if self.flatimages.pixelflat_bpm is not None:
                 return self.flatimages.pixelflat_bpm
             else:
-                msgs.warn("illumflat has no BPM - using the illumflat BPM")
+                msgs.warn("pixelflat has no BPM - using the illumflat BPM")
                 return self.flatimages.illumflat_bpm
 
     @property
-    def spat_bsplines(self, frametype='illum'):
+    def pixelflat(self):
+        return self.flatimages.pixelflat_norm
+
+    @property
+    def flat_model(self):
+        return self.flatimages.pixelflat_model
+
+    @property
+    def bpmflats(self):
+        return self.get_bpmflats('pixel')
+
+    @property
+    def spat_bsplines(self):
+        if self.flatimages.illumflat_spat_bsplines is not None:
+            return self.flatimages.illumflat_spat_bsplines
+        else:
+            msgs.warn("illumflat has no spatial bspline fit - using the pixelflat")
+            return self.flatimages.pixelflat_spat_bsplines
+
+    def get_spat_bsplines(self, frametype='illum'):
         # Check if both spat bsplines are none
         if self.flatimages.pixelflat_spat_bsplines is None and self.flatimages.illumflat_spat_bsplines is None:
             msgs.error("FlatImages contains no spatial bspline fit")
@@ -115,9 +126,9 @@ class Flats:
         Returns:
 
         """
-        illumflat = np.ones_like(self.pixelflat)
+        illumflat = np.ones(self.shape)
         # Load spatial bsplines
-        spat_bsplines = self.spat_bsplines(frametype=frametype)
+        spat_bsplines = self.get_spat_bsplines(frametype=frametype)
 
         # Loop
         for slit_idx in range(slits.nslits):
@@ -132,7 +143,6 @@ class Flats:
                                                       initial=initial,
                                                       slitid_img=_slitid_img,
                                                       flexure_shift=flexure_shift)
-
             illumflat[onslit] = spat_bsplines[slit_idx].value(spat_coo[onslit])[0]
         # TODO -- Update the internal one?  Or remove it altogether??
         return illumflat
@@ -167,14 +177,14 @@ class Flats:
             if self.flatimages.illumflat_spat_bsplines is not None:
                 illumflat_illum = self.fit2illumflat(slits, frametype='illum')
         # Decide which frames should be displayed
[…]
