+++ pypeit/flatfield.py
@@ -4,7 +4,6 @@ Implements the flat-field class.
 .. include common links, assuming primary doc root is up one directory
 .. include:: ../links.rst
 """
-import os
 import copy
 import inspect
 import numpy as np
@@ -20,7 +19,6 @@ from pypeit import utils
 from pypeit import ginga
 from pypeit import bspline
 
-from pypeit.par import pypeitpar
 from pypeit import datamodel
 from pypeit import masterframe
 from pypeit import wavecalib
@@ -28,186 +26,6 @@ from pypeit.core import flat
 from pypeit.core import tracewave
 from pypeit.core import basis
 from pypeit import slittrace
-
-
-class Flats:
-    """
-    A utility container to extract the desired properties from FlatImages
-
-    Args:
-        flatimages (:class:`pypeit.flatfield.FlatImages`):
-            An instance of FlatImages
-    """
-
-    def __init__(self, flatimages=None):
-        if flatimages is None:
-            flatimages = FlatImages()
-        self.flatimages = flatimages
-
-    @property
-    def shape(self):
-        if self.flatimages.pixelflat_raw is not None:
-            return self.flatimages.pixelflat_raw.shape
-        elif self.flatimages.illumflat_raw is not None:
-            return self.flatimages.illumflat_raw.shape
-        else:
-            msgs.error("Shape of FlatImages could not be determined")
-
-    def get_procflat(self, frametype='pixel'):
-        if frametype == 'illum':
-            return self.flatimages.illumflat_raw
-        else:
-            return self.flatimages.pixelflat_raw
-
-    def get_bpmflats(self, frametype='pixel'):
-        # Check if both BPMs are none
-        if self.flatimages.pixelflat_bpm is None and self.flatimages.illumflat_bpm is None:
-            msgs.warn("FlatImages contains no BPM - trying to generate one")
-            return np.zeros(self.shape, dtype=np.int)
-        # Now return the requested case, checking for None
-        if frametype == 'illum':
-            if self.flatimages.illumflat_bpm is not None:
-                return self.flatimages.illumflat_bpm
-            else:
-                msgs.warn("illumflat has no BPM - using the pixelflat BPM")
-                return self.flatimages.pixelflat_bpm
-        else:
-            if self.flatimages.pixelflat_bpm is not None:
-                return self.flatimages.pixelflat_bpm
-            else:
-                msgs.warn("pixelflat has no BPM - using the illumflat BPM")
-                return self.flatimages.illumflat_bpm
-
-    @property
-    def pixelflat(self):
-        return self.flatimages.pixelflat_norm
-
-    @property
-    def spec_illum(self):
-        return self.flatimages.pixelflat_spec_illum
-
-    @property
-    def flat_model(self):
-        return self.flatimages.pixelflat_model
-
-    @property
-    def bpmflats(self):
-        return self.get_bpmflats('pixel')
-
-    @property
-    def spat_bsplines(self):
-        if self.flatimages.illumflat_spat_bsplines is not None:
-            return self.flatimages.illumflat_spat_bsplines
-        else:
-            msgs.warn("illumflat has no spatial bspline fit - using the pixelflat")
-            return self.flatimages.pixelflat_spat_bsplines
-
-    def get_spat_bsplines(self, frametype='illum'):
-        # Check if both spat bsplines are none
-        if self.flatimages.pixelflat_spat_bsplines is None and self.flatimages.illumflat_spat_bsplines is None:
-            msgs.error("FlatImages contains no spatial bspline fit")
-        # Now return the requested case, checking for None
-        if frametype == 'illum':
-            if self.flatimages.illumflat_spat_bsplines is not None:
-                return self.flatimages.illumflat_spat_bsplines
-            else:
-                msgs.warn("illumflat has no spatial bspline fit - using the pixelflat")
[…]
