+++ pypeit/reduce.py
@@ -357,14 +357,8 @@ class Reduce(object):
         # Check if the user wants to overwrite the skymask with a pre-defined sky regions file
         skymask_init, usersky = self.load_skyregions(skymask_init)
 
-        # Define the trim regions for different pypelines
-        if self.spectrograph.pypeline == "IFU":
-            trim_edg = (0, 0)
-        else:
-            trim_edg = (3, 3)
-
         # Global sky subtract
-        self.initial_sky = self.global_skysub(skymask=skymask_init, trim_edg=trim_edg).copy()
+        self.initial_sky = self.global_skysub(skymask=skymask_init).copy()
 
         # Second pass object finding on sky-subtracted image
         if (not self.std_redux) and (not self.par['reduce']['findobj']['skip_second_find']):
@@ -384,8 +378,7 @@ class Reduce(object):
                     self.par['reduce']['findobj']['skip_second_find'] or usersky):
                 self.global_sky = self.initial_sky.copy()
             else:
-                self.global_sky = self.global_skysub(skymask=self.skymask,
-                                                     trim_edg=trim_edg, show=self.reduce_show)
+                self.global_sky = self.global_skysub(skymask=self.skymask, show=self.reduce_show)
             # Extract + Return
             self.skymodel, self.objmodel, self.ivarmodel, self.outmask, self.sobjs \
                 = self.extract(self.global_sky, self.sobjs_obj)
@@ -533,92 +526,24 @@ class Reduce(object):
         # Mask objects using the skymask? If skymask has been set by objfinding, and masking is requested, then do so
         skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
 
-        if self.par['reduce']['skysub']['joint_fit']:
-            msgs.info("Performing joint global sky subtraction")
-            thismask = (self.slitmask != 0)
-            inmask = ((self.sciImg.fullmask == 0) & thismask & skymask_now).astype(np.bool)
-            # Convert the wavelength image to A/pixel, registered at pixel 0 (this gives something like
-            # the tilts frame, but conserves wavelength position in each slit)
-            tilt_wave = (self.waveimg - self.waveimg.min()) / (self.waveimg.max() - self.waveimg.min())
+        # Loop on slits
+        for slit_idx in gdslits:
+            slit_spat = self.slits.spat_id[slit_idx]
+            msgs.info("Global sky subtraction for slit: {:d}".format(slit_idx))
+            thismask = self.slitmask == slit_spat
+            inmask = (self.sciImg.fullmask == 0) & thismask & skymask_now
             # Find sky
             self.global_sky[thismask] \
-                = skysub.global_skysub(self.sciImg.image, self.sciImg.ivar, tilt_wave,
-                                       thismask, self.slits_left, self.slits_right, inmask=inmask,
-                                       sigrej=sigrej, trim_edg=trim_edg,
-                                       bsp=self.par['reduce']['skysub']['bspline_spacing'],
-                                       no_poly=self.par['reduce']['skysub']['no_poly'],
-                                       pos_mask=(not self.ir_redux), show_fit=show_fit)
+                    = skysub.global_skysub(self.sciImg.image, self.sciImg.ivar, self.tilts,
+                                           thismask, self.slits_left[:,slit_idx],
+                                           self.slits_right[:,slit_idx],
+                                           inmask=inmask, sigrej=sigrej,
+                                           bsp=self.par['reduce']['skysub']['bspline_spacing'],
+                                           no_poly=self.par['reduce']['skysub']['no_poly'],
+                                           pos_mask=(not self.ir_redux), show_fit=show_fit)
             # Mask if something went wrong
             if np.sum(self.global_sky[thismask]) == 0.:
[…]
