### 2020-06-18 pypeit/PypeIt#949 @profxj
I had been assuming additional comments/requests were coming and was going to wait to address them all at once.  But let me know if you are waiting on us.  Thanks!
