### 2020-06-19 pypeit/PypeIt#949 @profxj
Got it.  Will do so soon.
