+++ pypeit/core/wavecal/wv_fitting.py
@@ -286,7 +286,7 @@ def iterative_fitting(spec, tcent, ifit, IDs, llist, disp,
             msgs.warn("Bad fit!!")
             return None
 
-        rms_ang = pypeitFit.calc_fit_rms()
+        rms_ang = pypeitFit.calc_fit_rms(apply_mask=True)
         rms_pix = rms_ang/disp
         if verbose:
             msgs.info('n_order = {:d}'.format(n_order) + ': RMS = {:g}'.format(rms_pix))
@@ -340,7 +340,7 @@ def iterative_fitting(spec, tcent, ifit, IDs, llist, disp,
     ions = all_idsion[ifit]
 #    ions = all_idsion[ifit][mask == 0]
     # Final RMS
-    rms_ang = pypeitFit.calc_fit_rms()#xfit[pypeitFit.gpm == 0] / xnspecmin1,
+    rms_ang = pypeitFit.calc_fit_rms(apply_mask=True)#xfit[pypeitFit.gpm == 0] / xnspecmin1,
                                      #yfit[pypeitFit.gpm == 0],
                                      #weights=wfit[pypeitFit.gpm == 0])
 #    rms_ang = utils.calc_fit_rms(xfit, yfit, fit, func,
