### 2020-06-22 a609b22
deprecate fit_slit
