+++ pypeit/spectrographs/keck_kcwi.py
@@ -6,6 +6,8 @@ import numpy as np
 from IPython import embed
 
 from astropy.io import fits
+from astropy import wcs, units
+from astropy.coordinates import SkyCoord
 
 from pypeit import msgs
 from pypeit import telescopes
@@ -38,6 +40,8 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         self.grating = None
         self.optical_model = None
         self.detector_map = None
+        self.slicescale = 0.00037718  # Degrees per 'large slicer' slice
+        self.platescale = 0.00004048  # Degrees per unbinned pixel
 
     @property
     def pypeline(self):
@@ -79,7 +83,7 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
                         xgap            = 0.,
                         ygap            = 0.,
                         ysize           = 1.,
-                        platescale      = 0.147,  # arcsec/pixel
+                        platescale      = 0.145728,  # arcsec/pixel
                         darkcurr        = None,  # <-- TODO : Need to set this
                         mincounts       = -1e10,
                         saturation      = 65535.,
@@ -554,7 +558,7 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         Returns
         -------
         bpix : ndarray
-          0 = ok; 1 = Mask
+            0 = ok; 1 = Mask
 
         """
         bpm_img = self.empty_bpm(filename, det, shape=shape)
@@ -609,49 +613,166 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
 
         return bpm_img
 
-    def set_wcs(self, hdr):
-        """Set the WCS for this spectrograph
+    @staticmethod
+    def is_nasmask(hdr):
+        """Determine if a frame used NAS
+
+        Parameters
+        ----------
+        hdr : fits header
+            The header of the raw frame.
+
+        Returns
+        -------
+        bool : True if NAS used
         """
-        # TODO :: Need to set all of these!!
-        msgs.error("Should be using astropy.wcs")
-        ra = None
-        dec = None
-        wave0 = None
-        crpix1 = None
-        crpix2 = None
-        crpix3 = None
-        cd11 = None
-        cd21 = None
-        cd12 = None
-        cd22 = None
-        dwout = None
-        # WCS keywords
-        hdr['WCSDIM'] = (3, 'number of dimensions in WCS')
-        hdr['WCSNAME'] = ('KCWI', 'Name of WCS')
-        hdr['EQUINOX'] = (2000, 'EQUINOX')
-        hdr['RADESYS'] = ('FK5', 'WCS system')
-        hdr['CTYPE1'] = ('RA---TAN', '')
-        hdr['CTYPE2'] = ('DEC--TAN', '')
-        hdr['CTYPE3'] = ('AWAV', 'Air Wavelengths')
-        hdr['CUNIT1'] = ('deg', 'RA units')
-        hdr['CUNIT2'] = ('deg', 'DEC units')
-        hdr['CUNIT3'] = ('Angstrom', 'Wavelength units')
-        hdr['CNAME1'] = ('KCWI RA', 'RA name')
-        hdr['CNAME2'] = ('KCWI DEC', 'DEC name')
-        hdr['CNAME3'] = ('KCWI Wavelength', 'Wavelength name')
-        hdr['CRVAL1'] = (ra, 'RA zeropoint')
-        hdr['CRVAL2'] = (dec, 'DEC zeropoint')
-        hdr['CRVAL3'] = (wave0, 'Wavelength zeropoint')
-        hdr['CRPIX1'] = (crpix1, 'RA reference pixel')
-        hdr['CRPIX2'] = (crpix2, 'DEC reference pixel')
[…]
