### 2020-06-23 1b5f2f1
Bump version: 1.0.5dev → 1.0.5
