+++ CHANGES.rst
@@ -1,5 +1,8 @@
 
 
+1.0.6dev
+--------
+
 1.0.5 (23 Jun 2020)
 -------------------
 
