### 2020-06-23 8f47cd2
Bump version: 1.0.5 → 1.0.6dev
