+++ pypeit/spectrographs/lbt_mods.py
@@ -10,6 +10,7 @@ from pypeit.core import framematch
 from pypeit.par import pypeitpar
 from pypeit.spectrographs import spectrograph
 from pypeit.core import parse
+from pypeit.images import detector_container
 
 # FW ToDo: test MODS1B and MODS2B
 
@@ -107,7 +108,7 @@ class LBTMODSSpectrograph(spectrograph.Spectrograph):
                    & (fitstbl['dispname'] != 'Flat')
         if ftype == 'bias':
             return good_exp  & (fitstbl['idname'] == 'BIAS')
-        if ftype in ['pixelflat', 'trace']:
+        if ftype in ['pixelflat', 'trace', 'illumflat']:
             # Flats and trace frames are typed together
             return good_exp  & (fitstbl['idname'] == 'FLAT') & (fitstbl['decker'] != 'Imaging')
         if ftype in ['pinhole', 'dark']:
@@ -150,7 +151,8 @@ class LBTMODSSpectrograph(spectrograph.Spectrograph):
         # TODO These parameters should probably be stored in the detector par
 
         # Number of amplifiers (could pull from DetectorPar but this avoids needing the spectrograph, e.g. view_fits)
-        numamp = 4
+        detector_par = self.get_detector_par(hdu, det if det is None else 1)
+        numamp = detector_par['numamplifiers']
 
         # get the x and y binning factors...
         xbin, ybin = head['CCDXBIN'], head['CCDYBIN']
@@ -188,7 +190,7 @@ class LBTMODSSpectrograph(spectrograph.Spectrograph):
         # Need the exposure time
         exptime = hdu[self.meta['exptime']['ext']].header[self.meta['exptime']['card']]
         # Return, transposing array back to orient the overscan properly
-        return np.flipud(array), hdu, exptime, np.flipud(rawdatasec_img), np.flipud(oscansec_img)
+        return detector_par,np.flipud(array), hdu, exptime, np.flipud(rawdatasec_img), np.flipud(oscansec_img)
 
 
 class LBTMODS1RSpectrograph(LBTMODSSpectrograph):
@@ -201,6 +203,35 @@ class LBTMODS1RSpectrograph(LBTMODSSpectrograph):
         self.spectrograph = 'lbt_mods1r'
         self.camera = 'MODS1R'
         self.numhead = 1
+
+    def get_detector_par(self, hdu, det):
+        # Binning
+        xbin = hdu[0].header['CCDXBIN']
+        ybin = hdu[0].header['CCDYBIN']
+        binning = '{:},{:}'.format(xbin,ybin)
+
+        # Detector 1
+        detector_dict = dict(
+            binning= binning,
+            det=1,
+            dataext         = 0,
+            specaxis        = 0,
+            specflip        = False,
+            spatflip        = False,
+            platescale      = 0.123,
+            darkcurr        = 0.4,
+            saturation      = 65535.,
+            nonlinear       = 0.99,
+            mincounts       = -1e10,
+            numamplifiers   = 4,
+            gain            = np.atleast_1d([2.38,2.50,2.46,2.81]),
+            ronoise         = np.atleast_1d([3.78,4.04,4.74,4.14]),
+            datasec         = np.atleast_1d('[:,:]'),
+            oscansec        = np.atleast_1d('[:,:]')
+            )
+        detector = detector_container.DetectorContainer(**detector_dict)
+        return detector
+
 
 
     def default_pypeit_par(self):
@@ -217,14 +248,12 @@ class LBTMODS1RSpectrograph(LBTMODSSpectrograph):
         #par['calibrations']['wavelengths']['lamps'] = ['XeI','ArII','ArI','NeI','KrI']]
         par['calibrations']['wavelengths']['lamps'] = ['ArI','NeI','KrI','XeI']
[…]
