### 2020-06-23 55cf88c
DataContainer version/class check
