### 2020-06-25 8716705
avoid crash in autoid
