+++ CHANGES.rst
@@ -1,4 +1,8 @@
 
+Hotfixes after 1.0.5
+--------------------
+
+ - Deal with wavecalib crash
 
 1.0.5 (23 Jun 2020)
 -------------------
@@ -9,8 +13,6 @@
  - KCWI reduction complete up to spec2D frames
  - Implemented new flatfield DataContainer to separate
    pixelflat and illumflat
- - Deal with wavecalib crash
-
 
 1.0.4 (27 May 2020)
 -------------------
