+++ pypeit/spectrographs/keck_kcwi.py
@@ -149,8 +149,8 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         """
         meta = {}
         # Required (core)
-        meta['ra'] = dict(ext=0, card='RA')
-        meta['dec'] = dict(ext=0, card='DEC')
+        meta['ra'] = dict(ext=0, compound=True)
+        meta['dec'] = dict(ext=0, compound=True)
         meta['target'] = dict(ext=0, card='TARGNAME')
         meta['dispname'] = dict(ext=0, card='BGRATNAM')
         meta['decker'] = dict(ext=0, card='IFUNAM')
@@ -261,6 +261,18 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
             binspatial, binspec = parse.parse_binning(headarr[0]['BINNING'])
             binning = parse.binning2string(binspec, binspatial)
             return binning
+        elif meta_key == 'ra' or meta_key == 'dec':
+            try:
+                if self.is_nasmask(headarr[0]):
+                    hdrstr = 'RABASE' if meta_key == 'ra' else 'DECBASE'
+                else:
+                    hdrstr = 'RA' if meta_key == 'ra' else 'DEC'
+            except KeyError:
+                try:
+                    hdrstr = 'TARGRA' if meta_key == 'ra' else 'TARGDEC'
+                except KeyError:
+                    hdrstr = ''
+            return headarr[0][hdrstr]
         else:
             msgs.error("Not ready for this compound meta")
 
@@ -669,7 +681,7 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
             slscl = self.slicescale/4.0
         return pxscl, slscl
 
-    def get_wcs(self, hdr, maxslitlen, wave0, dwv):
+    def get_wcs(self, hdr, slits, wave0, dwv):
         """Get the WCS for a frame
 
         Parameters
@@ -677,8 +689,8 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         hdr : fits header
             The header of the raw frame. The information in this
             header will be extracted and returned as a WCS.
-        maxslitlen : float
-            Maximum slit length (in pixels)
+        slits : :class:`pypeit.slittrace.SlitTraceSet`
+            Master slit edges
         wave0 : float
             wavelength zeropoint
         dwv : float
@@ -689,25 +701,15 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         astropy.wcs : An astropy WCS object.
         """
         msgs.info("Calculating the WCS")
-        # Get the x and y binning factors...
+        # Get the x and y binning factors, and the typical slit length
         xbin, ybin = self.get_binning(hdr)
         slscl, pxscl = self.get_scales(hdr)
+        slitlength = int(np.round(np.median(slits.get_slitlengths(initial=True, median=True))))
 
         # Get RA/DEC
-        try:
-            if self.is_nasmask(hdr):
-                rastr = hdr['RABASE']
-                decstr = hdr['DECBASE']
-            else:
-                rastr = hdr['RA']
-                decstr = hdr['DEC']
-        except KeyError:
-            try:
-                rastr = hdr['TARGRA']
-                decstr = hdr['TARGDEC']
-            except KeyError:
-                rastr = ''
-                decstr = ''
+        rastr = self.compound_meta([hdr], 'ra')
+        decstr = self.compound_meta([hdr], 'dec')
+
         # Create a coordinate
         if len(rastr) > 0 and len(decstr) > 0:
             coord = SkyCoord(rastr, decstr, unit=(units.hourangle, units.deg))
@@ -743,7 +745,7 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
[…]
