### 2020-06-28 40ef271
Fixing bugs for GNIRS_10L when orders shift.
