### 2020-06-28 0df2e37
Fixing bugs for GNIRS_10L when orders shift.
