### 2020-06-28 e43b057
Fixing bugs for GNIRS_10L when orders shift.
