### 2020-06-29 4478d43
update gui doc + lamps
