+++ pypeit/scripts/launch_identify.py
@@ -65,7 +65,7 @@ def main(args):
     par = spec.default_pypeit_par()['calibrations']['wavelengths']
 
     # Get the lamp list
-    if args.lamps == '':
+    if args.lamps is None:
         lamplist = par['lamps']
         if lamplist is None:
             print("ERROR :: Cannot determine the lamps")
