+++ doc/wave_calib.rst
@@ -109,11 +109,40 @@ calibration frame.
 pypeit_identify
 +++++++++++++++
 
+usage
+-----
+
+Here is the usage::
+
+    usage: pypeit_identify [-h] [--lamps LAMPS] [-s] [--wmin WMIN] [--wmax WMAX]
+                           [--slit SLIT] [--det DET] [--rmstol RMSTOL]
+                           arc_file slits_file
+
+    Launch PypeIt identify tool, display extractedMasterArc, and load linelist.Run
+    above the Masters/ folder
+
+    positional arguments:
+      arc_file         PypeIt MasterArc file
+      slits_file       PypeIt MasterSlits file
+
+    optional arguments:
+      -h, --help       show this help message and exit
+      --lamps LAMPS    Comma separated list of calibration lamps (no spaces)
+                       (default: None)
+      -s, --solution   Load a wavelength solution from the arc_file (if it exists)
+                       (default: False)
+      --wmin WMIN      Minimum wavelength range (default: 3000.0)
+      --wmax WMAX      Maximum wavelength range (default: 10000.0)
+      --slit SLIT      Which slit to load for wavelength calibration (default: 0)
+      --det DET        Detector index (default: 1)
+      --rmstol RMSTOL  RMS tolerance (default: 0.1)
+
+
 To launch the GUI, use the following command:
 
 .. code-block:: bash
 
-    pypeit_identify MasterArc_A_1_01.fits
+    pypeit_identify MasterArc_A_1_01.fits MasterSlits_A_1_01.fits.gz
 
 Instructions on how to use this GUI are available by pressing
 the '?' key while hovering your mouse over the plotting window.
