+++ doc/installing.rst
@@ -38,6 +38,7 @@ We recommend that you use `Anaconda <https://www.continuum.io/downloads/>`_ to i
 * `scikit-learn <https://scikit-learn.org/stable/>`_ -- version 0.20 or later
 * `IPython <https://ipython.org>`_ -- version 7.2.0 or later
 * `extension_helpers <https://pypi.org/project/extension-helpers/>`_ -- version 0.1 or later
+* `packaging <https://pypi.org/project/packaging/>`_ -- version 19.0 or later
 
 If you are using Anaconda, you can check the presence of these packages with::
 
