### 2020-06-29 pypeit/PypeIt#949 @profxj
@Gabriel-p -- Have made these edits:

- Added `packaging` to the depenency list
- Added a verison for `PyQT5`
- Edited the `linetools` installation note
- Commented out `pytest` in the `requirements`

Note that `extension_helpers` is written correctly

Last, we are aiming to have a conda bundle and/or
a `docker` for distribution.  Also, a standard `pip`
install with dependencies grabs everything, e.g.
`pip install -r path/requirements.txt`

Please continue with your review.
