+++ pypeit/pypeit.py
@@ -15,7 +15,8 @@ from pypeit.images import buildimage
 from pypeit import ginga
 from pypeit import reduce
 from pypeit import spec2dobj
-from pypeit.core import qa
+from pypeit.core import qa, flat
+from pypeit import utils
 from pypeit import specobjs
 from pypeit.spectrographs.util import load_spectrograph
 from pypeit import slittrace
@@ -657,10 +658,17 @@ class PypeIt(object):
         # Prep for manual extraction (if requested)
         manual_extract_dict = self.fitstbl.get_manual_extract(frames, det)
 
-        skymodel, objmodel, ivarmodel, outmask, sobjs, waveImg, tilts = self.redux.run(
+        skymodel, objmodel, ivarmodel, outmask, sobjs, scaleImg, waveImg, tilts = self.redux.run(
             std_trace=std_trace, manual_extract_dict=manual_extract_dict, show_peaks=self.show,
             basename=self.basename, ra=self.fitstbl["ra"][frames[0]], dec=self.fitstbl["dec"][frames[0]],
             obstime=self.obstime)
+
+        # Apply the scaleImg
+        if scaleImg.size != 1:
+            sciScl, varScl = flat.flatfield(sciImg.image.copy(), scaleImg, sciImg.fullmask,
+                                            varframe=utils.inverse(sciImg.ivar.copy()))
+            sciImg.image = sciScl.copy()
+            sciImg.ivar = utils.inverse(varScl)
 
         # TODO -- Save the slits yet again?
 
@@ -677,6 +685,7 @@ class PypeIt(object):
                                         skymodel=skymodel,
                                         objmodel=objmodel,
                                         ivarmodel=ivarmodel,
+                                        scaleimg=scaleImg,
                                         waveimg=waveImg,
                                         bpmmask=outmask,
                                         detector=sciImg.detector,
