### 2020-06-30 343a502
message and logical not
