+++ pypeit/spectrographs/spectrograph.py
@@ -47,6 +47,9 @@ from pypeit.par import pypeitpar
 
 from IPython import embed
 
+# TODO: Create an EchelleSpectrograph derived class that holds all of
+# the echelle specific methods.
+
 class Spectrograph(object):
     """
     Abstract base class whose derived classes dictate
@@ -873,22 +876,22 @@ class Spectrograph(object):
         """
         pass
 
-    def order_vec(self, slit_spat_pos):
-        """
-        Convert an array of slit_spat_pos values to order numbers
-
-        Args:
-            slit_spat_pos (np.ndarray): Slit positions
-
-        Returns:
-            np.ndarray: Order numbers
-
-        """
-        order_vec = np.zeros(slit_spat_pos.size, dtype=int)
-        for kk, ipos in enumerate(slit_spat_pos):
-            order_vec[kk], indx= self.slit2order(ipos)
-        # Return
-        return order_vec
+#    def order_vec(self, slit_spat_pos):
+#        """
+#        Convert an array of slit_spat_pos values to order numbers
+#
+#        Args:
+#            slit_spat_pos (np.ndarray): Slit positions
+#
+#        Returns:
+#            np.ndarray: Order numbers
+#
+#        """
+#        order_vec = np.zeros(slit_spat_pos.size, dtype=int)
+#        for kk, ipos in enumerate(slit_spat_pos):
+#            order_vec[kk], indx= self.slit2order(ipos)
+#        # Return
+#        return order_vec
 
     @property
     def norders(self):
@@ -926,36 +929,39 @@ class Spectrograph(object):
     def loglam_minmax(self):
         return None
 
-    def slit2order(self, slit_spat_pos):
-        """
-        This routine is only for fixed-format echelle spectrographs.
-        It returns the order of the input slit based on its slit_pos
-
-        Args:
-            slit_spat_pos (float):  Slit position (spatial at 1/2 the way up)
-
-        Returns:
-            order, indx
-
-            order (int): order number
-            indx  (int): order index
-
-        """
-        indx = np.arange(self.norders)
-        # Find closest
-        try:
-            iorder = [np.argmin(np.abs(slit-self.order_spat_pos)) for slit in slit_spat_pos]
-        except TypeError:
-            iorder = np.argmin(np.abs(slit_spat_pos-self.order_spat_pos))
-
-        # Check
-        if np.any(np.abs(self.order_spat_pos[iorder] - slit_spat_pos) > self.match_tol_spat_pos):
-            msgs.warn("Bad echelle format for VLT-XSHOOTER or you are performing a 2-d coadd with different order locations."
-                      "Returning order vector with the same number of orders you requested")
-            iorder = np.arange(slit_spat_pos.size)
-            return self.orders[iorder], indx[iorder]
-        else:
-            return self.orders[iorder], indx[iorder]
+#    def slit2order(self, slit_spat_pos):
+#        """
+#        This routine is only for fixed-format echelle spectrographs.
+#        It returns the order of the input slit based on its slit_pos
+#
+#        Args:
+#            slit_spat_pos (float):  Slit position (spatial at 1/2 the way up)
+#
+#        Returns:
+#            order, indx
+#
+#            order (int): order number
+#            indx  (int): order index
+#
+#        """
+#        indx = np.arange(self.norders)
+#        # Find closest
+#        try:
+#            iorder = [np.argmin(np.abs(slit-self.order_spat_pos)) for slit in slit_spat_pos]
+#        except TypeError:
+#            iorder = np.argmin(np.abs(slit_spat_pos-self.order_spat_pos))
+#
+#        embed(header='line 954 spectrograph.py')
+#        exit()
+#
+#        # Check
[…]
