+++ pypeit/edgetrace.py
@@ -169,9 +169,14 @@ class EdgeTraceBitMask(BitMask):
                    ('MASKINSERT', 'Trace was inserted based on drilled slit-mask locations'),
                  ('ORPHANINSERT', 'Trace was inserted to match an orphaned edge'),
                     ('SHORTSLIT', 'Slit formed by left and right edge is too short'),
+                      ('BOXSLIT', 'Slit formed by left and right edge is valid (large enough '
+                                  'to be a valid slit), but too short to be a science slit'),
                  ('ABNORMALSLIT', 'Slit formed by left and right edge has abnormal length'),
                    ('USERRMSLIT', 'Slit removed by user'),
-        ])
+                      ('NOORDER', 'Unable to associate this trace with an echelle order (echelle '
+                                  ' spectrographs only)'),
+                ('ORDERMISMATCH', 'Slit traces are not well matched to any echelle order (echelle '
+                                  ' spectrographs only)')])
         super(EdgeTraceBitMask, self).__init__(list(mask.keys()), descr=list(mask.values()))
 
     @property
@@ -190,10 +195,18 @@ class EdgeTraceBitMask(BitMask):
         return ['USERINSERT', 'SYNCINSERT', 'MASKINSERT', 'ORPHANINSERT']
 
     @property
+    def order_flags(self):
+        """
+        List of flags related to the echelle order.
+        """
+        return ['NOORDER', 'ORDERMISMATCH']
+
+    @property
     def exclude_flags(self):
         # We will not exclude SYNCINSERT slits edges from the masking
         #   These can easily crop up for star boxes and at the edge of the detector
-        return ['USERINSERT', 'MASKINSERT', 'ORPHANINSERT', 'SHORTSLIT']
+#        return ['USERINSERT', 'MASKINSERT', 'ORPHANINSERT', 'BOXSLIT']
+        return self.insert_flags + self.order_flags + ['BOXSLIT']
 
 
 # TODO -- Make a DataContainer
@@ -418,6 +431,11 @@ class EdgeTraceSet(object):
         self.binning = None             # Detector ordered:  spec, spat
 
         self.traceid = None             # The ID numbers for each trace
+        self.orderid = None             # For echelle spectrographs, this is the matched order ID
+                                        # of each trace. Left traces are negative, right traces
+                                        # are positive, and the absolute value of both is the
+                                        # echelle order.  See match_order().
+
         self.spat_img = None            # (Integer) Pixel nearest the slit edge for each trace
         self.spat_cen = None            # (Floating-point) Spatial coordinate of the slit edges
                                         # for each spectral pixel
@@ -458,6 +476,7 @@ class EdgeTraceSet(object):
         Convenience method to set all attributes related to trace data to `None`.
         """
         self.traceid = None
+        self.orderid = None
         self.spat_img = None
         self.spat_cen = None
         self.spat_err = None
@@ -686,7 +705,8 @@ class EdgeTraceSet(object):
         # Perform the initial edge detection and trace identification
         self.initial_trace(bpm=bpm, det=det, binning=binning, save=False)
         if show_stages:
-            self.show(thin=10, include_img=True, idlabel=True)
+            self.show(thin=10, include_img=True, idlabel=True,
+                      title='Initial identification and tracing of slit edges')
 
         # Initial trace can result in no edges found
         if not self.is_empty:
@@ -694,8 +714,8 @@ class EdgeTraceSet(object):
             # features in the Sobel-filtered image.
             self.centroid_refine()
             if show_stages:
-                self.show(thin=10, include_img=True, idlabel=True)
-
+                self.show(thin=10, include_img=True, idlabel=True,
[…]
