### 2020-07-01 e62d108
rm match_tol_spat_pos
