+++ pypeit/reduce.py
@@ -15,6 +15,7 @@ from astropy.io import fits
 from abc import ABCMeta
 
 from linetools import utils as ltu
+from scipy import interpolate
 
 from pypeit import specobjs, specobj
 from pypeit import ginga, msgs, utils
@@ -127,7 +128,7 @@ class Reduce(object):
         self.par = par
         self.caliBrate = caliBrate
         self.std_outfile = std_outfile
-        self.scaleimg = np.array([1])  # np.array([1]) applies no scale
+        self.scaleimg = np.array([1.0], dtype=np.float)  # np.array([1]) applies no scale
         # Parse
         # Slit pieces
         #   WARNING -- It is best to unpack here then pass around self.slits
@@ -147,6 +148,7 @@ class Reduce(object):
             msgs.error("Not ready for this objtype in Reduce")
 
         # Initialise the slits
+        msgs.info("Initialising slits")
         self.initialise_slits()
 
         # Internal bpm mask
@@ -341,6 +343,7 @@ class Reduce(object):
             tilt_flexure_shift = _spat_flexure - self.waveTilts.spat_flexure
         else:
             tilt_flexure_shift = self.spat_flexure_shift
+        msgs.info("Generating tilts image")
         self.tilts = self.waveTilts.fit2tiltimg(self.slitmask, flexure=tilt_flexure_shift)
 
         # Wavelengths (on unmasked slits)
@@ -545,6 +548,23 @@ class Reduce(object):
             # Mask if something went wrong
             if np.sum(self.global_sky[thismask]) == 0.:
                 self.reduce_bpm[slit_idx] = True
+
+        # wavefull=np.linspace(3950,4450,10000)
+        # import matplotlib.pylab as pl
+        # from matplotlib import pyplot as plt
+        # colors = pl.cm.jet(np.linspace(0, 1, gdslits.size))
+        # for sl, slit_idx in enumerate(gdslits):
+        #     slit_spat = self.slits.spat_id[slit_idx]
+        #     thismask = self.slitmask == slit_spat
+        #     wav = self.waveimg[thismask]
+        #     flx = self.global_sky[thismask]
+        #     argsrt = np.argsort(wav)
+        #     spl = interpolate.interp1d(wav[argsrt], flx[argsrt], bounds_error=False)
+        #     if sl == 0:
+        #         ref = spl(wavefull)
+        #         plt.plot(wavefull, ref/np.nanmedian(ref), color=colors[sl], linestyle=':')
+        #     plt.plot(wavefull, spl(wavefull)/ref, color=colors[sl])
+        # plt.show()
 
         if update_crmask:
             # Find CRs with sky subtraction
@@ -1188,6 +1208,39 @@ class IFUReduce(MultiSlitReduce, Reduce):
         else:
             return None, None, None
 
+    def apply_relative_scale(self, scaleImg):
+        """Apply a relative scale to the science frame (and correct the varframe, too)
+
+         Args:
+             scaleImg (np.ndarray):
+                scale image to divide the science frame by
+        """
+        # Check that scaleimg is set to the correct shape
+        if self.scaleimg.size == 1:
+            self.scaleimg = np.ones_like(self.sciImg.image)
+        # Correct the relative illumination of the science frame
+        msgs.info("Correcting science frame for relative spectral illumination")
+        scaleFact = scaleImg + (scaleImg == 0)
+        self.scaleimg *= scaleFact
+        sciImg, varImg = flat.flatfield(self.sciImg.image.copy(), scaleFact, self.sciImg.fullmask,
+                                        varframe=utils.inverse(self.sciImg.ivar.copy()))
+        self.sciImg.image = sciImg.copy()
+        self.sciImg.ivar = utils.inverse(varImg)
+        return
+
+    def get_closest(self, array, values):
[…]
