+++ pypeit/reduce.py
@@ -198,7 +198,7 @@ class Reduce(object):
 
         # Slitmask
         self.slitmask = self.slits.slit_img(initial=initial, flexure=self.spat_flexure_shift,
-                                           exclude_flag=self.slits.bitmask.exclude_for_reducing)
+                                            exclude_flag=self.slits.bitmask.exclude_for_reducing)
         # Now add the slitmask to the mask (i.e. post CR rejection in proc)
         # NOTE: this uses the par defined by EdgeTraceSet; this will
         # use the tweaked traces if they exist
@@ -650,7 +650,7 @@ class Reduce(object):
                 # Get the regions
                 status, regions = skysub.read_userregions(skyregtxt, self.slits.nslits, maxslitlength)
                 # Generate image
-                skymask_init = skysub.generate_mask(self.pypeline, regions, self.slits, self.slits_left, self.slits_right)
+                skymask_init = skysub.generate_mask(self.pypeline, regions, self.slits, self.slits_left, self.slits_right, spat_flexure=self.spat_flexure_shift)
                 usersky = True
         return skymask_init, usersky
 
@@ -1228,7 +1228,7 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.sciImg.ivar = utils.inverse(varImg)
         return
 
-    def get_closest(self, array, values):
+    def find_nearest(self, array, values):
         # make sure array is a numpy array
         array = np.array(array)
 
@@ -1275,8 +1275,8 @@ class IFUReduce(MultiSlitReduce, Reduce):
         rawimg = self.sciImg.image.copy()
         # Grab the GPM and the slit images
         gpm = (self.sciImg.fullmask == 0)
-        slitid_img_init = self.slits.slit_img(pad=0, initial=True)
-        slitid_img_trim = self.slits.slit_img(pad=-trim, initial=True)
+        slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
+        slitid_img_trim = self.slits.slit_img(pad=-trim, initial=True, flexure=self.spat_flexure_shift)
         # Find all good slits, and create a mask of pixels to include (True=include)
         wgd = self.slits.spat_id[np.where(self.slits.mask == 0)]
         # Obtain the minimum and maximum wavelength of all slits
@@ -1293,23 +1293,35 @@ class IFUReduce(MultiSlitReduce, Reduce):
         # Go through the slits and calculate the overlapping flux
         maxiter = 20
         scaleImg = np.ones_like(self.sciImg.image)
+        dwav = self.wv_calib[str(self.slits.spat_id[0])]['cen_disp']
         for rr in range(maxiter):
             relscl_model = np.ones_like(rawimg)
             # Prepare slit 0
             onslit_a = (slitid_img_trim == self.slits.spat_id[swslt[0]])
             onslit_a_olap = onslit_a & gpm & (self.waveimg >= minw) & (self.waveimg <= maxw) & skymask
             wsort = np.argsort(self.waveimg[onslit_a_olap])
-            # Generate a simple, quick spline for the first slit
-            wave_ref = self.waveimg[onslit_a_olap][wsort]
-            sky_ref = rawimg[onslit_a_olap][wsort]
-            #sclspl = interpolate.interp1d(, , kind='linear', bounds_error=False, fill_value="extrapolate")
+            # Generate a master sky frame
+            if rr == 0:
+                numsamp = gdslits.size * int((maxw - minw) / dwav)
+                bins = np.linspace(minw, maxw, numsamp)
+                hist, edge = np.histogram(self.waveimg, bins=bins, weights=global_sky_sep)
[…]
