+++ pypeit/tests/test_chk_calibs.py
@@ -21,14 +21,15 @@ def test_chk_calibs_not():
     droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/not_alfosc/grism4')
     droot += '/ALD'
     pargs = chk_calibs.parser(['-r', droot, '-s', 'not_alfosc'])
-    answers = chk_calibs.main(pargs)
+    answers, _ = chk_calibs.main(pargs)
 
-    assert np.all(answers['passfail']), 'One or more failures!'
+    assert answers['pass'][0], 'One or more failures!'
 
 @dev_suite_required
 def test_chk_calibs_deimos():
     droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/keck_deimos/*/')
     pargs = chk_calibs.parser(['-r', droot, '-s', 'keck_deimos'])
-    answers = chk_calibs.main(pargs)
+    answers, _ = chk_calibs.main(pargs)
 
-    assert np.all(answers['passfail']), 'One or more failures!'
+    assert np.all(answers['pass']), 'One or more failures!'
+
