### 2020-07-07 177c274
spatial+spectral scaling
