+++ pypeit/reduce.py
@@ -16,11 +16,12 @@ from abc import ABCMeta
 
 from linetools import utils as ltu
 from scipy import interpolate
+from scipy.optimize import least_squares
 
 from pypeit import specobjs, specobj
 from pypeit import ginga, msgs, utils
 from pypeit import masterframe
-from pypeit.core import skysub, extract, wave, flexure, flat
+from pypeit.core import skysub, extract, pixels, wave, flexure, flat
 from pypeit.images import buildimage
 from pypeit import wavecalib
 
@@ -1262,105 +1263,157 @@ class IFUReduce(MultiSlitReduce, Reduce):
         """
         # Generate a global sky sub for all slits separately
         global_sky_sep = Reduce.global_skysub(self, skymask=skymask, update_crmask=update_crmask, trim_edg=trim_edg,
-                                               show_fit=show_fit, show=show, show_objs=show_objs)
-        # If a joint fit is not being performed, return the separate slits result
-        if not self.par['reduce']['skysub']['joint_fit']:
+                                              show_fit=show_fit, show=show, show_objs=show_objs)
+        # If the joint fit or spec/spat sensitivity corrections are not being performed, return the separate slits sky
+        if not self.par['reduce']['skysub']['joint_fit'] and \
+                not self.par['scienceframe']['process']['use_specillum'] and \
+                not self.par['scienceframe']['process']['use_illumflat']:
             return global_sky_sep
 
-        msgs.info("Performing joint global sky subtraction")
-        # Grab some parameters
+        # Extract some parameters
         gdslits = np.where(np.logical_not(self.reduce_bpm))[0]
         trim = self.par['calibrations']['flatfield']['slit_trim']
         spec_samp_fine = self.par['calibrations']['flatfield']['spec_samp_coarse']
-        rawimg = self.sciImg.image.copy()
-        # Grab the GPM and the slit images
+        applyCorrection = False  # Boolean flag to see if spatial or spectral corrections have been performed
+
+        # Get the GPM and the slit images
         gpm = (self.sciImg.fullmask == 0)
         slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
         slitid_img_trim = self.slits.slit_img(pad=-trim, initial=True, flexure=self.spat_flexure_shift)
-        # Find all good slits, and create a mask of pixels to include (True=include)
-        wgd = self.slits.spat_id[np.where(self.slits.mask == 0)]
-        # Obtain the minimum and maximum wavelength of all slits
-        mnmx_wv = np.zeros((self.slits.nslits, 2))
-        for slit_idx, slit_spat in enumerate(self.slits.spat_id):
-            onslit_init = (slitid_img_init == slit_spat)
-            mnmx_wv[slit_idx, 0] = np.min(self.waveimg[onslit_init])
-            mnmx_wv[slit_idx, 1] = np.max(self.waveimg[onslit_init])
-        # Sort by increasing minimum wavelength
-        swslt = np.argsort(mnmx_wv[:, 0])
-        # Get the minimum and maximum wavelengths that are covered in *all* slits.
-        minw, maxw = np.max(mnmx_wv, axis=0)[0], np.min(mnmx_wv, axis=0)[1]
 
-        # Go through the slits and calculate the overlapping flux
-        maxiter = 20
-        scaleImg = np.ones_like(self.sciImg.image)
-        dwav = self.wv_calib[str(self.slits.spat_id[0])]['cen_disp']
-        for rr in range(maxiter):
-            relscl_model = np.ones_like(rawimg)
-            # Prepare slit 0
-            onslit_a = (slitid_img_trim == self.slits.spat_id[swslt[0]])
[…]
