### 2020-07-08 edf6a2d
updates to KCWI file
