+++ pypeit/reduce.py
@@ -1242,24 +1242,190 @@ class IFUReduce(MultiSlitReduce, Reduce):
 
         return idxs
 
+    def illum_profile_spatial(self, skymask=None, trim_edg=(0,0)):
+        msgs.info("Performing spatial sensitivity correction")
+        # Setup some helpful parameters
+        skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
+        hist_trim = 3  # Trim the edges of the histogram to take into account edge effects
+        gpm = (self.sciImg.fullmask == 0)
+        slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
+        spatScaleImg = np.ones_like(self.sciImg.image)
+        # For each slit, grab the spatial coordinates and a spline
+        # representation of te spatial profile from the illumflat
+        rawimg = self.sciImg.image.copy()
+        numbins = int(np.max(self.slits.get_slitlengths(initial=True, median=True)))
+        spatbins = np.linspace(0.0, 1.0, numbins + 1)
+        spat_slit = 0.5 * (spatbins[1:] + spatbins[:-1])
+        for sl, slitnum in enumerate(self.slits.spat_id):
+            msgs.info("Deriving spatial correction for slit {0:d}/{1:d}".format(sl + 1, self.slits.spat_id.size))
+            # Get the initial slit locations
+            onslit_b_init = (slitid_img_init == slitnum)
+
+            # Synthesize ximg, and edgmask from slit boundaries. Doing this outside this
+            # routine would save time. But this is pretty fast, so we just do it here to make the interface simpler.
+            spatcoord, edgmask = pixels.ximg_and_edgemask(self.slits_left[:, sl], self.slits_right[:, sl],
+                                                          onslit_b_init, trim_edg=trim_edg)
+            # Ignore skymask
+            coord_msk = onslit_b_init & gpm
+            hist, _ = np.histogram(spatcoord[coord_msk], bins=spatbins, weights=rawimg[coord_msk])
+            cntr, _ = np.histogram(spatcoord[coord_msk], bins=spatbins)
+            hist_slit_all = hist / (cntr + (cntr == 0))
+
+            # Repeat with skymask
+            coord_msk = onslit_b_init & gpm & skymask_now
+            hist, _ = np.histogram(spatcoord[coord_msk], bins=spatbins, weights=rawimg[coord_msk])
+            cntr, _ = np.histogram(spatcoord[coord_msk], bins=spatbins)
+            hist_slit = hist / (cntr + (cntr == 0))
+
+            # Prepare for fit - take the non-zero elements and trim slit edges
+            ww = (hist_slit[hist_trim:-hist_trim] != 0)
+            xfit = spat_slit[hist_trim:-hist_trim][ww]
+            yfit = hist_slit_all[hist_trim:-hist_trim][ww]
+
+            # Generate the model
+            model = self.caliBrate.flatimages.illumflat_spat_bsplines[sl].value(xfit)[0]
+            modev = self.caliBrate.flatimages.illumflat_spat_bsplines[sl].value(spatcoord[onslit_b_init])[0]
+            xnorm = self.caliBrate.flatimages.illumflat_spat_bsplines[sl].value(np.array([0.5]))[0][0]
+
+            # Fit the function
+            spat_func = lambda par, ydata, model: par[0] + par[1] * model - ydata
+            res_lsq = least_squares(spat_func, [np.median(yfit), 0.0], args=(yfit, model))
+            spatnorm = spat_func(res_lsq.x, 0.0, modev)
+            spatnorm /= spat_func(res_lsq.x, 0.0, xnorm)
[…]
