### 2020-07-09 be26de0
edits to KW, no bias for DEIMOS, CHANGES
