+++ CHANGES.rst
@@ -4,6 +4,8 @@
 --------
 
  - Fix class and version check for DataContainer objects.
+ - Script to check for calibration files
+ - No longer require bias frames as default for DEIMOS
 
 1.0.5 (23 Jun 2020)
 -------------------
