+++ pypeit/spectrographs/keck_lris.py
@@ -380,6 +380,18 @@ class KeckLRISBSpectrograph(KeckLRISSpectrograph):
         self.sky_file = 'sky_LRISb_600.fits'
 
     def get_detector_par(self, hdu, det):
+        """
+
+        Args:
+            hdu (`astropy.io.fits.HDUList`):
+                HDUList of the image of interest.
+                Ought to be the raw file, or else..
+            det (int):
+
+        Returns:
+            :class:`pypeit.images.detector_container.DetectorContainer`:
+
+        """
         # Binning
         binning = self.get_meta_value(self.get_headarr(hdu), 'binning')  # Could this be detector dependent??
 
@@ -554,19 +566,7 @@ class KeckLRISRSpectrograph(KeckLRISSpectrograph):
         super(KeckLRISRSpectrograph, self).__init__()
         self.spectrograph = 'keck_lris_red'
         self.camera = 'LRISr'
-#        # Allow for variable namps
-#        if ifile is not None:
-#            hdu = fits.open(ifile)
-#            head0 = hdu[0].header
-#            if head0['AMPPSIZE'] == '[1:1024,1:4096]':
-#                namps = 2
-#            elif head0['AMPPSIZE'] == '[1:2048,1:4096]':
-#                namps = 1
-#            else:
-#                msgs.error("Bad amp size (windowed??)!!")
-#        else: # Default
-#            namps = 2
-        #
+
     def get_detector_par(self, hdu, det):
         # Binning
         binning = self.get_meta_value(self.get_headarr(hdu), 'binning')  # Could this be detector dependent??
@@ -781,56 +781,111 @@ class KeckLRISRSpectrograph(KeckLRISSpectrograph):
         return bpm_img
 
 
-#class KeckLRISRLSpectrograph(KeckLRISRSpectrograph):
-#    """
-#    Child to handle Keck/LRISr in Long-slit readout mode (Vid1, Vid4)
-#    """
-#    def __init__(self):
-#        # Get it started
-#        super(KeckLRISRSpectrograph, self).__init__()
-#        self.spectrograph = 'keck_lris_red_longonly'
-#        self.camera = 'LRISr'
-#        self.detector = [
-#                # Detector 1
-#                pypeitpar.DetectorPar(
-#                            dataext         =1,
-#                            specaxis        =0,
-#                            specflip        = False,
-#                            xgap            =0.,
-#                            ygap            =0.,
-#                            ysize           =1.,
-#                            platescale      =0.135,
-#                            darkcurr        =0.0,
-#                            saturation      =65535.,  # Gain applied
-#                            nonlinear       =0.86,          # Modified by JXP to go higher
-#                            numamplifiers   =1,
-#                            gain            =[1.255],
-#                            ronoise         =[4.64],
-#                            datasec         = ['',''],
-#                            oscansec        = ['',''],
-#                            suffix          ='_01red'
-#                            ),
-#                #Detector 2
-#                pypeitpar.DetectorPar(
-#                            dataext         =2,
-#                            specaxis        =0,
-#                            specflip        = False,
-#                            xgap            =0.,
-#                            ygap            =0.,
-#                            ysize           =1.,
-#                            platescale      =0.135,
-#                            darkcurr        =0.,
-#                            saturation      =65535.,  # Gain applied
-#                            nonlinear       =0.86,
-#                            numamplifiers   =1,
-#                            gain            =[1.162],
[…]
