+++ pypeit/defs.py
@@ -2,7 +2,7 @@
 """
 # No imports to avoid circularity
 
-pypeit_spectrographs = ['keck_deimos', 'keck_lris_blue', 'keck_lris_red', 'keck_nires',
+pypeit_spectrographs = ['keck_deimos', 'keck_lris_blue', 'keck_lris_red', 'keck_lris_red_orig', 'keck_nires',
             'keck_nirspec_low', 'keck_mosfire', 'keck_hires_red', 'keck_kcwi',
             'shane_kast_blue', 'shane_kast_red', 'shane_kast_red_ret', 'tng_dolores', 'wht_isis_blue',
             'wht_isis_red', 'vlt_xshooter_uvb', 'vlt_xshooter_vis', 'vlt_xshooter_nir', 'vlt_fors2',
