+++ pypeit/scripts/coadd_datacube.py
@@ -9,20 +9,17 @@ and define the sky background regions interactively.
 Run above the Science/ folder.
 """
 
-import os
 import argparse
 from IPython import embed
 
 from astropy.io import fits
-from linetools import utils as ltu
 import numpy as np
 
 from pypeit import msgs
 from pypeit.spectrographs.util import load_spectrograph
 from pypeit.core.parse import get_dnum
-from pypeit.images.imagebitmask import ImageBitMask
+from pypeit.core.flux_calib import load_extinction_data, extinction_correction
 from pypeit import alignframe
-from pypeit import masterframe
 from pypeit import spec2dobj
 
 
@@ -39,6 +36,80 @@ def parser(options=None):
                         help='Overwrite any existing files/directories')
 
     return parser.parse_args() if options is None else parser.parse_args(options)
+
+from astropy.coordinates import AltAz, SkyCoord
+import astropy.units as units
+import scipy.optimize as opt
+from scipy.interpolate import interp1d
+
+
+def dar_fitfunc(radec, coord_ra, coord_dec, datfit, wave, obstime, location, pressure, temperature, rel_humidity):
+    (diff_ra, diff_dec) = radec
+    # Generate the coordinate with atmopheric conditions
+    coord_atmo = SkyCoord(coord_ra + diff_ra, coord_dec + diff_dec, unit=(units.deg, units.deg))
+    coord_altaz = coord_atmo.transform_to(AltAz(obstime=obstime, location=location, obswl=wave,
+                                         pressure=pressure, temperature=temperature,
+                                         relative_humidity=rel_humidity))
+    # Return chi-squared value
+    return np.sum((np.array([coord_altaz.alt.value, coord_altaz.az.value])-datfit)**2)
+
+
+def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, rel_humidity):
+    """Apply a differental atmospheric refraction correction to the input ra/dec.
+    This implementation is based on ERFA, which is called through astropy
+
+    wave_arr (ndarray):
+        wavelengths to obtain ra and dec offsets
+    coord (astropy SkyCoord):
+        ra, dec positions at the centre of the field
+    obstime (astropy Time):
+        time at the midpoint of observation
+    location (astropy EarthLocation):
+        observatory location
+    pressure (float):
+        Outside pressure at `location`
+    temperature (float):
+        Outside ambient air temperature at `location`
+    rel_humidity (float):
+        Outside relative humidity at `location`. This should be between 0 to 1.
+
+    Returns
+    -------
+    ra_diff (ndarray):
+        Relative RA shift at each wavelength given by `wave_arr`
+    dec_diff (ndarray):
+        Relative DEC shift at each wavelength given by `wave_arr`
+
+    TODO :: There's probably going to be issues when the RA angle is either side of RA=0
+    TODO :: Move this routine to the main PypeIt code
+    """
+    msgs.info("Performing differential atmospheric refraction correction")
+    numgrid = 100
+    # First create the reference frame and wavelength grid
+    coord_altaz = coord.transform_to(AltAz(obstime=obstime, location=location))
[…]
