+++ CHANGES.rst
@@ -6,6 +6,7 @@
  - Fix class and version check for DataContainer objects.
  - Script to check for calibration files
  - No longer require bias frames as default for DEIMOS
+ - Adds in Keck/LRISr with the original detector
 
 1.0.5 (23 Jun 2020)
 -------------------
