+++ doc/spectrographs.rst
@@ -19,6 +19,7 @@ gemini_gnirs            Gemini      GNIRS spectrometer
 keck_kcwi               Keck        KCWI slit-based IFU (BM, BH2)
 keck_lris_blue          Keck        :doc:`lris` spectrometer; blue camera
 keck_lris_red           Keck        :doc:`lris` spectrometer; red camera
+keck_lris_red_orig      Keck        :doc:`lris` spectrometer; red camera + original detector
 keck_mosfire            Keck        MOSFIRE spectrometer; J and Y gratings tested
 keck_nires              Keck        NIRES spectrometer
 keck_nirspec_low        Keck        NIRSPEC spectrometer; low-dispersion
