+++ pypeit/utils.py
@@ -1910,6 +1910,33 @@ def subsample(frame):
     return frame[tuple(indices)]
 
 
+def find_nearest(array, values):
+    """For all elements of values, find the index of the nearest value in array
+
+    Parameters
+    ----------
+    array : numpy.ndarray
+        Array of values
+    values : numpy.ndarray
+        Values to be compared with the elements of `array`
+
+    Return
+    ------
+    numpy.ndarray : indices of `array` that are closest to each element of value
+    """
+    # Make sure the input is a numpy array
+    array = np.array(array)
+
+    # get insert positions
+    idxs = np.searchsorted(array, values, side="left")
+
+    # find indexes where previous index is closer
+    prev_idx_is_less = ((idxs == len(array)) | (np.fabs(values - array[np.maximum(idxs - 1, 0)]) <
+                                                np.fabs(values - array[np.minimum(idxs, len(array) - 1)])))
+    idxs[prev_idx_is_less] -= 1
+
+    return idxs
+
 
 def yamlify(obj, debug=False):
     """
