+++ pypeit/flatfield.py
@@ -1226,6 +1226,7 @@ class FlatField(object):
         if gpm is None:
             gpm = np.ones_like(rawflat, dtype=bool) if self.rawflatimg.bpm is None else (
                     1 - self.rawflatimg.bpm).astype(bool)
+
         slitid_img_init = self.slits.slit_img(pad=0, initial=True)
         slitid_img_trim = self.slits.slit_img(pad=-trim, initial=True)
         # Find all good slits, and create a mask of pixels to include (True=include)
@@ -1240,35 +1241,8 @@ class FlatField(object):
         swslt = np.argsort(mnmx_wv[:, 0])
 
         ### STEP 1
-        # Go through the slits and calculate the overlapping flux
-        relscl_model = np.ones_like(rawflat)
-        # Get the minimum and maximum wavelengths that are covered in *all* slits.
-        minw, maxw = np.max(mnmx_wv, axis=0)[0], np.min(mnmx_wv, axis=0)[1]
-        # Prepare slit 0
-        # TODO :: Probably could do a bspline fit to slit 0... not sure about the small (< 1 pixel) extrapolation required
-        onslit_a = (slitid_img_trim == self.slits.spat_id[swslt[0]])
-        onslit_a_olap = onslit_a & gpm & (waveimg >= minw) & (waveimg <= maxw)
-        wsort = np.argsort(waveimg[onslit_a_olap])
-        # Generate a simple, quick spline for the first slit
-        sclspl = interpolate.interp1d(waveimg[onslit_a_olap][wsort], rawflat[onslit_a_olap][wsort],
-                                      kind='linear', bounds_error=False, fill_value="extrapolate")
-        # Now roughly calculate the flux scale relative to the first slit
-        # TODO :: Should iterate over this (see IFU.gloval_skysub() in reduce.py - could also use find_nearest instead of interpolating).
-        for slit_idx in range(1, self.slits.spat_id.size):
-            # Only use the overlapping regions of the slits, where the same wavelength range is covered
-            onslit_b = (slitid_img_trim == self.slits.spat_id[swslt[slit_idx]])
-            onslit_b_init = (slitid_img_init == self.slits.spat_id[swslt[slit_idx]])
-            onslit_b_olap = onslit_b & gpm & (waveimg >= minw) & (waveimg <= maxw)
-            speca = sclspl(waveimg[onslit_b_olap])
-            # Fit a low order polynomial
-            xfit = (waveimg[onslit_b_olap] - minw) / (maxw - minw)
-            yfit = speca / rawflat[onslit_b_olap]
-            # Rough outlier rejection
-            msk = (yfit > -10) & (yfit < 10)
-            coeff = utils.func_fit(xfit[msk], yfit[msk], "legendre", 2, minx=0, maxx=1)
-            relscl_model[onslit_b_init] = utils.func_val(coeff,
-                                                         (waveimg[onslit_b_init] - minw) / (maxw - minw),
-                                                         "legendre", minx=0, maxx=1)
+        relscl_model = illum_profile_spectral(rawflat, waveimg, self.slits, model=None, gpmask=gpm, skymask=None,
+                                              trim=trim, flexure=flex)
 
         ### STEP 2
         # Perform a simultaneous fit to all pixels in all slits to get a "global" shape of the flat spectrum.
@@ -1401,6 +1375,76 @@ def show_flats(image_list, wcs_match=True, slits=None):
         # Turn off clear
         if clear:
             clear = False
+
+
[…]
