+++ pypeit/reduce.py
@@ -5,22 +5,20 @@ Main driver class for skysubtraction and extraction
 .. include:: ../links.rst
 """
 
-import os
 import inspect
 import numpy as np
 import os
 
 from astropy import stats
-from astropy.io import fits
 from abc import ABCMeta
 
 from linetools import utils as ltu
 from scipy import interpolate
 from scipy.optimize import least_squares
 
-from pypeit import specobjs, specobj
+from pypeit import specobjs
 from pypeit import ginga, msgs, utils
-from pypeit import masterframe
+from pypeit import masterframe, flatfield
 from pypeit.core import skysub, extract, pixels, wave, flexure, flat
 from pypeit.images import buildimage
 from pypeit import wavecalib
@@ -1229,19 +1227,6 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.sciImg.ivar = utils.inverse(varImg)
         return
 
-    def find_nearest(self, array, values):
-        # make sure array is a numpy array
-        array = np.array(array)
-
-        # get insert positions
-        idxs = np.searchsorted(array, values, side="left")
-
-        # find indexes where previous index is closer
-        prev_idx_is_less = ((idxs == len(array)) | (np.fabs(values - array[np.maximum(idxs - 1, 0)]) < np.fabs(values - array[np.minimum(idxs, len(array) - 1)])))
-        idxs[prev_idx_is_less] -= 1
-
-        return idxs
-
     def illum_profile_spatial(self, skymask=None, trim_edg=(0,0)):
         msgs.info("Performing spatial sensitivity correction")
         # Setup some helpful parameters
@@ -1299,6 +1284,16 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.apply_relative_scale(spatScaleImg)
 
     def illum_profile_spectral(self, global_sky, skymask=None):
+        trim = self.par['calibrations']['flatfield']['slit_trim']
+        gpm = (self.sciImg.fullmask == 0)
+        scaleImg = flatfield.illum_profile_spectral(self.sciImg.image.copy(), self.waveimg, self.slits,
+                                                    model=global_sky, gpmask=gpm, skymask=skymask, trim=trim,
+                                                    flexure=self.spat_flexure_shift)
+        # Now perform the full correction to the science frame
+        msgs.info("Correcting science frame for relative spectral and spatial illumination")
+        self.apply_relative_scale(scaleImg)
+
+    def illum_profile_spectral_old(self, global_sky, skymask=None):
         msgs.info("Performing relative spectral sensitivity correction")
         # Setup some helpful parameters
         skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
@@ -1321,7 +1316,7 @@ class IFUReduce(MultiSlitReduce, Reduce):
         minw, maxw = np.max(mnmx_wv, axis=0)[0], np.min(mnmx_wv, axis=0)[1]
 
         # Go through the slits and calculate the overlapping flux
-        maxiter = 20
+        maxiter = 10
         dwav = self.wv_calib[str(self.slits.spat_id[0])]['cen_disp']
         for rr in range(maxiter):
             # Reset the relative scaling for this iteration
@@ -1344,14 +1339,14 @@ class IFUReduce(MultiSlitReduce, Reduce):
                 onslit_b = (slitid_img_trim == self.slits.spat_id[swslt[slit_idx]])
                 onslit_b_init = (slitid_img_init == self.slits.spat_id[swslt[slit_idx]])
                 onslit_b_olap = onslit_b & gpm & (self.waveimg >= minw) & (self.waveimg <= maxw) & skymask_now
[…]
