### 2020-07-11 c3879f1
rm specillum from reduce
