+++ pypeit/reduce.py
@@ -1293,89 +1293,6 @@ class IFUReduce(MultiSlitReduce, Reduce):
         msgs.info("Correcting science frame for relative spectral and spatial illumination")
         self.apply_relative_scale(scaleImg)
 
-    def illum_profile_spectral_old(self, global_sky, skymask=None):
-        msgs.info("Performing relative spectral sensitivity correction")
-        # Setup some helpful parameters
-        skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
-        gdslits = np.where(np.logical_not(self.reduce_bpm))[0]
-        gpm = (self.sciImg.fullmask == 0)
-        trim = self.par['calibrations']['flatfield']['slit_trim']
-        slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
-        slitid_img_trim = self.slits.slit_img(pad=-trim, initial=True, flexure=self.spat_flexure_shift)
-        scaleImg = np.ones_like(self.sciImg.image)
-        rawimg = self.sciImg.image.copy()
-        # Obtain the minimum and maximum wavelength of all slits
-        mnmx_wv = np.zeros((self.slits.nslits, 2))
-        for slit_idx, slit_spat in enumerate(self.slits.spat_id):
-            onslit_init = (slitid_img_init == slit_spat)
-            mnmx_wv[slit_idx, 0] = np.min(self.waveimg[onslit_init])
-            mnmx_wv[slit_idx, 1] = np.max(self.waveimg[onslit_init])
-        # Sort by increasing minimum wavelength
-        swslt = np.argsort(mnmx_wv[:, 0])
-        # Get the minimum and maximum wavelengths that are covered in *all* slits.
-        minw, maxw = np.max(mnmx_wv, axis=0)[0], np.min(mnmx_wv, axis=0)[1]
-
-        # Go through the slits and calculate the overlapping flux
-        maxiter = 10
-        dwav = self.wv_calib[str(self.slits.spat_id[0])]['cen_disp']
-        for rr in range(maxiter):
-            # Reset the relative scaling for this iteration
-            relscl_model = np.ones_like(rawimg)
-            # Generate a master sky frame
-            if rr == 0:
-                numsamp = gdslits.size * int((maxw - minw) / dwav)
-                bins = np.linspace(minw, maxw, numsamp)
-            else:
-                numsamp = gdslits.size * int((np.max(mnmx_wv) - np.min(mnmx_wv)) / dwav)
-                bins = np.linspace(np.min(mnmx_wv), np.max(mnmx_wv), numsamp)
-            # Histogram to get sky spectrum
-            hist, edge = np.histogram(self.waveimg, bins=bins, weights=global_sky)
-            cntr, edge = np.histogram(self.waveimg, bins=bins)
-            wave_ref = 0.5 * (edge[1:] + edge[:-1])
-            sky_ref = hist / cntr
-            # Using the reference sky, iterate to calculate the relative spectral response
-            for slit_idx in range(0, self.slits.spat_id.size):
-                # Only use the overlapping regions of the slits, where the same wavelength range is covered
-                onslit_b = (slitid_img_trim == self.slits.spat_id[swslt[slit_idx]])
-                onslit_b_init = (slitid_img_init == self.slits.spat_id[swslt[slit_idx]])
-                onslit_b_olap = onslit_b & gpm & (self.waveimg >= minw) & (self.waveimg <= maxw) & skymask_now
-                idxcls = utils.find_nearest(wave_ref, self.waveimg[onslit_b_olap])
-                speca = sky_ref[idxcls]
-                # Fit a low order polynomial
[…]
