### 2020-07-11 0511c14
edge sync and doc fixes
