+++ pypeit/spectrographs/keck_deimos.py
@@ -158,7 +158,7 @@ class KeckDEIMOSSpectrograph(spectrograph.Spectrograph):
         par['calibrations']['slitedges']['fit_order'] = 3
         par['calibrations']['slitedges']['minimum_slit_gap'] = 0.25
         par['calibrations']['slitedges']['minimum_slit_length_sci'] = 4.
-        par['calibrations']['slitedges']['sync_clip'] = False
+#        par['calibrations']['slitedges']['sync_clip'] = False
 
         # 1D wavelength solution
         par['calibrations']['wavelengths']['lamps'] = ['ArI','NeI','KrI','XeI']
