+++ pypeit/scripts/trace_edges.py
@@ -98,6 +98,13 @@ def main(args):
         if len(bias_files) == 0:
             bias_files = None
 
+        # Get the dark files, if requested
+        dark_rows = rdx.fitstbl.find_frames('dark', calib_ID=int(group), index=True)
+        dark_files = rdx.fitstbl.frame_paths(dark_rows)
+        dark_par = rdx.par['calibrations']['darkframe']
+        if len(dark_files) == 0:
+            dark_files = None
+
         # Set the QA path
         qa_path = rdx.qa_path
     else:
@@ -116,6 +123,9 @@ def main(args):
         bias_files = None
         bias_par = None
 
+        dark_files = None
+        dark_par = None
+
         # Set the QA path
         qa_path = os.path.join(os.path.abspath(os.path.split(files[0])[0]), 'QA')
 
@@ -133,23 +143,25 @@ def main(args):
             proc_par['process']['use_biasimage'] = False
             msbias = None
         else:
-            #biasFrame = biasframe.BiasFrame(spec, files=bias_files, det=det, par=bias_par,
-            #                                master_key=master_key, master_dir=master_dir)
-            #msbias = biasFrame.build_image()
             msbias = buildimage.buildimage_fromlist(spec, det, bias_par, bias_files)
+
+        # Get the dark frame if requested
+        if dark_files is None:
+            proc_par['process']['use_darkimage'] = False
+            msdark = None
+        else:
+            msdark = buildimage.buildimage_fromlist(spec, det, dark_par, dark_files) #, bias=msbias)
 
         msbpm = spec.bpm(files[0], det)
 
         # Build the trace image
-        #traceImage = traceimage.TraceImage(spec, files=files, det=det, par=proc_par, bias=msbias)
-        #traceImage.build_image(bias=msbias, bpm=msbpm)
         traceImage = buildimage.buildimage_fromlist(spec, det, proc_par, files, bias=msbias,
-                                                    bpm=msbpm)
+                                                    bpm=msbpm, dark=msdark)
 
         # Trace the slit edges
         t = time.perf_counter()
-        edges = edgetrace.EdgeTraceSet(traceImage, spec, trace_par, det=det, bpm=msbpm,
-                                       auto=True, debug=args.debug, show_stages=args.show,
+        edges = edgetrace.EdgeTraceSet(traceImage, spec, trace_par, bpm=msbpm, auto=True,
+                                       debug=args.debug, show_stages=args.show,
                                        qa_path=qa_path)
 
         print('Tracing for detector {0} finished in {1} s.'.format(det, time.perf_counter()-t))
