+++ pypeit/core/trace.py
@@ -1222,37 +1222,37 @@ def peak_trace(flux, ivar=None, bpm=None, trace_map=None, extract_width=None, sm
     after collapsing along the spectral axis.
 
     The image is either compressed directly or after rectification
-    using the supplied `trace_map`. The provided trace data *must*
-    have the same shape as the input `flux` image and map each
+    using the supplied ``trace_map``. The provided trace data *must*
+    have the same shape as the input ``flux`` image and map each
     spatial position as a function of spectral position. This can be
-    the output of :func:`pypeit.core.pca.pca_predict` where the
-    provided coordinates are `np.arange(flux.shape[1])`; see also
-    :func:`pypeit.edges.EdgeTracePCA.predict`. The rectification of
-    the input `flux` is done using a boxcar extraction along the
-    provided traces with a width of `extract_width`; see
-    :func:`pypeit.core.moment.moment1d`.
+    the output of :func:`~pypeit.core.pca.pca_predict` where the
+    provided coordinates are ``np.arange(flux.shape[1])``; see also
+    :func:`~pypeit.tracepca.TracePCA.predict`. The rectification of
+    the input ``flux`` data is done using a boxcar extraction along
+    the provided traces with a width of ``extract_width``; see
+    :func:`~pypeit.core.moment.moment1d`.
 
-    The (rectified) image is collapsed spectrally (see `smash_range`)
+    The (rectified) image is collapsed spectrally (see ``smash_range``)
     giving the sigma-clipped mean flux as a function of spatial
     position. Peaks are then isolated in this vector (see
-    :func:`pypeit.core.arc.detect_lines`).
+    :func:`~pypeit.core.arc.detect_lines`).
 
     Traces that pass through these peak positions are then passed to
     two iterations of :func:`fit_trace`, which both remeasures the
     centroids of the trace and fits a polynomial to those trace data.
     The first iteration determines the centroids with uniform
-    weighting, passing `fwhm=fwhm_uniform` to :func:`fit_trace`, and
-    the second uses Gaussian weighting for the centroid measurements
-    (passing `fwhm=fwhm_gaussian` to :func:`fit_trace`). The results
-    of this second iteration of :func:`fit_trace` are the data
-    returned.
+    weighting, passing ``fwhm=fwhm_uniform`` to :func:`fit_trace`,
+    and the second uses Gaussian weighting for the centroid
+    measurements (passing ``fwhm=fwhm_gaussian`` to
+    :func:`fit_trace`). The results of this second iteration of
+    :func:`fit_trace` are the data returned.
 
     Troughs in the image can also be traced, which is done by
     flipping the sign of the image about its median and then
     repeating the "peak" finding and :func:`fit_trace` iterations. If
-    troughs are fit, the traces are order with the set of peak traces
-    first (the number of which is given by the last returned object
-    of the function), followed by the trough traces.
+    troughs are fit, the traces are ordered with the set of peak
+    traces first (the number of which is given by the last returned
+    object of the function), followed by the trough traces.
 
     Args:
         flux (`numpy.ndarray`_):
@@ -1260,45 +1260,46 @@ def peak_trace(flux, ivar=None, bpm=None, trace_map=None, extract_width=None, sm
         ivar (`numpy.ndarray`_, optional):
             Inverse variance of the image intensity.  If not provided,
             unity variance is used.  If provided, must have the same
-            shape as `flux`.
+            shape as ``flux``.
         bpm (`numpy.ndarray`_, optional):
             Boolean array with the input bad-pixel mask for the
-            image. If not provided, all values in `flux` are
+            image. If not provided, all values in ``flux`` are
             considered valid. If provided, must have the same shape
-            as `flux`.
+            as ``flux``.
[…]

[…]
