### 2020-07-11 939657a
magellan_mage fix
