+++ pypeit/reduce.py
@@ -1218,7 +1218,7 @@ class IFUReduce(MultiSlitReduce, Reduce):
         if self.scaleimg.size == 1:
             self.scaleimg = np.ones_like(self.sciImg.image)
         # Correct the relative illumination of the science frame
-        msgs.info("Correcting science frame for relative spectral illumination")
+        msgs.info("Correcting science frame for relative illumination")
         scaleFact = scaleImg + (scaleImg == 0)
         self.scaleimg *= scaleFact
         sciImg, varImg = flat.flatfield(self.sciImg.image.copy(), scaleFact, self.sciImg.fullmask,
@@ -1227,7 +1227,9 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.sciImg.ivar = utils.inverse(varImg)
         return
 
-    def illum_profile_spatial(self, skymask=None, trim_edg=(0,0)):
+#    def illum_profile_spatial_fitall(self, skymask=None, trim_edg=(0,0)):
+    def illum_profile_spatial(self, skymask=None, trim_edg=(0, 0)):
+
         msgs.info("Performing spatial sensitivity correction")
         # Setup some helpful parameters
         skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
@@ -1241,6 +1243,106 @@ class IFUReduce(MultiSlitReduce, Reduce):
         numbins = int(np.max(self.slits.get_slitlengths(initial=True, median=True)))
         spatbins = np.linspace(0.0, 1.0, numbins + 1)
         spat_slit = 0.5 * (spatbins[1:] + spatbins[:-1])
+
+        # For KCWI there are two groups of slits to fit together
+        spat_fitfunc = lambda par, ydata, slit, model: (par[0] + par[1]*slit + par[2]*slit**2)*(1 + (par[3] + slit*par[4]) * model) - ydata
+        fitall=False
+        for gr in range(2):
+            msgs.info("Deriving spatial correction for slits in group {0:d}/2".format(gr + 1))
+            # Setup fit for this group
+            xfit, yfit, slitfit, modfit = np.array([]), np.array([]), np.array([]), np.array([])
+            spls = [None for all in self.slits.spat_id]
+            for sl in range(gr, self.slits.spat_id.size, 2):
+                # Get the initial slit locations
+                spatid = self.slits.spat_id[sl]
+                onslit_b_init = (slitid_img_init == spatid)
+
+                # Synthesize ximg, and edgmask from slit boundaries.
+                spatcoord, edgmask = pixels.ximg_and_edgemask(self.slits_left[:, sl], self.slits_right[:, sl],
+                                                              onslit_b_init, trim_edg=trim_edg)
+                # Ignore skymask
+                coord_msk = onslit_b_init & gpm
+                hist, _ = np.histogram(spatcoord[coord_msk], bins=spatbins, weights=rawimg[coord_msk])
+                cntr, _ = np.histogram(spatcoord[coord_msk], bins=spatbins)
+                hist_slit_all = hist / (cntr + (cntr == 0))
+
+                # Repeat with skymask
+                coord_msk = onslit_b_init & gpm & skymask_now
+                hist, _ = np.histogram(spatcoord[coord_msk], bins=spatbins, weights=rawimg[coord_msk])
+                cntr, _ = np.histogram(spatcoord[coord_msk], bins=spatbins)
+                hist_slit = hist / (cntr + (cntr == 0))
+
+                # Prepare for fit - take the non-zero elements and trim slit edges
+                ww = (hist_slit[hist_trim:-hist_trim] != 0)
+                xfit = np.append(xfit, spat_slit[hist_trim:-hist_trim][ww])
[…]
