### 2020-07-17 66564e4
det=2 fix
