+++ pypeit/calibrations.py
@@ -450,7 +450,7 @@ class Calibrations(object):
 
         # Check if a bias frame exists, and if a BPM should be generated
         msbias = None
-        if self.par['bpm_usebias'] and self._cached('bias', self.master_key_dict['bias']):
+        if self.par['bpm_usebias']:
             msbias = self.msbias
         # Build it
         self.msbpm = self.spectrograph.bpm(sci_image_file, self.det, msbias=msbias)
