### 2020-07-18 a1ad7b9
Fix typos in HolyGrail
