### 2020-07-18 b92ac10
Change `is` to `==` in string comparisons
