+++ pypeit/spectrographs/p200_dbsp.py
@@ -1,0 +1,435 @@
+""" Module for P200/DBSP specific codes
+"""
+import numpy as np
+
+from astropy.io import fits
+from astropy.coordinates import Angle
+from astropy.time import Time
+from pkg_resources import resource_filename
+
+from pypeit import msgs
+from pypeit import telescopes
+from pypeit.core import framematch
+from pypeit.par import pypeitpar
+from pypeit.spectrographs import spectrograph
+from pypeit.core import parse
+from pypeit.images import detector_container
+
+from pypeit import debugger
+
+
+class P200DBSPSpectrograph(spectrograph.Spectrograph):
+    """
+    Child to handle P200/DBSP specific code
+    """
+    ndet = 1
+
+    def __init__(self):
+        # Get it started
+        super(P200DBSPSpectrograph, self).__init__()
+        self.spectrograph = 'p200_dbsp'
+        self.telescope = telescopes.P200TelescopePar()
+
+    def configuration_keys(self):
+        """
+        Return the metadata keys that defines a unique instrument
+        configuration.
+
+        This list is used by :class:`pypeit.metadata.PypeItMetaData` to
+        identify the unique configurations among the list of frames read
+        for a given reduction.
+
+        Returns:
+            list: List of keywords of data pulled from meta
+        """
+        return ['dispname', 'binning', 'dispangle', 'dichroic']
+
+    def init_meta(self):
+        """
+        Generate the meta data dict
+        Note that the children can add to this
+
+        Returns:
+            self.meta: dict (generated in place)
+
+        """
+        meta = {}
+        # Required (core)
+        meta['ra'] = dict(ext=0, card='RA')
+        meta['dec'] = dict(ext=0, card='DEC')
+        meta['target'] = dict(ext=0, card='OBJECT')
+
+        meta['dispname'] = dict(ext=0, card='GRATING')
+        meta['decker'] = dict(ext=0, card='APERTURE')
+        meta['binning'] = dict(card=None, compound=True)
+
+        meta['mjd'] = dict(card=None, compound=True)
+        meta['exptime'] = dict(ext=0, card='EXPTIME')
+        meta['airmass'] = dict(ext=0, card='AIRMASS')
+
+        # Extras for config and frametyping
+        meta['dichroic'] = dict(ext=0, card='DICHROIC')
+        meta['dispangle'] = dict(card=None, rtol=1e-2, compound=True)
+        meta['slitwid'] = dict(ext=0, card='APERTURE')
+        meta['idname'] = dict(ext=0, card='IMGTYPE')
+        # Lamps
+        meta['lampstat01'] = dict(ext=0, card='LAMPS')
+
+        # Ingest
+        self.meta = meta
+
+    def compound_meta(self, headarr, meta_key):
+        if meta_key == 'mjd':
+            return Time(headarr[0]['UTSHUT']).mjd
+        elif meta_key == 'dispangle':
+            try:
+                return Angle(headarr[0]['ANGLE'], ).deg
+            except Exception as e:
+                print(headarr[0]['ANGLE'])
+                raise e
+        else:
+            return None
+            #msgs.error("Not ready for this compound meta")
+
+    def pypeit_file_keys(self):
+        pypeit_keys = super(P200DBSPSpectrograph, self).pypeit_file_keys()
+        pypeit_keys += ['slitwid']
+        return pypeit_keys
+
+
+class P200DBSPBlueSpectrograph(P200DBSPSpectrograph):
+    """
+    Child to handle P200/DBSP blue specific code
+    """
+    def __init__(self):
+        # Get it started
+        super(P200DBSPBlueSpectrograph, self).__init__()
+        self.spectrograph = 'p200_dbsp_blue'
+        self.camera = 'DBSPb'
+    
+    def compound_meta(self, headarr, meta_key):
[…]
