+++ pypeit/par/pypeitpar.py
@@ -2040,42 +2040,46 @@ class WavelengthSolutionPar(ParSet):
         defaults['method'] = 'holy-grail'
         options['method'] = WavelengthSolutionPar.valid_methods()
         dtypes['method'] = str
-        descr['method'] = 'Method to use to fit the individual arc lines. Most of these methods are now deprecated ' \
-                          'as they fail most of the time without significant parameter tweaking. ' \
-                          '\'holy-grail\' attempts to get a first guess at line IDs by looking for patterns in the ' \
-                          'line locations. It is fully automated and works really well excpet for when it does not' \
-                          '\'reidentify\' is now the preferred method, however it requires that an archive of ' \
-                          'wavelength solution has been constructed for your instrument/grating combination'' \
-                          ''Options are: {0}'.format(', '.join(options['method']))
-#        descr['method'] = 'Method to use to fit the individual arc lines.  ' \
-#                          '\'fit\' is likely more accurate, but \'simple\' uses a polynomial ' \
-#                          'fit (to the log of a gaussian) and is fast and reliable.  ' \
-#                          '\'arclines\' uses the arclines python package.' \
-#                          'Options are: {0}'.format(', '.join(options['method']))
+        descr['method'] = 'Method to use to fit the individual arc lines.  Note that most of ' \
+                          'the available methods should not be used; they are unstable and ' \
+                          'require significant parameter tweaking to succeed.  You should use' \
+                          'either \'holy-grail\' or \'reidentify\': \'holy-grail\' attempts to ' \
+                          'get a first guess at line IDs by looking for patterns in the line ' \
+                          'locations.  It is fully automated.  When it works, it works well; ' \
+                          'however, it can fail catastrophically.  Instead, \'reidentify\' is ' \
+                          'the preferred method.  It requires an archived wavelength solution ' \
+                          'for your specific instrument/grating combination as a reference.  ' \
+                          'This is used to anchor the wavelength solution for the data being ' \
+                          'reduced.  All options are: {0}'.format(', '.join(options['method']))
 
         # Echelle wavelength calibration stuff
         # TODO: Is this needed? I.e., where do we need this parameter
         # when we don't have access to spectrograph.pypeline?
         defaults['echelle'] = False
         dtypes['echelle'] = bool
-        descr['echelle'] = 'Is this an echelle spectrograph? If yes an additional 2-d fit wavelength fit will be performed as a function ' \
-                           'of spectral pixel and order number to improve the wavelength solution'
+        descr['echelle'] = 'Is this an echelle spectrograph? If yes an additional 2-d fit ' \
+                           'wavelength fit will be performed as a function of spectral pixel ' \
+                           'and order number to improve the wavelength solution'
 
         defaults['ech_nspec_coeff'] = 4
         dtypes['ech_nspec_coeff'] = int
-        descr['ech_nspec_coeff'] = 'For echelle spectrographs, order of the final 2d fit to the spectral dimension. ' \
-                                   'You should choose this to be the n_final of the fits to the individual orders.'
+        descr['ech_nspec_coeff'] = 'For echelle spectrographs, this is the order of the final ' \
+                                   '2d fit to the spectral dimension.  You should choose this ' \
+                                   'to be the n_final of the fits to the individual orders.'
 
         defaults['ech_norder_coeff'] = 4
         dtypes['ech_norder_coeff'] = int
-        descr['ech_norder_coeff'] = 'For echelle spectrographs, order of the final 2d fit to the order dimension.'
+        descr['ech_norder_coeff'] = 'For echelle spectrographs, this is the order of the final ' \
[…]
