+++ pypeit/reduce.py
@@ -1227,8 +1227,8 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.sciImg.ivar = utils.inverse(varImg)
         return
 
-#    def illum_profile_spatial_fitall(self, skymask=None, trim_edg=(0,0)):
-    def illum_profile_spatial(self, skymask=None, trim_edg=(0, 0)):
+    def illum_profile_spatial_fitall(self, skymask=None, trim_edg=(0,0)):
+#    def illum_profile_spatial(self, skymask=None, trim_edg=(0, 0)):
 
         msgs.info("Performing spatial sensitivity correction")
         # Setup some helpful parameters
@@ -1327,22 +1327,28 @@ class IFUReduce(MultiSlitReduce, Reduce):
         # Apply the relative scale correction
         self.apply_relative_scale(spatScaleImg)
 
-    def illum_profile_spatial_separateSlits(self, skymask=None, trim_edg=(0,0)):
+#def spat_func(par, ydata, coords, spatbins, ww, spl, hist_trim, slitlength):
+ #   return par[0]*(1 + par[1] * model) - ydata
+
+#    def illum_profile_spatial_separateSlits(self, skymask=None, trim_edg=(0,0)):
+    def illum_profile_spatial(self, skymask=None, trim_edg=(0, 0)):
+
         msgs.info("Performing spatial sensitivity correction")
         # Setup some helpful parameters
         skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
-        hist_trim = 3  # Trim the edges of the histogram to take into account edge effects
+        hist_trim = 0  # Trim the edges of the histogram to take into account edge effects
         gpm = (self.sciImg.fullmask == 0)
         slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
         spatScaleImg = np.ones_like(self.sciImg.image)
         # For each slit, grab the spatial coordinates and a spline
-        # representation of te spatial profile from the illumflat
+        # representation of the spatial profile from the illumflat
         rawimg = self.sciImg.image.copy()
         numbins = int(np.max(self.slits.get_slitlengths(initial=True, median=True)))
         spatbins = np.linspace(0.0, 1.0, numbins + 1)
         spat_slit = 0.5 * (spatbins[1:] + spatbins[:-1])
-        embed()
+        slitlength = np.median(self.slits.get_slitlengths(median=True))
         coeff_fit = np.zeros((self.slits.nslits, 2))
+        #embed()
         for sl, slitnum in enumerate(self.slits.spat_id):
             msgs.info("Deriving spatial correction for slit {0:d}/{1:d}".format(sl + 1, self.slits.spat_id.size))
             # Get the initial slit locations
@@ -1352,11 +1358,20 @@ class IFUReduce(MultiSlitReduce, Reduce):
             # routine would save time. But this is pretty fast, so we just do it here to make the interface simpler.
             spatcoord, edgmask = pixels.ximg_and_edgemask(self.slits_left[:, sl], self.slits_right[:, sl],
                                                           onslit_b_init, trim_edg=trim_edg)
+
+            # Make the model histogram
+            xspl = np.linspace(0.0, 1.0, 10 * int(slitlength))  # Sub sample each pixel with 10 subpixels
+            modspl = self.caliBrate.flatimages.illumflat_spat_bsplines[sl].value(xspl)[0]
+            gradspl = interpolate.interp1d(xspl, np.gradient(modspl) / modspl, kind='linear', bounds_error=False,
[…]

[…]
