+++ pypeit/edgetrace.py
@@ -118,14 +118,14 @@ from astropy.stats import sigma_clipped_stats
 from astropy import table
 
 from pypeit import msgs
+from pypeit import io
 from pypeit import utils
 from pypeit import sampling
 from pypeit import ginga
 from pypeit import masterframe
-from pypeit import io
 from pypeit import slittrace
 # TODO: Commented until EdgeTraceSet becomes a DataContainer
-#from pypeit.datamodel import DataContainer
+from pypeit.datamodel import DataContainer
 from pypeit.bitmask import BitMask
 from pypeit.par.pypeitpar import EdgeTracePar
 from pypeit.core import parse, pydl, procimg, pca, trace
@@ -160,8 +160,6 @@ class EdgeTraceBitMask(BitMask):
                 ('DISCONTINUOUS', 'Pixel included in a trace but part of a discontinuous segment'),
                     ('DUPLICATE', 'Trace is a duplicate based on trace matching tolerance'),
                    ('SHORTRANGE', 'Trace does not meet the minimum spectral range criterion'),
-#                ('SHORTDETRANGE', 'Trace length does not meet trace detection threshold.'),
-#                ('SHORTFITRANGE', 'Trace length does not meet fitting/PCA threshold.'),
                        ('HITMIN', 'Trace crosses the minimum allowed spatial column'),
                        ('HITMAX', 'Trace crosses the maximum allowed spatial column'),
                   ('OFFDETECTOR', 'Trace lands off, or within `det_buffer` of, the detector edge'),
@@ -205,14 +203,13 @@ class EdgeTraceBitMask(BitMask):
 
     @property
     def exclude_flags(self):
-        # We will not exclude SYNCINSERT slits edges from the masking
-        #   These can easily crop up for star boxes and at the edge of the detector
-#        return ['USERINSERT', 'MASKINSERT', 'ORPHANINSERT', 'BOXSLIT']
+        """
+        List of flags to exclude when finding bad trace data.
+        """
         return self.insert_flags + self.order_flags + ['BOXSLIT']
 
 
-# TODO -- Make a DataContainer
-class EdgeTraceSet(object):
+class EdgeTraceSet(DataContainer):
     r"""
     Core class that identifies, traces, and pairs edges in an image
     to define the slit apertures.
@@ -226,25 +223,10 @@ class EdgeTraceSet(object):
     value of `auto`. To automatically have the instantiation save the
     results, set `save=True` on instantiation.
 
-    To load an existing master file with the result of a trace, you
-    can either use the :attr:`from_file` method::
+    To load an existing master file with the result of a trace, 
+    use the :attr:`from_file` method::
 
         edges = EdgeTraceSet.from_file(file)
-
-    or instantiate the object as would have been done initially and
-    then check if the trace exists and load it::
-    
-        edges = EdgeTraceSet(spec, par)
-        if edges.exists:
-            edges.load()
-
-    or you can attempt to load directly on instantiation::
-
-        edges = EdgeTraceSet(spec, par, load=True)
-
-    In the latter case, note that the `load` argument takes
-    precedence and an exception is raised if one provides both `img`
-    and sets `load=True`.
 
     Most commonly, one will use the automatic tracing routine to
     trace the slit edges; see the description of the steps used
@@ -304,9 +286,6 @@ class EdgeTraceSet(object):
             :func:`show`.
         save (:obj:`bool`, optional):
             Save the result to the master frame.
-        load (:obj:`bool`, optional):
-            Attempt to load existing output. If True and the file
-            does not exist, an error is raised.
 
     Attributes:
         trace_img
@@ -394,13 +373,40 @@ class EdgeTraceSet(object):
             A list of strings indicating the main methods applied
[…]
