### 2020-07-22 d1fffd1
Bump version: 1.0.6dev → 1.0.6
