+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.0.6dev
+current_version = 1.0.6
 commit = False
 tag = False
 tag_name = {new_version}
