+++ CHANGES.rst
@@ -1,4 +1,7 @@
 
+1.0.7dev
+--------
+
 1.0.6 (22 Jul 2020)
 -------------------
 
