### 2020-07-22 81f86e9
Bump version: 1.0.6 → 1.0.7dev
