+++ doc/conf.py
@@ -90,9 +90,9 @@ author = u'Prochaska, Hennawi, Cooke, and Associates'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0.6'
+version = '1.0.7dev'
 # The full version, including alpha/beta/rc tags.
-release = '1.0.6'
+release = '1.0.7dev'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
