+++ pypeit/core/wavecal/autoid.py
@@ -1229,7 +1229,7 @@ class ArchiveReid:
                 msgs.warn(badmsg)
                 continue
             st = str(slit)
-            if self.wv_calib[st] is None or len(self.wv_calib[st]) == 0:
+            if self.wv_calib[st] is None: # or len(self.wv_calib[st]) == 0:
                 msgs.warn('Bad solution for slit: {}'.format(st))
                 continue
             if self.all_patt_dict[st]['sign'] == +1:
