+++ pypeit/core/wavecal/wv_fitting.py
@@ -277,8 +277,11 @@ def iterative_fitting(spec, tcent, ifit, IDs, llist, disp,
         # Fit with rejection
         xfit, yfit, wfit = tcent[ifit], all_ids[ifit], weights[ifit]
         maxiter = xfit.size - n_order - 2
-        #mask, fit = utils.robust_polyfit(xfit/xnspecmin1, yfit, n_order, function=func, sigma=sigrej_first,
-        #                                 minx=fmin, maxx=fmax, verbose=verbose, weights=wfit)
+        #
+        if xfit.size == 0:
+            msgs.warn("All points rejected !!")
+            return None
+        # Fit
         pypeitFit = fitting.robust_fit(xfit/xnspecmin1, yfit, n_order, function=func, maxone=True, maxiter=maxiter,
                                        lower=sigrej_first, upper=sigrej_first, maxrej=1, sticky=False,
                                        minx=fmin, maxx=fmax, weights=wfit)
