### 2020-07-22 84512ea
corrected decker card
