+++ pypeit/spectrographs/gemini_gnirs.py
@@ -221,7 +221,7 @@ class GeminiGNIRSSpectrograph(spectrograph.Spectrograph):
         meta['ra'] = dict(ext=0, card='RA')
         meta['dec'] = dict(ext=0, card='DEC')
         meta['target'] = dict(ext=0, card='OBJECT')
-        meta['decker'] = dict(ext=0, card='SLIT')
+        meta['decker'] = dict(ext=0, card='decker')
 
         meta['binning'] = dict(ext=0, card=None, default='1,1')
         meta['mjd'] = dict(ext=0, card='MJD_OBS')
