+++ pypeit/tracepca.py
@@ -14,7 +14,7 @@ from astropy.io import fits
 
 from pypeit import msgs
 from pypeit import utils
-from pypeit.io import init_record_array, rec_to_bintable
+from pypeit.io import hdu_iter_by_ext
 from pypeit.core import trace
 from pypeit.core import pca
 from pypeit.datamodel import DataContainer
@@ -65,7 +65,7 @@ class TracePCA(DataContainer):
     datamodel = {'reference_row': dict(otype=int, 
                                        descr='The row (spectral position) used as the reference ' \
                                              'coordinate system for the PCA.'),
-                 'trace_coo': dict(otype=np.ndarray, atype=np.floating,
+                 'trace_coo': dict(otype=np.ndarray, atype=(float,np.floating),
                                    descr='Trace coordinates.  Shape must be '
                                          r':math:`(N_{\rm spec},N_{\rm trace}).'),
                  'nspec': dict(otype=int,
@@ -74,22 +74,22 @@ class TracePCA(DataContainer):
                  'input_npca': dict(otype=int,
                                     descr='Requested number of PCA components if provided.'),
                  'npca': dict(otype=int, descr='Number of PCA components used.'),
-                 'input_pcav': dict(otype=float,
+                 'input_pcav': dict(otype=(float,np.floating),
                                     descr='Requested variance accounted for by PCA decomposition.'),
-                 'pca_coeffs': dict(otype=np.ndarray, atype=np.floating,
+                 'pca_coeffs': dict(otype=np.ndarray, atype=(float,np.floating),
                                     descr=r'PCA component coefficients. If the PCA decomposition '
                                           r'used :math:`N_{\rm comp}` components for '
                                           r':math:`N_{\rm vec}` vectors, the shape of this array '
                                           r'must be :math:`(N_{\rm vec}, N_{\rm comp})`. The '
                                           r'array can be 1D with shape :math:`(N_{\rm vec},)` if '
                                           r'there was only one PCA component.'),
-                 'pca_components': dict(otype=np.ndarray, atype=np.floating,
+                 'pca_components': dict(otype=np.ndarray, atype=(float,np.floating),
                                         descr=r'Vectors with the PCA components.  Shape must be '
                                               r':math:`(N_{\rm comp}, N_{\rm spec})`.'),
-                 'pca_mean': dict(otype=np.ndarray, atype=np.floating,
+                 'pca_mean': dict(otype=np.ndarray, atype=(float,np.floating),
                                   descr=r'The mean offset of the PCA decomposotion for each '
                                         r' spectral pixel. Shape is :math:`(N_{\rm spec},)`.'),
-                 'pca_bpm': dict(otype=np.ndarray, atype=np.bool,
+                 'pca_bpm': dict(otype=np.ndarray, atype=(bool, np.bool_),
                                  descr='A boolean numpy.ndarray masking data PCA component '
                                        'coefficients that were rejected during the polynomial '
                                        'fitting. Shape is the same as ``pca_coeff``.'),
@@ -103,17 +103,17 @@ class TracePCA(DataContainer):
                                          r'the PCA-component coefficients. The number of function '
                                          r'coefficients is typically :math:`N_{\rm coeff} = o+1`, '
                                          r'where :math:`o` is the order of the fit.'),
-                 'minx': dict(otype=float,
+                 'minx': dict(otype=(float,np.floating),
[…]

[…]
