+++ pypeit/scripts/coadd_datacube.py
@@ -17,12 +17,15 @@ from astropy.io import fits
 from astropy.coordinates import AltAz, SkyCoord
 import scipy.optimize as opt
 from scipy.interpolate import interp1d
+import scipy.signal
 import numpy as np
+import copy
 
 from pypeit import msgs
 from pypeit.spectrographs.util import load_spectrograph
 from pypeit.core.parse import get_dnum
 from pypeit.core.flux_calib import load_extinction_data, extinction_correction
+from pypeit.core.procimg import grow_masked
 from pypeit import alignframe
 from pypeit import spec2dobj
 
@@ -53,7 +56,8 @@ def dar_fitfunc(radec, coord_ra, coord_dec, datfit, wave, obstime, location, pre
     return np.sum((np.array([coord_altaz.alt.value, coord_altaz.az.value])-datfit)**2)
 
 
-def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, rel_humidity):
+def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, rel_humidity,
+                   wave_ref=None, numgrid = 100):
     """Apply a differental atmospheric refraction correction to the input ra/dec.
     This implementation is based on ERFA, which is called through astropy
 
@@ -71,6 +75,10 @@ def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, re
         Outside ambient air temperature at `location`
     rel_humidity (float):
         Outside relative humidity at `location`. This should be between 0 to 1.
+    wave_ref (float):
+        Reference wavelength (The DAR correction will be performed relative to this wavelength)
+    numgrid (int):
+        Number of grid points to evaluate the DAR correction.
 
     Returns
     -------
@@ -83,7 +91,10 @@ def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, re
     TODO :: Move this routine to the main PypeIt code
     """
     msgs.info("Performing differential atmospheric refraction correction")
-    numgrid = 100
+
+    if wave_ref is None:
+        wave_ref = 0.5*(wave_arr.min() + wave_arr.max())
+
     # First create the reference frame and wavelength grid
     coord_altaz = coord.transform_to(AltAz(obstime=obstime, location=location))
     wave_grid = np.linspace(wave_arr.min(), wave_arr.max(), numgrid) * units.AA
@@ -105,8 +116,8 @@ def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, re
     spl_dec = interp1d(wave_grid, dec_grid, kind='cubic')
 
     # Evaluate the differentials at the input wave_arr
-    ra_diff = spl_ra(wave_arr)
-    dec_diff = spl_dec(wave_arr)
+    ra_diff = spl_ra(wave_arr) - spl_ra(wave_ref)
+    dec_diff = spl_dec(wave_arr) - spl_dec(wave_ref)
 
     return ra_diff, dec_diff
 
@@ -131,6 +142,41 @@ def generate_masterWCS(crval, cdelt, equinox=2000.0):
     return w
 
 
+def twoD_Gaussian(tup, amplitude, xo, yo, sigma_x, sigma_y, theta, offset):
+    (x, y) = tup
+    xo = float(xo)
+    yo = float(yo)
+    a = (np.cos(theta)**2)/(2*sigma_x**2) + (np.sin(theta)**2)/(2*sigma_y**2)
+    b = -(np.sin(2*theta))/(4*sigma_x**2) + (np.sin(2*theta))/(4*sigma_y**2)
+    c = (np.sin(theta)**2)/(2*sigma_x**2) + (np.cos(theta)**2)/(2*sigma_y**2)
[…]
