### 2020-07-23 ad2968c
extcorr + varcube
