### 2020-07-24 1883cc3
fix skysub script
