+++ pypeit/reduce.py
@@ -551,23 +551,6 @@ class Reduce(object):
             if np.sum(self.global_sky[thismask]) == 0.:
                 self.reduce_bpm[slit_idx] = True
 
-        # wavefull=np.linspace(3950,4450,10000)
-        # import matplotlib.pylab as pl
-        # from matplotlib import pyplot as plt
-        # colors = pl.cm.jet(np.linspace(0, 1, gdslits.size))
-        # for sl, slit_idx in enumerate(gdslits):
-        #     slit_spat = self.slits.spat_id[slit_idx]
-        #     thismask = self.slitmask == slit_spat
-        #     wav = self.waveimg[thismask]
-        #     flx = self.global_sky[thismask]
-        #     argsrt = np.argsort(wav)
-        #     spl = interpolate.interp1d(wav[argsrt], flx[argsrt], bounds_error=False)
-        #     if sl == 0:
-        #         ref = spl(wavefull)
-        #         plt.plot(wavefull, ref/np.nanmedian(ref), color=colors[sl], linestyle=':')
-        #     plt.plot(wavefull, spl(wavefull)/ref, color=colors[sl])
-        # plt.show()
-
         if update_crmask:
             # Find CRs with sky subtraction
             self.sciImg.build_crmask(self.par['scienceframe']['process'],
@@ -1235,110 +1218,6 @@ class IFUReduce(MultiSlitReduce, Reduce):
         self.sciImg.ivar = utils.inverse(varImg)
         return
 
-    def illum_profile_spatial_fitall(self, skymask=None, trim_edg=(0,0)):
-#    def illum_profile_spatial(self, skymask=None, trim_edg=(0, 0)):
-
-        msgs.info("Performing spatial sensitivity correction")
-        # Setup some helpful parameters
-        skymask_now = skymask if (skymask is not None) else np.ones_like(self.sciImg.image, dtype=bool)
-        hist_trim = 3  # Trim the edges of the histogram to take into account edge effects
-        gpm = (self.sciImg.fullmask == 0)
-        slitid_img_init = self.slits.slit_img(pad=0, initial=True, flexure=self.spat_flexure_shift)
-        spatScaleImg = np.ones_like(self.sciImg.image)
-        # For each slit, grab the spatial coordinates and a spline
-        # representation of te spatial profile from the illumflat
-        rawimg = self.sciImg.image.copy()
-        numbins = int(np.max(self.slits.get_slitlengths(initial=True, median=True)))
-        spatbins = np.linspace(0.0, 1.0, numbins + 1)
-        spat_slit = 0.5 * (spatbins[1:] + spatbins[:-1])
-
-        # For KCWI there are two groups of slits to fit together
-        spat_fitfunc = lambda par, ydata, slit, model: (par[0] + par[1]*slit + par[2]*slit**2)*(1 + (par[3] + slit*par[4]) * model) - ydata
-        fitall=False
-        for gr in range(2):
-            msgs.info("Deriving spatial correction for slits in group {0:d}/2".format(gr + 1))
-            # Setup fit for this group
-            xfit, yfit, slitfit, modfit = np.array([]), np.array([]), np.array([]), np.array([])
-            spls = [None for all in self.slits.spat_id]
-            for sl in range(gr, self.slits.spat_id.size, 2):
-                # Get the initial slit locations
-                spatid = self.slits.spat_id[sl]
-                onslit_b_init = (slitid_img_init == spatid)
-
-                # Synthesize ximg, and edgmask from slit boundaries.
[…]
