+++ pypeit/pypeit.py
@@ -376,7 +376,7 @@ class PypeIt(object):
 
         # Check if this is an IFU reduction. If so, make a datacube
         if self.spectrograph.pypeline == "IFU" and self.par['reduce']['cube']['make_cube']:
-            msgs.info("Generating datacube")
+            msgs.work("Generate datacube")
 
         # Finish
         self.print_end_time()
