+++ pypeit/scripts/coadd_datacube.py
@@ -16,15 +16,14 @@ from astropy.io import fits
 from astropy.coordinates import AltAz, SkyCoord
 import scipy.optimize as opt
 from scipy.interpolate import interp1d
-import scipy.signal
 import numpy as np
 import copy
 
 from pypeit import msgs
 from pypeit.spectrographs.util import load_spectrograph
-from pypeit.core.parse import get_dnum
 from pypeit.core.flux_calib import load_extinction_data, extinction_correction
 from pypeit.core.procimg import grow_masked
+from pypeit.core.flexure import calculate_image_offset
 from pypeit import alignframe
 from pypeit import spec2dobj
 
@@ -45,6 +44,32 @@ def parser(options=None):
 
 
 def dar_fitfunc(radec, coord_ra, coord_dec, datfit, wave, obstime, location, pressure, temperature, rel_humidity):
+    """ Generates a fitting function to calculate the offset due to differential atmospheric refraction
+
+    Args:
+        radec (tuple):
+            A tuple containing two floats representing the shift in ra and dec due to DAR.
+        coord_ra (float):
+            RA in degrees
+        coord_dec (float):
+            Dec in degrees
+        datfit (`numpy.ndarray`_):
+            The RA and DEC that the model needs to match
+        wave (float):
+            Wavelength to calculate the DAR
+        location (astropy EarthLocation):
+            observatory location
+        pressure (float):
+            Outside pressure at `location`
+        temperature (float):
+            Outside ambient air temperature at `location`
+        rel_humidity (float):
+            Outside relative humidity at `location`. This should be between 0 to 1.
+
+    Returns:
+        chisq (float):
+            chi-squared difference between datfit and model
+    """
     (diff_ra, diff_dec) = radec
     # Generate the coordinate with atmopheric conditions
     coord_atmo = SkyCoord(coord_ra + diff_ra, coord_dec + diff_dec, unit=(units.deg, units.deg))
@@ -60,33 +85,31 @@ def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, re
     """Apply a differental atmospheric refraction correction to the input ra/dec.
     This implementation is based on ERFA, which is called through astropy
 
-    Parameters
-    ----------
-    wave_arr (ndarray):
-        wavelengths to obtain ra and dec offsets
-    coord (astropy SkyCoord):
-        ra, dec positions at the centre of the field
-    obstime (astropy Time):
-        time at the midpoint of observation
-    location (astropy EarthLocation):
-        observatory location
-    pressure (float):
-        Outside pressure at `location`
-    temperature (float):
-        Outside ambient air temperature at `location`
-    rel_humidity (float):
-        Outside relative humidity at `location`. This should be between 0 to 1.
-    wave_ref (float):
-        Reference wavelength (The DAR correction will be performed relative to this wavelength)
-    numgrid (int):
-        Number of grid points to evaluate the DAR correction.
+    Args:
+        wave_arr (`numpy.ndarray`_):
+            wavelengths to obtain ra and dec offsets
+        coord (astropy SkyCoord):
+            ra, dec positions at the centre of the field
+        obstime (astropy Time):
+            time at the midpoint of observation
+        location (astropy EarthLocation):
+            observatory location
+        pressure (float):
+            Outside pressure at `location`
[…]
