### 2020-07-27 9f4ec7a
fixed docstrings
