+++ pypeit/spectrographs/keck_kcwi.py
@@ -794,51 +794,3 @@ class KeckKCWISpectrograph(spectrograph.Spectrograph):
         w.wcs.latpole = 0.0  # Native latitude of the Celestial pole
 
         return w
-
-    def get_radec_image(self, slits, wcs, flexure=None, trace_cen=None):
-        """Generate an RA and DEC image for every pixel in the frame
-
-        Parameters
-        ----------
-        slits : :class:`pypeit.slittrace.SlitTraceSet`
-            Master slit edges
-        wcs : astropy.wcs
-            The World Coordinate system of a science frame
-        maxslitlen : int
-            This is the slit length in pixels, and it should be the same
-            value that was passed to get_wcs() to generate the WCS that
-            is passed into this function as an argument.
-        flexure : float, optional
-            If provided, offset each slit by this amount.
-        trace_cen : `numpy.ndarray`_, optional
-            Central traces of each slit. Shape should be (slits.nspec, slits.nslits).
-            If None, the average of the left and right slit edges will be used
-
-        Returns
-        -------
-        ndimage, ndimage : Two 2D numpy array of shape (nspec, nspat), where the
-                           first ndarray is the RA image, and the second ndarray
-                           is the DEC image. RA and DEC are in units degrees.
-        """
-        # Grab the alignments
-        if trace_cen is None:
-            left, right, _ = slits.select_edges(initial=True, flexure=flexure)
-            trace_cen = 0.5*(left+right)
-        # Initialise the output
-        raimg = np.zeros((slits.nspec, slits.nspat))
-        decimg = np.zeros((slits.nspec, slits.nspat))
-        minmax = np.zeros((slits.nslits, 2))
-        # Get the slit information
-        slitid_img_init = slits.slit_img(pad=0, initial=True, flexure=flexure)
-        for slit_idx, spatid in enumerate(slits.spat_id):
-            onslit = (slitid_img_init == spatid)
-            onslit_init = np.where(onslit)
-            evalpos = onslit_init[1]-trace_cen[onslit_init[0], slit_idx]
-            minmax[:, 0] = np.min(evalpos)
-            minmax[:, 1] = np.max(evalpos)
-            slitID = np.ones(evalpos.size) * slit_idx - wcs.wcs.crpix[0]
-            world_ra, world_dec, _ = wcs.wcs_pix2world(slitID, evalpos, onslit_init[0], 0)
-            # Set the RA first and DEC next
-            raimg[onslit] = world_ra.copy()
-            decimg[onslit] = world_dec.copy()
-        return raimg, decimg, minmax
