### 2020-07-28 #2308 @kthyng
Hi @Gabriel-p, @astrom-tom! This is just a friendly reminder about this JOSS review. Will you be able to work on this in the next week or so?
