### 2020-07-30 13bec03
Fixing bugs for GNIRS_10L when orders shift.
