+++ pypeit/core/fitting.py
@@ -27,23 +27,25 @@ class PypeItFit(DataContainer):
     datamodel = {
         'xval': dict(otype=np.ndarray, atype=np.floating, desc='x inputs'),
         'yval': dict(otype=np.ndarray, atype=np.floating, desc='y inputs'),
+        'order': dict(otype=np.ndarray, atype=np.integer, desc='The order of the polynomial to be used in the fitting. This is a 2d array for 2d fits'),
+        'x2': dict(otype=np.ndarray, atype=np.floating, desc='x2 inputs, second independent variable'),
         'weights': dict(otype=np.ndarray, atype=np.floating, desc='Weights'),
         'fitc': dict(otype=np.ndarray, atype=np.floating, desc='Fit coefficients'),
         'fitcov': dict(otype=np.ndarray, atype=np.floating, desc='Covariance of the coefficients'),
         'gpm': dict(otype=np.ndarray, atype=np.integer, desc='Mask (1=good)'),
-        'func': dict(otype=str, desc='Fit function (polynomial, legendre, chebyshev, bspline, gauss)'),
+        'func': dict(otype=str, desc='Fit function (polynomial, legendre, chebyshev, polynomial2d, legendre2d, gauss)'),
         'minx': dict(otype=float,
                      desc='minimum value in the array (or the left limit for a legendre / chebyshev polynomial)'),
         'maxx': dict(otype=float,
                      desc='maximum value in the array (or the right limit for a legendre / chebyshev polynomial)'),
-        'minx2': dict(otype=float, desc='To be written by JFH'),
-        'maxx2': dict(otype=float, desc='To be written by JFH'),
+        'minx2': dict(otype=float, desc='Same as minx for the second independent variable x2'),
+        'maxx2': dict(otype=float, desc='Same as maxx for the second independent variable x2'),
     }
 
     # This needs to contain all datamodel items
-    def __init__(self, xval=None, yval=None, weights=None, fitc=None,
+    def __init__(self, xval=None, yval=None, order=None, x2=None, weights=None, fitc=None,
                  fitcov=None, func=None, minx=None, maxx=None, minx2=None,
-                 maxx2=None, gpm=None):
+                 maxx2=None, gpm=None, guesses=None, return_errors=False):
         # Setup the DataContainer
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         _d = {k: values[k] for k in args[1:]}
@@ -75,7 +77,147 @@ class PypeItFit(DataContainer):
         # Do it
         return super(PypeItFit, self).to_hdu(**_d)
 
-    def val(self, x, x2=None): #, minx=None, maxx=None, minx2=None, maxx2=None):
+    def func_fit(self):
+        """
+
+        Args:
+            x (`numpy.ndarray`_):
+            y (`numpy.ndarray`_):
+            func (:obj:`str`):
+                polynomial, legendre, chebyshev, gaussian
+            deg (:obj:`int`):
+                degree of the fit
+            x2 (`numpy.ndarray`_, optional):
+            minx:
+            maxx:
+            minx2:
+            maxx2:
+            w (`numpy.ndarray`_, optional):
+            inmask (`numpy.ndarray`_, optional):
+            guesses:
+            return_errors:
+
+        Returns:
+            PypeItFit:
+
+        """
+        if self.return_errors:
+            msgs.error("Need to deal with this")
+        # Init
+        self.fitcov = None
+
+        # If the user provided an inmask apply it. The logic below of evaluating the fit only at the non-masked
+        # pixels is preferable to the other approach of simply setting the weights to zero. The reason for that is that
[…]
