### 2020-07-31 36b1988
Fixing bugs for GNIRS_10L when orders shift.
