+++ pypeit/core/fitting.py
@@ -33,6 +33,7 @@ class PypeItFit(DataContainer):
         'fitc': dict(otype=np.ndarray, atype=np.floating, desc='Fit coefficients'),
         'fitcov': dict(otype=np.ndarray, atype=np.floating, desc='Covariance of the coefficients'),
         'gpm': dict(otype=np.ndarray, atype=np.integer, desc='Mask (1=good)'),
+        'success': dict(otype=int, desc='Flag indicating whether fit was successful (success=1) or if it failed (success=0)'),
         'func': dict(otype=str, desc='Fit function (polynomial, legendre, chebyshev, polynomial2d, legendre2d, gauss)'),
         'minx': dict(otype=float,
                      desc='minimum value in the array (or the left limit for a legendre / chebyshev polynomial)'),
@@ -45,7 +46,7 @@ class PypeItFit(DataContainer):
     # This needs to contain all datamodel items
     def __init__(self, xval=None, yval=None, order=None, x2=None, weights=None, fitc=None,
                  fitcov=None, func=None, minx=None, maxx=None, minx2=None,
-                 maxx2=None, gpm=None, guesses=None, return_errors=False):
+                 maxx2=None, gpm=None, success=None):
         # Setup the DataContainer
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         _d = {k: values[k] for k in args[1:]}
@@ -77,49 +78,53 @@ class PypeItFit(DataContainer):
         # Do it
         return super(PypeItFit, self).to_hdu(**_d)
 
-    def func_fit(self):
+    @property
+    def bool_gpm(self):
+        return self.gpm.astype(bool) if self.gpm is not None else None
+
+    def fit(self, guesses=None, return_errors=False):
         """
 
         Args:
-            x (`numpy.ndarray`_):
-            y (`numpy.ndarray`_):
-            func (:obj:`str`):
-                polynomial, legendre, chebyshev, gaussian
-            deg (:obj:`int`):
-                degree of the fit
-            x2 (`numpy.ndarray`_, optional):
-            minx:
-            maxx:
-            minx2:
-            maxx2:
-            w (`numpy.ndarray`_, optional):
-            inmask (`numpy.ndarray`_, optional):
-            guesses:
+            guesses (array):
+                Guesses for Gaussian fits
             return_errors:
 
-        Returns:
-            PypeItFit:
-
         """
-        if self.return_errors:
+
+        # TODO Fix this!
+        if return_errors:
             msgs.error("Need to deal with this")
         # Init
         self.fitcov = None
 
-        # If the user provided an inmask apply it. The logic below of evaluating the fit only at the non-masked
+        # If the user provided an gpm apply it. The logic below of evaluating the fit only at the non-masked
         # pixels is preferable to the other approach of simply setting the weights to zero. The reason for that is that
         # the fits use a least-square optimization approach using matrix algebra, and lots of zero weights are
         # 1) more costly, and 2) will not produce exactly the same result (due to roundoff error) as actually
         # removing the locations you want to mask.
-        if self.inmask is not None:
-            x_out = self.xval[self.gpm]
-            y_out = self.yval[self.gpm]
+
+        # This block ensures sensible zero coefficient outputs are returned if the fits was successful
+        if not np.any(self.bool_gpm):
+            if self.func == "gaussian":
+                self.fitc = np.zeros(self.order[0])
+            elif '2d' in self.func:
+                self.fitc = np.zeros(self.order[0] + 1, self.order[1] + 1)
+            else:
+                self.fitc = np.zeros(self.order[0] + 1)
[…]
