### 2020-07-31 fd3efba
Fixing bugs for GNIRS_10L when orders shift.
