+++ pypeit/core/arc.py
@@ -55,8 +55,8 @@ def fit2darc(all_wv,all_pix,all_orders,nspec, nspec_coeff=4,norder_coeff=4,sigre
 
     Returns
     -------
-    fit_dict : dict
-        Dictionary with the fitting results
+    pypeitFit : pypeit.fitting.PypeItFit
+        Fitting results
 
     """
 
@@ -99,25 +99,25 @@ def fit2darc(all_wv,all_pix,all_orders,nspec, nspec_coeff=4,norder_coeff=4,sigre
     #                                           minx=min_spec, maxx=max_spec, minx2=min_order, maxx2=max_order)
     #resid = (wv_order_mod[fitmask]-all_wv_order[fitmask])
     #fin_rms = np.std(resid)
-    fin_rms = pypeitFit.calc_fit_rms(apply_mask=True)
+    fin_rms = pypeitFit.calc_fit_rms(x2=all_orders, apply_mask=True)
     msgs.info("RMS: {0:.5f} Ang*Order#".format(fin_rms))
 
     orders = np.unique(all_orders)
 
-    fit_dict = dict(coeffs=coeff2, orders=orders,
-                    nspec_coeff=nspec_coeff, norder_coeff=norder_coeff,
-                    min_spec=min_spec, max_spec=max_spec,
-                    min_order=min_order, max_order=max_order,
-                    nspec=nspec, all_pix=all_pix, all_wv=all_wv,
-                    func2d=func2d,xnorm=xnspecmin1,
-                    all_orders=all_orders, all_mask=fitmask)
+    #fit_dict = dict(coeffs=coeff2, orders=orders,
+    #                nspec_coeff=nspec_coeff, norder_coeff=norder_coeff,
+    #                min_spec=min_spec, max_spec=max_spec,
+    #                min_order=min_order, max_order=max_order,
+    #                nspec=nspec, all_pix=all_pix, all_wv=all_wv,
+    #                func2d=func2d,xnorm=xnspecmin1,
+    #                all_orders=all_orders, all_mask=fitmask)
 
 
     if debug:
-        fit2darc_global_qa(fit_dict)
-        fit2darc_orders_qa(fit_dict)
+        fit2darc_global_qa(pypeitFit)
+        fit2darc_orders_qa(pypeitFit)
 
-    return fit_dict
+    return pypeitFit
 
 
 
