+++ pypeit/wavecalib.py
@@ -46,6 +46,7 @@ class WaveCalib(datamodel.DataContainer):
     datamodel = {
         'wv_fits': dict(otype=np.ndarray, atype=wv_fitting.WaveFit,
                               desc='WaveFit to each 1D wavelength solution'),
+        'wv_fit2d': dict(otype=fitting.PypeItFit, desc='2D wavelength solution (echelle)'),
         'nslits': dict(otype=int, desc='Total number of slits.  This can include masked slits'),
         'spat_id': dict(otype=np.ndarray, atype=np.integer, desc='Slit spat_id '),
         'PYP_SPEC': dict(otype=str, desc='PypeIt spectrograph name'),
@@ -53,7 +54,7 @@ class WaveCalib(datamodel.DataContainer):
     }
 
     def __init__(self, wv_fits=None, nslits=None, spat_id=None, PYP_SPEC=None,
-                 strpar=None):
+                 strpar=None, wv_fit2d=None):
         # Parse
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         d = dict([(k,values[k]) for k in args[1:]])
@@ -105,6 +106,8 @@ class WaveCalib(datamodel.DataContainer):
                     kwv_fit.hdu_prefix = 'SPAT_ID-{}_'.format(self.spat_id[ss])
                     # Save
                     _d.append({dkey: kwv_fit})
+            elif key == 'wv_fit2d':
+                _d.append({key: self[key]})
             else: # Add to header of the spat_id image
                 _d[0][key] = self[key]
         # Return
@@ -136,6 +139,8 @@ class WaveCalib(datamodel.DataContainer):
                 i0 = ihdu.name.find('ID-')
                 i1 = ihdu.name.find('_WAV')
                 spat_ids.append(int(ihdu.name[i0+3:i1]))
+            elif ihdu.name == 'PYPEITFIT': # 2D fit
+                _d['wv_fit2d'] = fitting.PypeItFit.from_hdu(ihdu)
         # Check
         if spat_ids != _d['spat_id'].tolist():
             msgs.error("Bad parsing of the MasterFlat")
@@ -192,8 +197,9 @@ class WaveCalib(datamodel.DataContainer):
         # If this is echelle print out a status message and do some error checking
         if self.par['echelle']:
             msgs.info('Evaluating 2-d wavelength solution for echelle....')
-            if len(wv_calib['fit2d']['orders']) != np.sum(ok_slits):
-                msgs.error('wv_calib and ok_slits do not line up. Something is very wrong!')
+            # TODO UPDATE THIS!!
+            #if len(wv_calib['fit2d']['orders']) != np.sum(ok_slits):
+            #    msgs.error('wv_calib and ok_slits do not line up. Something is very wrong!')
 
         # Unpack some 2-d fit parameters if this is echelle
         for islit in np.where(ok_slits)[0]:
@@ -205,7 +211,11 @@ class WaveCalib(datamodel.DataContainer):
                 embed(header='195 of wavecalib')
                 # TODO: Put this in `SlitTraceSet`?
                 order, indx = spectrograph.slit2order(slit_spat_pos[slits.spatid_to_zero(slit_spat)])
-                # evaluate solution
+                # evaluate solution --
+                embed(header='211 of wavecalib')
+                image[thismask] = self.wv_fit2d.eval(tilts[thismask], x2 = np.ones_like(tilts[thismask]) * order)
+
+
                 image[thismask] = utils.func_val(wv_calib['fit2d']['coeffs'],
                                                  tilts[thismask],
                                                  wv_calib['fit2d']['func2d'],
@@ -479,8 +489,7 @@ class BuildWaveCalib(object):
[…]
