### 2020-07-31 915b775
running on GNIRS, but wavelengths may be wrong
