### 2020-08-01 9ad6436
Added PypeIt ginga plugin to show wavelengths.
