### 2020-08-01 2eb981a
Added PypeIt ginga plugin to show wavelengths.
