### 2020-08-01 cc2d682
Added PypeIt ginga plugin to show wavelengths.
