+++ pypeit/tests/test_scripts.py
@@ -17,7 +17,7 @@ from pypeit.scripts import setup, show_1dspec, coadd_1dspec, chk_edges, view_fit
 from pypeit.scripts import trace_edges, run_pypeit, ql_mos, show_2dspec, tellfit, flux_setup
 from pypeit.tests.tstutils import dev_suite_required, cooked_required
 from pypeit import edgetrace
-from pypeit import ginga
+from pypeit.ginga import ginga
 
 
 def data_path(filename):
