### 2020-08-01 b2a4b0e
Added a note on testing with interactive debugging, since I always manage to forget how to do this.
