### 2020-08-01 68e0cdb
Added PypeIt ginga plugin to show wavelengths.
