### 2020-08-01 2e94068
Added PypeIt ginga plugin to show wavelengths.
