### 2020-08-01 8acf2fd
Added PypeIt ginga plugin to show wavelengths.
