### 2020-08-01 2784e4e
Added PypeIt ginga plugin to show wavelengths.
