### 2020-08-01 2658715
Added PypeIt ginga plugin to show wavelengths.
