### 2020-08-02 813a41f
Added PypeIt ginga plugin to show wavelengths.
