### 2020-08-02 75ae03f
Added PypeIt ginga plugin to show wavelengths.
