### 2020-08-02 440b1fb
Added PypeIt ginga plugin to show wavelengths.
