+++ CHANGES.rst
@@ -4,7 +4,7 @@
  - Implemented a PypeIt ginga plugin kindly provided by the developers of ginga.
    It allows us to display images with a secondary wavelength image. It removes the need to
    constantly use @profxj ginga fork, and fixes a bug whereby WCS image registration
-   was broken when using the forked ginga. 
+   was broken when using the forked ginga.
 
 1.0.6 (22 Jul 2020)
 -------------------
