### 2020-08-02 26e94c8
Added PypeIt ginga plugin to show wavelengths.
