### 2020-08-02 0ad6edf
Added PypeIt ginga plugin to show wavelengths.
