+++ pypeit/scripts/coadd_datacube.py
@@ -34,8 +34,6 @@ def parser(options=None):
                                      formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
     parser.add_argument('file', type = str, default=None, help='ascii file with list of spec2D files to combine')
-    parser.add_argument('--list', default=False, help='List the extensions only?',
-                        action='store_true')
     parser.add_argument('--det', default=1, type=int, help="Detector")
     parser.add_argument('-o', '--overwrite', default=False, action='store_true',
                         help='Overwrite any existing files/directories')
@@ -181,20 +179,33 @@ def generate_masterWCS(crval, cdelt, equinox=2000.0):
 
 
 def main(args):
-
-    # List only?
-    if args.list:
-        hdu = fits.open(args.file)
-        hdu.info()
-        return
-
     # Get a list of files for the combination
     files = open(args.file, 'r').readlines()
+    filelist = []
+    for fil in files:
+        filelist.append(fil.rstrip("\n"))
+
+    # Coadd the files
+    coadd_cube(filelist, det=args.det, overwrite=args.overwrite)
+
+
+def coadd_cube(files, det=0, overwrite=False):
+    """ Main routine to coadd spec2D files
+
+    Args:
+        files (list):
+            List of all spec2D files
+        det (int):
+            detector
+        overwrite (bool):
+            Overwrite the output file, if it exists?
+    """
+    # prep
     numfiles = len(files)
     combine = True if numfiles > 1 else False
 
     all_ra, all_dec, all_wave = np.array([]), np.array([]), np.array([])
-    all_sci, all_ivar, all_idx = np.array([]), np.array([]), np.array([])
+    all_sci, all_ivar, all_idx, all_wghts = np.array([]), np.array([]), np.array([]), np.array([])
     all_wcs = []
     dspat = None  # binning size on the sky (in arcsec)
     ref_scale = None  # This will be used to correct relative scaling among the various input frames
@@ -202,8 +213,7 @@ def main(args):
     weights = np.ones(numfiles)  # Weights to use when combining cubes
     for ff, fil in enumerate(files):
         # Load it up
-        spec2DObj = spec2dobj.Spec2DObj.from_file(fil.rstrip("\n"), args.det)
-        print(spec2DObj.sci_spat_flexure)
+        spec2DObj = spec2dobj.Spec2DObj.from_file(fil, det)
 
         # Load the spectrograph
         specname = spec2DObj.head0['SPECTROG']
@@ -289,7 +299,7 @@ def main(args):
 
         # Calculate the weights relative to the zeroth cube
         if ff != 0:
-            weights[ff] = np.median(flux_sav[resrt]/ivar_sav[resrt])
+            weights[ff] = np.median(flux_sav[resrt]*np.sqrt(ivar_sav[resrt]))**2
 
         # Store the information
         numpix = raimg[onslit_gpm].size
@@ -299,6 +309,7 @@ def main(args):
         all_sci = np.append(all_sci, flux_sav[resrt].copy())
         all_ivar = np.append(all_ivar, ivar_sav[resrt].copy())
         all_idx = np.append(all_idx, ff*np.ones(numpix))
+        all_wghts = np.append(all_wghts, weights[ff]*np.ones(numpix))
 
     # Grab cos(dec) for convenience
     cosdec = np.cos(np.mean(all_dec) * np.pi / 180.0)
@@ -318,17 +329,16 @@ def main(args):
 
[…]
