### 2020-08-03 pypeit/PypeIt#949 @Gabriel-p
Hi again. The installation procedure is a bit confusing:

1. First you recommend to use `conda`, but give no one-line command to install all the dependencies (which is possible) or a `yaml` file with all the dependencies. Why is this? Could you provide one?
2. The possibility of using `pip install -r path/requirements.txt` is mentioned *after* the instructions to install `pypeit`. Shouldn't this be an alternative to the `conda` installation? Why is it located here?
