### 2020-08-03 pypeit/PypeIt#949 @Gabriel-p
3. **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
I guess this is covered by "*Python based data reduction pipeline (DRP) written oringinally for echelle spectroscopy and since expanded to low-resolution spectrometers*", but it could be expanded a bit with some more details. Also, notice the typo in "oringinally".
