### 2020-08-03 2c0890f
Added PypeIt ginga plugin to show wavelengths.
