### 2020-08-03 0d238e9
Added PypeIt ginga plugin to show wavelengths.
