### 2020-08-03 9b184e1
Added PypeIt ginga plugin to show wavelengths.
