+++ pypeit/images/rawimage.py
@@ -14,6 +14,7 @@ from pypeit.core import flat
 from pypeit.core import flexure
 from pypeit.images import pypeitimage
 from pypeit import utils
+from pypeit.ginga import ginga
 
 from IPython import embed
 
@@ -226,7 +227,6 @@ class RawImage(object):
                 msgs.error("Need to provide slits to create illumination flat")
             illum_flat = flatimages.fit2illumflat(slits, flexure_shift=self.spat_flexure_shift)
             if debug:
-                from pypeit.ginga import ginga
                 left, right = slits.select_edges(flexure=self.spat_flexure_shift)
                 viewer, ch = ginga.show_image(illum_flat, chname='illum_flat')
                 ginga.show_slits(viewer, ch, left, right)  # , slits.id)
