+++ pypeit/spectrographs/spectrograph.py
@@ -36,9 +36,8 @@ from pkg_resources import resource_filename
 import numpy as np
 from astropy.io import fits
 
-from linetools import utils as ltu
-
 from pypeit import msgs
+from pypeit import utils
 from pypeit.core.wavecal import wvutils
 from pypeit.core import parse
 from pypeit.core import procimg
@@ -50,7 +49,8 @@ from IPython import embed
 # TODO: Create an EchelleSpectrograph derived class that holds all of
 # the echelle specific methods.
 
-class Spectrograph(object):
+class Spectrograph:
+    # TODO: This docstring needs to be updated.
     """
     Abstract base class whose derived classes dictate
     instrument-specific behavior in PypeIt.
@@ -95,7 +95,6 @@ class Spectrograph(object):
         self.dispname = None
         self.detector = None
         self.naxis = None
-#        self.raw_naxis = None
         self.rawdatasec_img = None
         self.oscansec_img = None
         self.slitmask = None
@@ -106,12 +105,8 @@ class Spectrograph(object):
         # Default extension with the primary header data
         #   used by arsave.save_2d_images
         self.primary_hdrext = 0
-        #self.numhead = 0
 
-        self.minexp = 0  # NEED TO TIE TO INSTRUMENT PAR INSTEAD
-
-        # Init Calibrations Par
-#        self._set_calib_par()
+        self.minexp = 0  # TODO: NEED TO TIE TO INSTRUMENT PAR INSTEAD
 
         # Init meta
         self.meta_data_model = meta.get_meta_data_model()
@@ -516,8 +511,10 @@ class Spectrograph(object):
         # Raw image
         raw_img = hdu[detector['dataext']].data.astype(float)
         # TODO -- Move to FLAMINGOS2 spectrograph
-        # raw data from some spectrograph (i.e. FLAMINGOS2) have an addition extention, so I add the following two lines.
-        # it's easier to change here than writing another get_rawimage function in the spectrograph file.
+        # Raw data from some spectrograph (i.e. FLAMINGOS2) have an
+        # addition extention, so I add the following two lines. It's
+        # easier to change here than writing another get_rawimage
+        # function in the spectrograph file.
         if raw_img.ndim == 3:
             raw_img = raw_img[0]
 
@@ -619,18 +616,11 @@ class Spectrograph(object):
             value: value or list of values
 
         """
-        if isinstance(inp, str):
-            headarr = self.get_headarr(inp)
-        else:
-            headarr = inp
+        headarr = self.get_headarr(inp) if isinstance(inp, str) else inp
 
         # Loop?
         if isinstance(meta_key, list):
-            values = []
-            for mdict in meta_key:
-                values.append(self.get_meta_value(headarr, mdict, required=required))
-            #
-            return values
+            return [self.get_meta_value(headarr, mdict, required=required)) for mdict in meta_key]
 
         # Are we prepared to provide this meta data?
         if meta_key not in self.meta.keys():
@@ -639,6 +629,7 @@ class Spectrograph(object):
             else:
                 msgs.warn("Requested meta data does not exist...")
                 return None
+
         # Is this not derivable?  If so, use the default
         #   or search for it as a compound method
         value = None
@@ -655,6 +646,7 @@ class Spectrograph(object):
                 value = headarr[self.meta[meta_key]['ext']][self.meta[meta_key]['card']]
             except (KeyError, TypeError):
                 value = None
+
         # Return now?
         if no_fussing:
             return value
@@ -665,8 +657,9 @@ class Spectrograph(object):
                                 self.get_meta_value(headarr, 'dec', no_fussing=True))
             value = ra if meta_key == 'ra' else dec
 
[…]

[…]
