+++ CHANGES.rst
@@ -5,6 +5,7 @@
 1.0.6 (22 Jul 2020)
 -------------------
 
+ - (Hotfix) Deal with chk_calibs crash
  - (Hotfix) Deal with wavecalib crash
  - Fix class and version check for DataContainer objects.
  - Script to check for calibration files
