+++ CHANGES.rst
@@ -3,6 +3,7 @@
 --------
 
 - Script to generate combined datacubes for IFU data.
+- Changed numpy (> 1.18.0) and scipy (> 1.4.0) version requirements
 
 1.0.6 (22 Jul 2020)
 -------------------
