### 2020-08-04 4e76ac3
degrade astropy
