+++ pypeit/viewer/__init__.py
@@ -6,8 +6,8 @@ import os.path
 from ginga.misc.Bunch import Bunch
 
 def setup_SlitWavelength():
-    return Bunch(path=os.path.join(os.path.split(__file__)[0], 'ginga_plugin.py'),
-                 module='ginga_plugin', klass='SlitWavelength',
+    return Bunch(path=os.path.join(os.path.split(__file__)[0], 'ginga_plugins.py'),
+                 module='ginga_plugins', klass='SlitWavelength',
                  ptype='global', workspace='right', start=False,
                  category='PypeIt', menu='SlitWavelength', tab='SlitWavelength')
 
