### 2020-08-04 97908f7
deprecate settings and debugger
