+++ pypeit/wavetilts.py
@@ -14,7 +14,7 @@ from matplotlib import pyplot as plt
 from astropy import stats, visualization
 
 from pypeit import msgs, datamodel
-from pypeit.ginga import ginga
+from pypeit import display
 from pypeit.core import arc
 from pypeit.core import tracewave
 
@@ -588,7 +588,7 @@ class BuildWaveTilts(object):
 
         # TODO sort out show methods for debugging
         if show:
-            viewer,ch = ginga.show_image(self.mstilt.image*(self.slitmask > -1),chname='tilts')
+            viewer,ch = display.show_image(self.mstilt.image*(self.slitmask > -1),chname='tilts')
 
         # Loop on all slits
         for slit_idx, slit_spat in enumerate(self.slits.spat_id):
@@ -619,7 +619,7 @@ class BuildWaveTilts(object):
             # TODO: Show the traces before running the 2D fit
 
             if show:
-                ginga.show_tilts(viewer, ch, self.trace_dict)
+                display.show_tilts(viewer, ch, self.trace_dict)
 
             self.spat_order[slit_idx] = self._parse_param(self.par, 'spat_order', slit_idx)
             self.spec_order[slit_idx] = self._parse_param(self.par, 'spec_order', slit_idx)
