+++ pypeit/core/extract.py
@@ -20,7 +20,7 @@ from pypeit import specobj
 from pypeit import specobjs
 from pypeit import tracepca
 from pypeit import bspline
-from pypeit import display
+from pypeit.display import display
 from pypeit.core import pydl
 from pypeit.core import pixels
 from pypeit.core import arc
