+++ pypeit/reduce.py
@@ -19,7 +19,7 @@ from linetools import utils as ltu
 from pypeit import specobjs, specobj
 from pypeit import msgs, utils
 from pypeit import masterframe
-from pypeit import display
+from pypeit.display import display
 from pypeit.core import skysub, extract, wave, flexure, flat
 from pypeit.images import buildimage
 from pypeit import wavecalib
