+++ pypeit/core/flexure.py
@@ -17,7 +17,7 @@ from linetools.spectra import xspectrum1d
 
 from pypeit import msgs
 from pypeit import utils
-from pypeit import display
+from pypeit.display import display
 from pypeit.core import arc
 from pypeit.core import qa
 
