+++ pypeit/display/__init__.py
@@ -5,6 +5,8 @@ import os.path
 
 from ginga.misc.Bunch import Bunch
 
+required_plugins = ['SlitWavelength']
+
 def setup_SlitWavelength():
     return Bunch(path=os.path.join(os.path.split(__file__)[0], 'ginga_plugins.py'),
                  module='ginga_plugins', klass='SlitWavelength',
