+++ pypeit/requirements.txt
@@ -15,7 +15,7 @@ PyYAML>=5.1
 configobj>=5.0.6
 scikit-learn>=0.20
 IPython>=7.2.0
-#ginga>=2.9  -- Can't include ginga as pip will not grab the JXP fork
+ginga>=3.0
 
 # linetools -- This causes readthedocs to barf
 #linetools>=0.2
