+++ pypeit/datamodel.py
@@ -898,7 +898,7 @@ class DataContainer:
         # Allow user to provide single HDU
         if isinstance(hdu, (fits.ImageHDU, fits.BinTableHDU)):
             if ext is not None:
-                warnings.warn('Only one HDU provided; extension number/name is irrelevant.')
+                msgs.warn('Only one HDU provided; extension number/name is irrelevant.')
             ext = [0]
             _hdu = [hdu]
         else:
