### 2020-08-04 fa611d3
doc updates
