+++ doc/api/pypeit.display.ginga_plugins.rst
@@ -1,0 +1,8 @@
+pypeit.display.ginga\_plugins module
+====================================
+
+.. automodule:: pypeit.display.ginga_plugins
+   :members:
+   :private-members:
+   :undoc-members:
+   :show-inheritance:
