+++ doc/installing.rst
@@ -1,30 +1,27 @@
-=================
-Installing PypeIt
-=================
+============
+Installation
+============
 
 This document describes how to install PypeIt for both users and developers.
 
 Installing Dependencies
 =======================
 
-Though we have tried to keep the number of dependencies low, there
-are a few packages that need to be installed (various python packages
-and linetools).
+Installing pypeit will not automatically ensure that all the
+dependencies (and their appropriate versions) are installed and
+up-to-date.  This must be done manually.
 
-We highly recommend that you use Anaconda for the majority of these
-installations.
+The required packages for pypeit are listed in the `requirements.txt
+file
+<https://github.com/pypeit/PypeIt/blob/master/pypeit/requirements.txt>`__.
+If you download the file, you can install/update all the dependencies
+by executing::
 
-Python Dependencies
--------------------
+    pip install -r requirements.txt
 
-The 
-
-PypeIt depends on the following list of Python packages.
-Install these first.
-
-We recommend that you use `Anaconda
-<https://www.continuum.io/downloads/>`_ to install and/or update
-these packages.
+However, we highly recommend that you use `Anaconda
+<https://www.continuum.io/downloads/>`_, as described below. The
+package and version requirements for pypeit are:
 
 * `python <http://www.python.org/>`_ version 3.7 or later
 * `numpy <http://www.numpy.org/>`_ version 1.16.0 or later
@@ -43,97 +40,94 @@ these packages.
 * `linetools <https://pypi.org/project/linetools/>`_ -- version 0.2 or later (see also below)
 * `pytest <https://pypi.org/project/pytest/>`_ -- version 3.0.7 or later
 
-
-If you are using Anaconda, you can check the presence of these packages with::
+With Anaconda, you can check the presence of these packages with::
 
 	conda list "^python$|numpy|astropy$|scipy$|matplotlib|numba|PyQT|ginga|pyyaml|h5py"
 
-If the packages have been installed, this command should print out all the packages and their version numbers.
+If the packages have been installed, this command should print out
+all the packages and their version numbers.
 
-If any of the packages are out of date, they can be updated with a command like::
+If any of the packages are out-of-date, they can be updated with a
+command like::
 
 	conda update scipy
 
+The only exception (true for ginga?) is that you *must* use ``pip
+install`` to install ``linetools``::
 
-Use conda's *pip install* to install the following:
+    pip install linetools
 
-* `linetools <https://github.com/linetools/linetools/>`_ -- version 0.2 or later
+Installing PypeIt
+=================
 
-GitHub cloning
-++++++++++++++
+Please read all of the text in this sub-section before choosing how you
+wish to install ``PypeIt``.
 
-The following packages need to be installed by cloning from GitHub:
+There are two methods you can use to install ``PypeIt``:
 
-* `ginga <https://github.com/profxj/ginga>`_ JXP's fork of Ginga
-
-Do **not** use pip install for this.
-
-To remind you, install via GitHub with a sequence like::
-
-    git clone https://github.com/profxj/ginga
-    cd ginga
-    python setup.py install
-
-This will push the code into your Python distribution.
-
-PypeIt
-======
-
-pip
----
-
-Read all of the text in this sub-section before choosing how you
-wish to install `PypeIt`.
+Install using pip
+-----------------
 
 If you are not using code on the edge of development, then
-we recommend that you install PypeIt with `pip`::
+we recommend that you install ``PypeIt`` with ``pip``::
 
     pip install pypeit
 
-Nuff said, or so we thought because this does not seem to work on all
-OS systems. If that includes you then do::
+This has been known to fail on some systems (and we're working to fix
+the issue). If you have problems, instead try::
 
[…]
