### 2020-08-05 e81a5d1
propogate to chk_flats and chk_edges
