### 2020-08-05 5c6276d
float32 and compressed
