+++ pypeit/spec2dobj.py
@@ -42,7 +42,7 @@ class Spec2DObj(datamodel.DataContainer):
             Primary header if instantiated from a FITS file
 
     """
-    version = '1.0.1'
+    version = '1.0.2'
 
     # TODO 2d data model should be expanded to include:
     # waveimage  --  flexure and heliocentric corrections should be applied to the final waveimage and since this is unique to
@@ -51,9 +51,7 @@ class Spec2DObj(datamodel.DataContainer):
     # tslits_dict -- flexure compensation implies that each frame will have a unique set of slit boundaries, so we probably need to
     #                 write these for each file as well. Alternatively we could just write the offsets to the header.
 
-    # TODO -- Hold, save the non-wavelength images as FLOAT32 ??
-
-    # Becase we are including nested DataContainers, be careful not to duplicate variable names!!
+    # Because we are including nested DataContainers, be careful not to duplicate variable names!!
     datamodel = {
         'sciimg': dict(otype=np.ndarray, atype=np.floating, desc='2D processed science image'),
         'ivarraw': dict(otype=np.ndarray, atype=np.floating, desc='2D processed inverse variance image'),
@@ -143,7 +141,10 @@ class Spec2DObj(datamodel.DataContainer):
             # Array?
             if self.datamodel[key]['otype'] == np.ndarray:
                 tmp = {}
-                tmp[key] = self[key]
+                if self.datamodel[key]['atype'] == np.floating and key not in ['waveimg', 'tilts']:
+                    tmp[key] = self[key].astype(np.float32)
+                else:
+                    tmp[key] = self[key]
                 d.append(tmp)
             # Detector
             elif key == 'detector':
