### 2020-08-05 e7a1571
rm chk_version in _parse
