+++ pypeit/core/arc.py
@@ -1107,10 +1107,13 @@ def fit_arcspec(xarray, yarray, pixt, fitp):
 #            continue  # Probably won't be a good solution
         # Fit the gaussian
         try:
-            pypeitFit = fitting.func_fit(xarray[pmin:pmax], yarray[pmin:pmax], "gaussian", 3)#, return_errors=True)
+            #pypeitFit = fitting.func_fit(xarray[pmin:pmax], yarray[pmin:pmax], "gaussian", 3)#, return_errors=True)
             #popt, pcov = utils.func_fit(xarray[pmin:pmax], yarray[pmin:pmax], "gaussian", 3, return_errors=True)
-            ampl[p], cent[p], widt[p] = pypeitFit.fitc
-            centerr[p] = pypeitFit.fitcov[1, 1]
+            #ampl[p], cent[p], widt[p] = pypeitFit.fitc
+            fitc, fitcov = fitting.fit_gauss(xarray[pmin:pmax], yarray[pmin:pmax])
+            ampl[p], cent[p], widt[p] = fitc  #fitting.fit_gauss(xarray[pmin:pmax], yarray[pmin:pmax])
+            centerr[p] = fitcov[1, 1]
+            #centerr[p] = pypeitFit.fitcov[1, 1]
             #popt, pcov = utils.func_fit(xarray[pmin:pmax], yarray[pmin:pmax], "gaussian", 4, return_errors=True)
             #b[p]    = popt[0]
             #ampl[p] = popt[1]
