### 2020-08-05 cd2b38d
DBSP hacky workaround for missing RA/Dec/AM in calib files
I think the functionality of fixing FITS headers should be moved elsewhere in the future, this is just for prototyping now.
