### 2020-08-05 884a3d3
a start at automating DBSP reduction
to be moved elsewhere later
