### 2020-08-05 64df672
gaussian removed!
