### 2020-08-05 8b3546b
Added PypeIt ginga plugin to show wavelengths.
