### 2020-08-05 76fd90b
Added PypeIt ginga plugin to show wavelengths.
