+++ pypeit/core/arc.py
@@ -90,6 +90,7 @@ def fit2darc(all_wv,all_pix,all_orders,nspec, nspec_coeff=4,norder_coeff=4,sigre
     pypeitFit = fitting.robust_fit(all_pix/xnspecmin1, all_wv_order, (nspec_coeff, norder_coeff), x2=all_orders,
                                    function=func2d, maxiter=100, lower=sigrej, upper=sigrej, minx=min_spec,maxx=max_spec,
                                    minx2=min_order, maxx2=max_order, use_mad=True, sticky=False)
+    embed()
     wv_order_mod = pypeitFit.eval(all_pix/xnspecmin1, x2=all_orders)
     #fitmask, coeff2 = fitting.robust_polyfit_djs(all_pix/xnspecmin1, all_wv_order, (nspec_coeff, norder_coeff),x2=all_orders,
     #                                           function=func2d, maxiter=100, lower=sigrej, upper=sigrej,
@@ -102,8 +103,7 @@ def fit2darc(all_wv,all_pix,all_orders,nspec, nspec_coeff=4,norder_coeff=4,sigre
     fin_rms = pypeitFit.calc_fit_rms(x2=all_orders, apply_mask=True)
     msgs.info("RMS: {0:.5f} Ang*Order#".format(fin_rms))
 
-    orders = np.unique(all_orders)
-
+    #orders = np.unique(all_orders)
     #fit_dict = dict(coeffs=coeff2, orders=orders,
     #                nspec_coeff=nspec_coeff, norder_coeff=norder_coeff,
     #                min_spec=min_spec, max_spec=max_spec,
@@ -112,16 +112,16 @@ def fit2darc(all_wv,all_pix,all_orders,nspec, nspec_coeff=4,norder_coeff=4,sigre
     #                func2d=func2d,xnorm=xnspecmin1,
     #                all_orders=all_orders, all_mask=fitmask)
 
-    debug=True
+    #debug=True
     if debug:
-        fit2darc_global_qa(pypeitFit)
-        fit2darc_orders_qa(pypeitFit)
+        fit2darc_global_qa(pypeitFit, nspec)
+        fit2darc_orders_qa(pypeitFit, nspec)
 
     return pypeitFit
 
 
 
-def fit2darc_global_qa(pypeitFit, outfile=None):
+def fit2darc_global_qa(pypeitFit, nspec, outfile=None):
     """ QA on 2D fit of the wavelength solution.
 
     Parameters
@@ -140,26 +140,18 @@ def fit2darc_global_qa(pypeitFit, outfile=None):
     utils.pyplot_rcparams()
 
     # Extract info from pypeitFit
-    embed()
-    nspec = pypeitFit['nspec']
-    orders = pypeitFit['orders']
-    nspec_coeff = pypeitFit['nspec_coeff']
-    norder_coeff = pypeitFit['norder_coeff']
-    all_wv = pypeitFit['all_wv']
-    all_pix = pypeitFit['all_pix']
-    all_orders = pypeitFit['all_orders']
-    fitmask = pypeitFit['all_mask']
-    coeffs = pypeitFit['coeffs']
-    func2d = pypeitFit['func2d']
-    min_spec = pypeitFit['min_spec']
-    max_spec = pypeitFit['max_spec']
-    min_order = pypeitFit['min_order']
-    max_order = pypeitFit['max_order']
-    xnorm = pypeitFit['xnorm']
+    xnspecmin1 = float(nspec - 1)
+    all_orders = pypeitFit['x2']
+    orders = np.unique(pypeitFit['x2'])
+    all_wv = pypeitFit['yval']/pypeitFit['x2']
[…]
