### 2020-08-05 09490d8
Added PypeIt ginga plugin to show wavelengths.
