### 2020-08-05 22fdf27
Created pypeit_clean script
pypeit_clean removes files created from running tests (except for those which require Cooked)
