+++ CHANGES.rst
@@ -1,6 +1,16 @@
 
 1.0.7dev
 --------
+ - Refactor WaveCalib into a DataContainer
+ - Refactor fitting
+ - Coadd2D bug fixes
+ - Coadd2D without spec1d files
+ - Coadd2D offsets
+ - Coadd2D docs
+ - Manual extraction
+ - Improve LBT/LUCI
+ - Add MMT/MMIRS
+ - QL script for Keck/MOSFIRE
 
 1.0.6 (22 Jul 2020)
 -------------------
@@ -44,14 +54,6 @@
  - Add a script (pypeit_flux_setup) for creating fluxing,
    coadd1d and tellfit pypeit files
  - Add telluric fitting script, pypeit_tellfit
- - Refactor WaveCalib into a DataContainer
- - Refactor fitting
-
- - Coadd2D bug fixes
- - Coadd2D without spec1d files
- - Coadd2D offsets
- - Coadd2D docs
- - Manual extraction
 
 
 
