### 2020-08-06 21c75a0
fix flexure bugs;  turn it back on;  QA is off
