+++ pypeit/core/flexure.py
@@ -253,13 +253,13 @@ def spec_flex_shift(obj_skyspec, arx_skyspec, mxshft=20):
     pypeitFit_obj, _ = fitting.iterfit(obj_skyspec.wavelength.value, obj_skyspec.flux.value,
                                        nord = 3,  kwargs_bspline={'everyn': everyn}, kwargs_reject={'groupbadpix':True,'maxrej':1},
                                        maxiter = 15, upper = 3.0, lower = 3.0)
-    obj_sky_cont = pypeitFit_obj.value(obj_skyspec.wavelength.value)
+    obj_sky_cont, _ = pypeitFit_obj.value(obj_skyspec.wavelength.value)
 
     obj_sky_flux = obj_skyspec.flux.value - obj_sky_cont
     pypeitFit_sky, _ = fitting.iterfit(arx_skyspec.wavelength.value, arx_skyspec.flux.value,
                                        nord = 3,  kwargs_bspline={'everyn': everyn}, kwargs_reject={'groupbadpix':True,'maxrej':1},
                                        maxiter = 15, upper = 3.0, lower = 3.0)
-    arx_sky_cont = pypeitFit_sky.value(arx_skyspec.wavelength.value)
+    arx_sky_cont, _ = pypeitFit_sky.value(arx_skyspec.wavelength.value)
     #pypeitFit_sky= fitting.robust_fit(arx_skyspec.wavelength.value, arx_skyspec.flux.value, 3,
     #                                    function='bspline', lower=3., upper=3., bspline_par=bspline_par)
     #arx_sky_cont = pypeitFit_sky.val(arx_skyspec.wavelength.value)
@@ -286,7 +286,7 @@ def spec_flex_shift(obj_skyspec, arx_skyspec, mxshft=20):
         fit.fit()
         #fit = fitting.func_fit(subpix_grid, corr[subpix_grid.astype(np.int)], 'polynomial', 2)
         success = True
-        max_fit = -0.5 * fit[1] / fit[2]
+        max_fit = -0.5 * fit.fitc[1] / fit.fitc[2]
     else:
         fit = fitting.PypeItFit(xval=subpix_grid, yval=0.0*subpix_grid,
                                 funct='polynomial', order=np.atleast_1d(2))
