### 2020-08-06 ca1a135
pypeit_clean cleans more files
