+++ pypeit/tests/test_chk_calibs.py
@@ -18,6 +18,7 @@ def data_path(filename):
 
 @dev_suite_required
 def test_chk_calibs_not():
+    os.chdir(data_path(''))
     droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/not_alfosc/grism4')
     droot += '/ALD'
     pargs = chk_calibs.parser([droot, '-s', 'not_alfosc'])
@@ -27,6 +28,7 @@ def test_chk_calibs_not():
 
 @dev_suite_required
 def test_chk_calibs_deimos():
+    os.chdir(data_path(''))
     # 830G
     droot = os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA/keck_deimos/830G_M_8600/')
     pargs = chk_calibs.parser([droot, '-s', 'keck_deimos'])
