### 2020-08-06 d19d51d
Fixing more bugs in full_template.
