### 2020-08-06 7aa230a
rm spec1d_b27-J1217p3905_KASTb_2015May20T045733.560.fits
