+++ pypeit/core/wavecal/defs.py
@@ -46,31 +46,30 @@ class LinesBitMask(BitMask):
             ('ZnI', 'Comment'),
             ('HeI', 'Comment'),
             ('FeI', 'Comment'),
+            ('ThAr', 'Comment'),
             ('OH', 'Comment'),
             ('UNKNWN', 'Comment'),
-            ('OH_R24000', 'Comment'),
-            ('OH_triplespec', 'Comment'),
+            #('OH_R24000', 'Comment'),
+            #('OH_triplespec', 'Comment'),
             ('CuI', 'Comment'),
             ('ArII', 'Comment'),
-            ('OH_XSHOOTER', 'Comment'),
-            ('OH_GNIRS', 'Comment'),
-            ('OH_NIRES', 'Comment'),
-            ('ThAr_XSHOOTER_VIS', 'Comment'),
-            ('OH_GMOS', 'Comment'),
-            ('OH_MODS', 'Comment'),
-            ('ThAr_MagE', 'Comment'),
-            ('OH_FIRE_Echelle', 'Comment'),
-            ('Ar_IR_GNIRS', 'Comment'),
+            #('OH_XSHOOTER', 'Comment'),
+            #('OH_GNIRS', 'Comment'),
+            #('OH_NIRES', 'Comment'),
+            #('OH_GMOS', 'Comment'),
+            #('OH_MODS', 'Comment'),
+            #('OH_FIRE_Echelle', 'Comment'),
+            #('Ar_IR_GNIRS', 'Comment'),
+            ('Ar', 'This is for Ar_IR_GNIRS which should specify the real ion'),
         ])
         super(LinesBitMask, self).__init__(list(mask.keys()), descr=list(mask.values()))
 
 
 def lines():
     """
-    Dict of lines included in this database
+    Dict of line lists included in this database
 
     .. warning::
-
         Modifying any of the following is a *bad* idea.  Adding is ok
 
     Returns
@@ -102,8 +101,6 @@ def lines():
     line_dict['OH_FIRE_Echelle'] = 2**19  # R=6000
     line_dict['Ar_IR_GNIRS'] = 2**20  # R=6000
     line_dict['UNKNWN'] = 2**21
-
-
 
     #
     return line_dict
