+++ pypeit/core/fitting.py
@@ -141,52 +141,19 @@ class PypeItFit(DataContainer):
         # For two-d fits x = x, y = x2, y = z
         if ('2d' in self.func) and (x2_out is not None):
             # Is this a 2d fit?
-            self.fitc = polyfit2d_general(x_out, x2_out, y_out, self.order, w=w_out, function=self.func[:-2],
-                                     minx=self.minx, maxx=self.maxx,
-                                     miny=self.minx2, maxy=self.maxx2)
+            self.fitc, self.minx, self.maxx, self.minx2, self.maxx2 = polyfit2d_general(
+                x_out, x2_out, y_out, self.order, w=w_out, function=self.func[:-2],
+                minx=self.minx, maxx=self.maxx,
+                miny=self.minx2, maxy=self.maxx2)
         elif self.func == "polynomial":
             self.fitc = np.polynomial.polynomial.polyfit(x_out, y_out, self.order[0], w=w_out)
-        elif self.func == "legendre":
-            if self.minx is None or self.maxx is None:
-                if np.size(x_out) == 1:
-                    xmin, xmax = -1.0, 1.0
-                else:
-                    xmin, xmax = np.min(x_out), np.max(x_out)
-            else:
-                xmin, xmax = self.minx, self.maxx
-            xv = 2.0 * (x_out - xmin) / (xmax - xmin) - 1.0
-            self.fitc = np.polynomial.legendre.legfit(xv, y_out, self.order[0], w=w_out)
-        elif self.func == "chebyshev":
-            if self.minx is None or self.maxx is None:
-                if np.size(x_out) == 1:
-                    xmin, xmax = -1.0, 1.0
-                else:
-                    xmin, xmax = np.min(x_out), np.max(x_out)
-            else:
-                xmin, xmax = self.minx, self.maxx
-            xv = 2.0 * (x_out - xmin) / (xmax - xmin) - 1.0
-            self.fitc = np.polynomial.chebyshev.chebfit(xv, y_out, self.order[0], w=w_out)
-#        elif self.func == "moffat":
-#            # Guesses
-#            if guesses is None:
-#                ampl, cent, sigma = guess_gauss(x_out, y_out)
-#                p0 = ampl
-#                p2 = 3.  # Standard guess
-#                p1 = (2.355 * sigma) / (2 * np.sqrt(2 ** (1. / p2) - 1))
-#            else:
-#                p0, p1, p2 = guesses
-#            # Error
-#            if w_out is not None:
-#                sig_y = 1. / w_out
-#            else:
-#                sig_y = None
-#            if self.order[0] == 3:  # Standard 3 parameters
-#                self.fitc, self.fitcov = curve_fit(moffat, x_out, y_out, p0=[p0, p1, p2], sigma=sig_y)
-#            else:
-#                msgs.error("Not prepared for deg={:d} for Moffat fit".format(self.order[0]))
+        elif self.func == "legendre" or self.func == "chebyshev":
+            xv, self.minx, self.maxx = scale_minmax(x_out, minx=self.minx, maxx=self.maxx)
+            self.fitc = np.polynomial.legendre.legfit(xv, y_out, self.order[0], w=w_out)  \
+                if self.func == "legendre" else np.polynomial.chebyshev.chebfit(xv, y_out, self.order[0], w=w_out)
         else:
             msgs.error("Fitting function '{0:s}' is not implemented yet" + msgs.newline() +
-                       "Please choose from 'polynomial', 'legendre', 'chebyshev','bspline'")
[…]
