### 2020-08-06 502e9f6
Fixing more bugs in full_template.
