+++ pypeit/core/tracewave.py
@@ -649,6 +649,7 @@ def fit_tilts(trc_tilt_dict, thismask, slit_cen, spat_order=3, spec_order=4, max
 
     tilts_ivar = utils.inverse((tilts_sigma.flatten() / xnspecmin1) ** 2)
     #fitmask, coeff2 = fitting.robust_fit(tilts_spec.flatten() / xnspecmin1,
+    embed(header='652 of tracewave')
     pypeitFit = fitting.robust_fit(tilts_spec.flatten() / xnspecmin1,
                                                (tilts.flatten() - tilts_spec.flatten()) / xnspecmin1,
                                                fitxy, x2=tilts_dspat.flatten() / xnspatmin1,
@@ -656,7 +657,7 @@ def fit_tilts(trc_tilt_dict, thismask, slit_cen, spat_order=3, spec_order=4, max
                                                function=func2d, maxiter=maxiter, lower=sigrej,
                                                upper=sigrej, maxdev=maxdev_pix / xnspecmin1,
                                                minx=-0.0, maxx=1.0, minx2=-1.0, maxx2=1.0,
-                                               use_mad=False, sticky=False)
+                                               use_mad=False, sticky=False, debug=True)
     fitmask = pypeitFit.bool_gpm.reshape(tilts_dspat.shape)
     # Compute a rejection mask that we will use later. These are
     # locations that were fit but were rejectedK
