+++ CHANGES.rst
@@ -7,7 +7,7 @@
  - Script to generate combined datacubes for IFU data.
  - Changed numpy (> 1.18.0) and scipy (> 1.4.0) version requirements
  - Allow show2d_spec, chk_edges, chk_flats to load older Spec2DObj datamodel versions
- - Modify Spec2DObj 2D model for float32 images and compressed
+ - Modify Spec2DObj 2D model for float32 images
 
 1.0.6 (22 Jul 2020)
 -------------------
