### 2020-08-06 9c1cf58
DBSP generate and use template arc spectra
