### 2020-08-06 85e1743
Removed DBSP-specific automation files
