+++ table_edit.py
@@ -1,192 +1,0 @@
-# Interactive Table Editor
-import tkinter as tk
-from tkinter.ttk import Treeview
-
-from astropy.table import Table
-import matplotlib.pyplot as plt
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-from PyQt5.QtCore import Qt
-
-from astropy.coordinates import Angle, SkyCoord, EarthLocation, AltAz, ICRS
-from astropy.time import Time
-import astropy
-from astropy import units as u
-
-# Some Thoughts on how this could work
-# have modified table and path variable
-# on exit, for each modified row, edit the corresponding FITS file
-
-
-# TODO: allow for deleting row. low priority
-
-loc = EarthLocation.of_site('Palomar')
-
-def update_airmass(row: astropy.table.Row):
-    skycoord = SkyCoord(row['ra'], row['dec'], unit=u.deg)
-    time = Time(row['mjd'], format='mjd')
-    altaz = AltAz(obstime=time, location=loc)
-    row['airmass'] = skycoord.transform_to(altaz).secz
-
-def get_zenith_ra_dec(time) -> SkyCoord:
-    time = Time(time, format='mjd')
-    altaz = AltAz(alt=Angle(90, unit=u.deg), az=Angle(0, unit=u.deg), obstime=time, location=loc)
-    return altaz.transform_to(ICRS)
-
-class TableModel(QtCore.QAbstractTableModel):
-    def __init__(self, data: Table, cols = None):
-        super(TableModel, self).__init__()
-        self._data = data #Table(data, masked = True, copy=False)
-        self._cols = cols if cols is not None else data.colnames
-        self._mask = self._data.mask
-        self._rowCount = len(self._data)
-        self._colCount = len(self._cols)
-    
-    def data(self, index, role):
-        if role == Qt.DisplayRole or role == Qt.EditRole:
-            col = self._cols[index.column()]
-            value = self._data[col][index.row()]
-            if col == 'ra':
-                ra = Angle(value, unit=u.deg)
-                value = ra.to_string(unit=u.hour)
-            elif col == 'dec':
-                dec = Angle(value, unit=u.deg)
-                value = dec.to_string()
-            elif col == 'dispangle':
-                ang = Angle(value, unit=u.deg)
-                value = ang.to_string()
-            elif col == 'mjd' or col == 'airmass':
-                value = f'{value:.4f}'
-            return str(value)
-        if role == Qt.BackgroundRole:
-            col = self._cols[index.column()]
-            masked = self._mask[col][index.row()]
-            if masked:
-                return QtGui.QColor(Qt.red)
-    
-    def setData(self, index: QtCore.QModelIndex, value: QtCore.QVariant, role = Qt.EditRole):
-        try:
-            col = self._cols[index.column()]
-            if col == 'ra':
-                ra = Angle(value, unit=u.hour)
-                self._data[col][index.row()] = ra.degree
-                if self._data['dec'][index.row()]:
-                    update_airmass(self._data[index.row()])
-                self.dataChanged.emit(index, index)
-                return True
-            if col == 'dec':
-                dec = Angle(value, unit=u.deg)
-                self._data[col][index.row()] = dec.degree
-                if self._data['dec'][index.row()]:
-                    update_airmass(self._data[index.row()])
-                self.dataChanged.emit(index, index)
-                return True
-            self._data[col][index.row()] = value
-            self.dataChanged.emit(index, index)
-            return True
-        except:
-            return False
-    
-    def flags(self, index):
[…]
