### 2020-08-07 ee19dcd
Address some PR comments
