+++ pypeit/spectrographs/p200_dbsp.py
@@ -3,10 +3,9 @@
 import numpy as np
 
 from astropy.io import fits
-from astropy.coordinates import Angle, SkyCoord, EarthLocation, AltAz, ICRS
+from astropy.coordinates import Angle
 from astropy import units as u
 from astropy.time import Time
-from pkg_resources import resource_filename
 
 from pypeit import msgs
 from pypeit import telescopes
@@ -16,15 +15,10 @@ from pypeit.spectrographs import spectrograph
 from pypeit.core import parse
 from pypeit.images import detector_container
 
-from pypeit import debugger
 from typing import List
 
-loc = EarthLocation.of_site('Palomar')
-
-def get_zenith_ra_dec(time) -> SkyCoord:
-    time = Time(time)
-    altaz = AltAz(alt=Angle(90, unit=u.deg), az=Angle(0, unit=u.deg), obstime=time, location=loc)
-    return altaz.transform_to(ICRS)
+def flip_fits_slice(s: str) -> str:
+    return '[' + ','.join(s.strip('[]').split(',')[::-1]) + ']'
 
 class P200DBSPSpectrograph(spectrograph.Spectrograph):
     """
@@ -63,11 +57,8 @@ class P200DBSPSpectrograph(spectrograph.Spectrograph):
         """
         meta = {}
         # Required (core)
-        # VERY HACKY!!!!
-        meta['ra'] = dict(card=None, compound=True)
-        meta['dec'] = dict(card=None, compound=True)
-        #meta['ra'] = dict(ext=0, card='RA')
-        #meta['dec'] = dict(ext=0, card='DEC')
+        meta['ra'] = dict(ext=0, card='RA', required_ftypes=['science', 'standard'])
+        meta['dec'] = dict(ext=0, card='DEC', required_ftypes=['science', 'standard'])
         meta['target'] = dict(ext=0, card='OBJECT')
 
         meta['dispname'] = dict(ext=0, card='GRATING')
@@ -76,9 +67,7 @@ class P200DBSPSpectrograph(spectrograph.Spectrograph):
 
         meta['mjd'] = dict(card=None, compound=True)
         meta['exptime'] = dict(ext=0, card='EXPTIME')
-        #meta['airmass'] = dict(ext=0, card='AIRMASS')
-        # VERY HACKY!!!!
-        meta['airmass'] = dict(card=None, compound=True)
+        meta['airmass'] = dict(ext=0, card='AIRMASS', required_ftypes=['science', 'standard'])
 
         # Extras for config and frametyping
         meta['dichroic'] = dict(ext=0, card='DICHROIC')
@@ -91,7 +80,23 @@ class P200DBSPSpectrograph(spectrograph.Spectrograph):
         # Ingest
         self.meta = meta
 
-    def compound_meta(self, headarr: List[fits.Header], meta_key):
+    def compound_meta(self, headarr: List[fits.Header], meta_key: str):
+        """
+        Methods to generate meta in a more complex manner than simply
+        reading from the header.
+
+        mjd is converted from UTSHUT header
+        dispangle is parsed from ANGLE header
+
+        Args:
+            headarr: List[fits.Header]
+              List of headers
+            meta_key: str
+
+        Returns:
+            value:
+
+        """
         if meta_key == 'mjd':
             return Time(headarr[0]['UTSHUT']).mjd
         elif meta_key == 'dispangle':
@@ -100,37 +105,32 @@ class P200DBSPSpectrograph(spectrograph.Spectrograph):
             except Exception as e:
                 print(headarr[0]['ANGLE'])
                 raise e
-        elif meta_key == 'ra':
-            ra = headarr[0].get('RA', default=None)
-            frametype = headarr[0]['IMGTYPE']
-            if ra is None and frametype in ['bias', 'flat', 'cal']:
[…]
