### 2020-08-10 #2308 @astrom-tom
Hi @harpolea
I am reviewing this software, it seems that there is no paper.md in the repository. Is that normal?
