### 2020-08-10 #2308 @harpolea
Hi @astrom-tom – sometimes authors choose to put their paper in a different branch of the repository so as not to clutter the main repository. For this repo, the paper is in the `joss-paper-branch` branch: https://github.com/pypeit/PypeIt/tree/joss-paper-branch/paper
