### 2020-08-10 #2308 @astrom-tom
@harpolea ,

Sorry to bug you, this is the first time I am reviewing for JOSS.
- For each identified problem/remarks I need to open an issue on the repo right?
- Once I am done with the review, what should I do?
