### 2020-08-10 #2308 @harpolea
@astrom-tom that's totally fine – feel free to ask about anything you're unsure of!

Yes, if you identify any problems or have any suggestions, then please open an issue in the original repo. It's normally useful for the repo maintainers if you include 'JOSS' in the issue title so they can easily identify what the issue relates to, but obviously different organisations handle this in different ways, so go with whatever the repo's local style is! Also, if you include this issue URL somewhere in the issues you open, then the issue should appear in this thread as being 'mentioned'.

Once you are finished with your review, it would be helpful if you can give a summary here. If you have ticked off everything in the checklist and don't think there's anything that needs to be changed or added, then you can simply say that you've looked over everything and recommend it for being accepted. If you have opened some issues and feel some changes should be made before the submission is accepted, then please give a brief overview of what changes you'd like to see (feel free to link to any issues you've opened, but if you've linked them as I suggested above then that won't necessary as they will already appear in this thread).
