### 2020-08-10 #2308 @astrom-tom
Dear @kthyng, @harpolea

I went throught the software and documentation, you can find my review below. Please let know if there is something missing. I am going to open issue for each of the 4 points highlighted below:

**1-Paper**
The paper is nicely written. I have a single remark about it: it could be useful to mention that pypeit is a set of commands design to perform the reduction (i.e. no need to know how to code in Python to use it).

**2-Installation**
I could install it with  the git clone section of the installation page. Although, I got an error:

```
Traceback (most recent call last):
  File "setup.py", line 12, in <module>
    from extension_helpers import get_extensions
ModuleNotFoundError: No module named 'extension_helpers'
```

After installing it manually, the python setup.py develop --user worked. **This should be added to the setup.py** of the package so it is installed automatically. The pip install from pypi crashed.

**3-Documentation:**
I believe a substantial work could be done about the documentation. I believe the authors could take example from existing documentation (e.g. https://fpavogt.github.io/fcmaker/index.html). My general impression is that the documentation is massive and that it is very hard to navigate in it. The documentation aims at providing help to people who want to use your code. For the moment it is not the case. I provide some details below where some improvements could be implemented.

_-Front page_

- I believe the front page could start with a quick overview of the code, with a list of instruments that are supported. It could be a quick summary of the paper.
- A link back to the repository should be added as well.
- A list of contributors could be added, as well a part about 'contributing' to the software.
- A 'change log' section can also be added.
- The quick link sections of the front page (Running PypeIt, Data Products, Calibrations, etc...) can safely be removed. You can make a page for each section and link if you want from the front page. But just providing the list of links is not helpful at all. For example, the section 'spectrograph' links to multiple pages. Why not making a single page about spectrograph with each instrument as sections? This would make the documentation more compact and clearer (and would remove a lot of pages from the doctree, see below).
- Why is VLT-XShooter/Gemini GMOS in the Orphaned docs? Should be in the spectrograph documentation maybe?

_-About the Menu (left black):_

- The shortcut Menu on the left could be largely cleaned up. As most pages are links from PypeIT Cookbook pages, I don't think it is necessary to add them in the Menu (it makes it very confusing). Here should just go the general sections, but not all individual pages.

 -   An optimal doctree could be :
        - Home (back to front page)
        - Installing PypeIt
        - PypeIt Cookbook
        - Spectrograph
        - Data Products
        - Calibrations
        - Master Frames (Although this could be part of Calibrations)
        - Reduction Algorithm
        - Miscellaneous (Documentation)
        - For developers (including testing)

-Code of conduct could go at the bottom of the doctree as it is intended firstly to developers.

_-Functionality documentation:_
    The 'pypeit package' page, with the list of modules, returns empty pages (just the name of the modules as page titles). The same goes for the 'PypeIt API' page. Why did you separate 'PypeIt API' from 'PypeIt Modules'?

_-Installation page:_
    At the bottom, in the blue note, what is the difference between 'all' and 'kast' options for tests?
    Please state it in that note.

_-Example usage:_
It would be nice, for the user (and also for the purpose of this review), to **add a 'step-by-step' page,** where you explicit everything from A-to-Z with a set of data that you provide and the expected results. **I will wait to have something like in order to comment on software functionality.**

**4-Testing:**

What is the difference between [2-Run the PypeIt unit tests] and [3-Try the test suite]? It should be clearly stated. For the moment I ran only the [2-Run the PypeIt unit tests] but I would like to know if I need to run the [3-Try the test suite]?.

I tried to run the unit tests with 'python setup.py tests'. It complained about missing numba, pytest, ginga and linetools. **This should be installed automatically during the pypeIt installation, please add it to the dependencies in the documentation AND in the setup.py.**

Results: -----144 passed, 89 skipped, 536 warnings in 33.60s------

Then I cloned the dev suite. The tests gave me the following error:

```
Traceback (most recent call last):
  File "./pypeit_test", line 772, in <module>
    raise EnvironmentError('Enviroment must include PYPEIT_DEV, which points to the '
OSError: Enviroment must include PYPEIT_DEV, which points to the top-level directory with the dev-suite raw data.
```

Should this be taken care of with the installation? It is only after that I realised that the full instruction for running the dev suite is back on github! Please create a page with testing instruction on the documentation (see above). Downloading the data is not simple: The author proposes to use Google File stream which is not available to me ('To use Drive File Stream, you need to have a Google Account through work or school'). I will wait to have a clarification on the difference between [2-Run the PypeIt unit tests] and [3-Try the test suite] before running these tests.
