### 2020-08-10 pypeit/PypeIt#992 @astrom-tom
JOSS review: paper
Dear @profxj ,

First of all, sorry for the long delay for the review. I am opening few ticket related to the submission of your code to JOSS [https://github.com/openjournals/joss-reviews/issues/2308](url)

Here is the first one about the paper itself:

The paper is nicely written. I have a single remark about it: it could be useful to mention that pypeit is a set of commands design to perform the reduction (i.e. no need to know how to code in Python to use it).
