### 2020-08-10 pypeit/PypeIt#993 @astrom-tom
JOSS review: Installation
Review of paper : [https://github.com/openjournals/joss-reviews/issues/2308](url)

Here is the one about installation:

I could install it with the git clone section of the installation page. Although, I got an error:

```
Traceback (most recent call last):
  File "setup.py", line 12, in <module>
    from extension_helpers import get_extensions
ModuleNotFoundError: No module named 'extension_helpers'
```

After installing it manually, the python setup.py develop --user worked. This package should be added to the setup.py of the package so it is installed automatically. The pip install from pypi crashed.
