### 2020-08-10 pypeit/PypeIt#994 @astrom-tom
JOSS review: documentation
Review of paper : [https://github.com/openjournals/joss-reviews/issues/2308](url)

On the documentation:

3-Documentation:
I believe a substantial work could be done about the documentation. I believe the authors could take example from existing documentation (e.g. https://fpavogt.github.io/fcmaker/index.html). My general impression is that the documentation is massive and that it is very hard to navigate in it. The documentation aims at providing help to people who want to use your code. For the moment it is not the case. I provide some details below where some improvements could be implemented.

-Front page

- I believe the front page could start with a quick overview of the code, with a list of instruments that are supported. It could be a quick summary of the paper.
- A link back to the repository should be added as well.
- A list of contributors could be added, as well a part about 'contributing' to the software.
- A 'change log' section can also be added.
- The quick link sections of the front page (Running PypeIt, Data Products, Calibrations, etc...) can safely be removed. You can make a page for each section and link if you want from the front page. But just providing the list of links is not helpful at all. For example, the section 'spectrograph' links to multiple pages. Why not making a single page about spectrograph with each instrument as sections? This would make the documentation more compact and clearer (and would remove a lot of pages from the doctree, see below).
- Why is VLT-XShooter/Gemini GMOS in the Orphaned docs? Should be in the spectrograph documentation maybe?

-About the Menu (left side):

- The shortcut Menu on the left could be largely cleaned up. As most pages are links from PypeIT Cookbook pages, I don't think it is necessary to add them in the Menu (it makes it very confusing). Here should just go the general sections, but not all individual pages.

 - An optimal doctree could be :
        - Home (back to front page)
        - Installing PypeIt
        - PypeIt Cookbook
        - Spectrograph
        - Data Products
        - Calibrations
        - Master Frames (Although this could be part of Calibrations)
        - Reduction Algorithm
        - Miscellaneous (Documentation)
        - For developers (including testing)

-Code of conduct could go at the bottom of the doctree as it is intended firstly to developers.

-Functionality documentation:
The 'pypeit package' page, with the list of modules, returns empty pages (just the name of the modules as page titles). The same goes for the 'PypeIt API' page. Why did you separate 'PypeIt API' from 'PypeIt Modules'?

-Installation page:
At the bottom, in the blue note, what is the difference between 'all' and 'kast' options for tests?
Please state it in that note.

-Example usage:
It would be nice, for the user (and also for the purpose of this review), to add a 'step-by-step' page, where you explicit everything from A-to-Z with a set of data that you provide and the expected results. I will wait to have something like in order to comment on software functionality.
