### 2020-08-10 pypeit/PypeIt#995 @astrom-tom
JOSS review: testing
Review of paper : [https://github.com/openjournals/joss-reviews/issues/2308
](url)
About testing:

What is the difference between [2-Run the PypeIt unit tests] and [3-Try the test suite]? It should be clearly stated. For the moment I ran only the [2-Run the PypeIt unit tests] but I would like to know if I need to run the [3-Try the test suite]?.

I tried to run the unit tests with 'python setup.py tests'. It complained about missing **numba, pytest, ginga and linetools**. This should be installed automatically during the pypeIt installation, please add it to the dependencies in the documentation AND in the setup.py.

Results: -----144 passed, 89 skipped, 536 warnings in 33.60s------

Then I cloned the dev suite. The tests gave me the following error:

```
Traceback (most recent call last):
  File "./pypeit_test", line 772, in <module>
    raise EnvironmentError('Enviroment must include PYPEIT_DEV, which points to the '
OSError: Enviroment must include PYPEIT_DEV, which points to the top-level directory with the dev-suite raw data.
```

Should this be taken care of with the installation? It is only after that I realised that the full instruction for running the dev suite is back on github! Please create a page with testing instruction on the documentation (see above). Downloading the data is not simple: The author proposes to use Google File stream which is not available to me ('To use Drive File Stream, you need to have a Google Account through work or school'). I will wait to have a clarification on the difference between [2-Run the PypeIt unit tests] and [3-Try the test suite] before running these tests.
@astrom-tom
