### 2020-08-11 bc4bf5c
Make tests clean up after themselves
All test scripts cleanup any files they write.
