+++ pypeit/tests/test_scripts.py
@@ -35,6 +35,7 @@ def test_quicklook():
         raise IOError("You need to get the CALIBS folder as described above!!")
 
     # Define the output directories (HARDCODED!!)
+    cdir = os.getcwd()
     os.chdir(data_path(''))
     outdir = data_path('keck_lris_blue_A')
     # Remove them if they already exist
@@ -49,6 +50,10 @@ def test_quicklook():
                                '--user_pixflat={0}'.format(
                                    os.path.join(calib_dir,
                                         'PYPEIT_LRISb_pixflat_B600_2x2_17sep2009.fits.gz'))]))
+    
+    # Cleanup
+    shutil.rmtree(outdir)
+    os.chdir(cdir)
 
 @dev_suite_required
 def test_trace_edges():
