### 2020-08-11 e0d1795
Address more PR comments
