### 2020-08-11 7cbc2aa
Clean up DBSP template creation
