### 2020-08-11 52cc458
Improve performance of spec_flex_shift and spec_flexure_obj
