+++ pypeit/edgetrace.py
@@ -345,7 +345,7 @@ class EdgeTraceSet(DataContainer):
             trace centroid; see :class:`EdgeTraceBitMask`.
         spat_fit (`numpy.ndarray`_):
             A model fit to the `spat_cen` data.
-        spat_fit_type (:obj:`str`):
+        fittype (:obj:`str`):
             An informational string identifier for the type of model
             used to fit the trace data.
         pca (:obj:`list`, :class:`pypeit.tracepca.TracePCA`):
@@ -355,7 +355,7 @@ class EdgeTraceSet(DataContainer):
             :class:`pypeit.tracepca.TracePCA` objects if the PCA
             decomposition is peformed for the left (`pca[0]`) and
             right (`pca[1]`) traces separately.
-        pca_type (:obj:`str`)
+        pcatype (:obj:`str`)
             An informational string indicating which data were used
             in the PCA decomposition, 'center' for `spat_cen` or
             'fit' for `spat_fit`.
@@ -386,30 +386,57 @@ class EdgeTraceSet(DataContainer):
     """DataContainer datamodel version."""
 
     datamodel = {'PYP_SPEC': dict(otype=str, descr='PypeIt spectrograph name'),
+                 'dispname': dict(otype=str,
+                                  descr='Spectrograph disperser name.  Primarily needed for '
+                                        'reloading an existing MasterEdge file.'),
                  'trace_img': dict(otype=TraceImage,
                                    descr='Image used to construct the edge traces.'),
                  'nspec': dict(otype=int, descr='Image pixels in the spectral direction.'),
                  'nspat': dict(otype=int, descr='Image pixels in the spatial direction.'),
                  'bpm': dict(otype=np.ndarray, atype=(bool, np.bool_),
-                             descr='Bad-pixel mask for trace image'
-
-
-                 'specmax': dict(otype=np.ndarray, atype=np.floating,
-                                descr='Maximum spectral position allowed for each slit/order.  '
-                                      'Shape is Nslits.')}
+                             descr='Bad-pixel mask for trace image'),
+                 'sobel_sig': dict(otype=np.ndarray, atype=(float, np.floating),
+                                   descr='Sobel-filtered image used to detect edges'),
+                 'traceid': dict(otype=np.ndarray, atype=(int, np.integer),
+                                 descr='ID number for the edge traces.  Negative and positive '
+                                       'IDs are for, respectively, left and right edges.'),
+                 'orderid': dict(otype=np.ndarray, atype=(int, np.integer),
+                                 descr='For echelle spectrographs, this is the order ID number '
+                                       'for the edge traces.  Negative and positive IDs are for, '
+                                       'respectively, left and right edges.'),
+                 'spat_cen': dict(otype=np.ndarray, atype=(float, np.floating),
+                                  descr='(Floating-point) Measured spatial coordinate of the '
+                                        'edge traces for each spectral pixel.  Shape is '
+                                        '(Nspec,Ntrace).'),
+                 'spat_err': dict(otype=np.ndarray, atype=(float, np.floating),
+                                  descr='Error in the measured spatial coordinate edge traces.'),
+                 'spat_msk': dict(otype=np.ndarray, atype=bitmask.minimum_dtype(),
+                                  descr='Bitmask for the edge trace positions.'),
+                 'spat_fit': dict(otype=np.ndarray, atype=(float, np.floating),
+                                  descr='The best-fit model result for the trace edge.'),
+                 'fittype': dict(otype=str,
+                                 descr='An informational string identifying the type of model '
+                                       'used to fit the trace data.  Either ``pca`` for a PCA '
+                                       'decomposition or the polynomial function type and order'),
+                 'pca': dict(otype=TracePCA,
[…]
