### 2020-08-12 pypeit/PypeIt#992 @profxj
Added in this commit:  https://github.com/pypeit/PypeIt/commit/315e54224125bd0cf23e1ef2b9b39258e63d5b37#diff-e2778e7674f45b806282a9611faa7220
