+++ doc/installing.rst
@@ -34,7 +34,6 @@ package and version requirements for pypeit are:
 * `configobj <https://pypi.org/project/configobj/>`_ -- version 5.0.6 or later
 * `scikit-learn <https://scikit-learn.org/stable/>`_ -- version 0.20 or later
 * `IPython <https://ipython.org>`_ -- version 7.2.0 or later
-* `extension-helpers <https://pypi.org/project/extension-helpers/>`_ -- version 0.1 or later
 * `ginga <https://pypi.org/project/ginga/>`_ -- version 3.0 or later
 * `packaging <https://pypi.org/project/packaging/>`_ -- version 19.0 or later
 * `linetools <https://pypi.org/project/linetools/>`_ -- version 0.2 or later (see also below)
@@ -43,7 +42,7 @@ package and version requirements for pypeit are:
 
 With Anaconda, you can check the presence of these packages with::
 
-	conda list "^python$|numpy|astropy$|scipy$|matplotlib|numba|PyQT|ginga|pyyaml|h5py"
+	conda list "^python$|numpy|astropy$|scipy$|matplotlib|numba|PyQT|ginga|pyyaml"
 
 If the packages have been installed, this command should print out
 all the packages and their version numbers.
@@ -225,9 +224,9 @@ To run the test::
 
 	``pypeit_test`` can also, e.g., take the argument ``kast``
 	instead of ``develop`` to only test data from the Shane Kast
-	spectrograph.
+	spectrograph instead of the full development suite.
 
-The test takes a while to run but should run without issue if all the
+The test takes a (long) while to run but should run without issue if all the
 packages have been properly installed.
 
 Developers
