+++ doc/pypeit_par.rst
@@ -163,6 +163,8 @@ Current PypeItPar Parameter Hierarchy
 
         ``[[extraction]]``: `ExtractionPar Keywords`_
 
+            ``[[[manual]]]``: `ManualExtractionPar Keywords`_
+
         ``[[cube]]``: `CubePar Keywords`_
 
     ``[flexure]``: `FlexurePar Keywords`_
@@ -432,6 +434,7 @@ Key                   Type        Options  Default     Description
 ====================  ==========  =======  ==========  ======================================================================================================================================================================================================================================================================================================================================================================================================
 ``coaddfile``         str         ..       ..          Output filename                                                                                                                                                                                                                                                                                                                                                                                       
 ``ex_value``          str         ..       ``OPT``     The extraction to coadd, i.e. optimal or boxcar. Must be either 'OPT' or 'BOX'                                                                                                                                                                                                                                                                                                                        
+``extrap_sens``       bool        ..       False       If True (default), the code will barf in Echelle mode if one tries to use sensfunc beyond its allowed space.                                                                                                                                                                                                                                                                                          
 ``filter``            str         ..       ``none``    Filter for scaling.  See flux_calib.load_fitler_file() for naming.  Ignore if none                                                                                                                                                                                                                                                                                                                    
 ``filter_mag``        float       ..       ..          Magnitude of the source in the given filter                                                                                                                                                                                                                                                                                                                                                           
 ``filter_mask``       str, list   ..       ..          List of wavelength regions to mask when doing the scaling (ie. occasional junk pixels).Colon and comma separateed, e.g.   5552:5559,6010:6030                                                                                                                                                                                                                                                         
@@ -464,12 +467,13 @@ Coadd2DPar Keywords
 
 Class Instantiation: :class:`pypeit.par.pypeitpar.Coadd2DPar`
 
-===========  =========  =======  ========  ===========================================================================
-Key          Type       Options  Default   Description                                                                
-===========  =========  =======  ========  ===========================================================================
-``offsets``  list       ..       ..        User-input list of offsets for the images being combined.                  
-``weights``  str, list  ..       ``auto``  Mode for the weights used to coadd images.  See coadd2d.py for all options.
-===========  =========  =======  ========  ===========================================================================
+====================  =========  =======  ========  ===========================================================================
+Key                   Type       Options  Default   Description                                                                
+====================  =========  =======  ========  ===========================================================================
+``offsets``           list       ..       ..        User-input list of offsets for the images being combined (spat pixels).    
+``use_slits4wvgrid``  bool       ..       False     If True, use the slits to set the trace down the center                    
+``weights``           str, list  ..       ``auto``  Mode for the weights used to coadd images.  See coadd2d.py for all options.
+====================  =========  =======  ========  ===========================================================================
 
 
 ----
@@ -510,19 +514,19 @@ ReduxPar Keywords
 
 Class Instantiation: :class:`pypeit.par.pypeitpar.ReduxPar`
 
-======================  ==========  =====================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================  ====================================  ======================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================
-Key                     Type        Options                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Default                               Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
-======================  ==========  =====================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================  ====================================  ======================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================
[…]
