+++ doc/step-by-step.rst
@@ -1,0 +1,291 @@
+====================
+Step by Step Example
+====================
+
+Overview
+========
+
+This doc goes through a full run of *PypeIt* on one of the
+Shane Kastb datasets in the Development Suite.
+
+Setup
+=====
+
+Organize data
+-------------
+
+Place all of the files in a single folder.
+Mine is named /home/xavier/Projects/PypeIt-development-suite/RAW_DATA/shane_kast_blue/600_4310_d55
+(which I will refer to as RAW_PATH)
+and the files within are:
+
+b10.fits.gz  b15.fits.gz  b1.fits.gz   b24.fits.gz  b4.fits.gz  b9.fits.gz
+b11.fits.gz  b16.fits.gz  b20.fits.gz  b27.fits.gz  b5.fits.gz
+b12.fits.gz  b17.fits.gz  b21.fits.gz  b28.fits.gz  b6.fits.gz
+b13.fits.gz  b18.fits.gz  b22.fits.gz  b2.fits.gz   b7.fits.gz
+b14.fits.gz  b19.fits.gz  b23.fits.gz  b3.fits.gz   b8.fits.gz
+
+Run pypeit_setup
+----------------
+
+The first script you will run with PypeIt is :ref:`pypeit_setup` which
+examines your raw files and generates a sorted list and (when instructed)
+one :doc:`pypeit_file` per instrument configuration.
+
+Complete instructions are provided in :doc:`setup`.
+
+Here is my call for these data::
+
+    cd folder_for_reducing   # *not* the raw data folder
+    pypeit_setup -r RAW_PATH/b -s shane_kast_blue -c A
+
+This creates a :doc:`pypeit_file` in the folder named
+*shane_kast_blue_A* beneath where the script was run.
+
+It looks like this::
+
+    # Auto-generated PypeIt file
+    # Wed 12 Aug 2020 10:00:54
+
+    # User-defined execution parameters
+    [rdx]
+    spectrograph = shane_kast_blue
+
+    # Setup
+    setup read
+     Setup A:
+       --:
+         binning: 1,1
+         dichroic: d55
+         disperser:
+           angle: none
+           name: 600/4310
+         slit:
+           decker: 2.0 arcsec
+           slitlen: none
+           slitwid: none
+    setup end
+
+    # Read in the data
+    data read
+     path /home/xavier/Projects/PypeIt-development-suite/RAW_DATA/shane_kast_blue/600_4310_d55
+    |    filename |                 frametype |                 ra |                dec |     target | dispname |     decker | binning |                mjd |        airmass | exptime | dichroic |
+    |  b1.fits.gz |                  arc,tilt | 140.44166666666663 |  37.43222222222222 |       Arcs | 600/4310 | 0.5 arcsec |     1,1 |  57162.06664467593 |            1.0 |    30.0 |      d55 |
+    | b14.fits.gz |                      bias | 172.34291666666664 |  36.86833333333333 |       Bias | 600/4310 | 2.0 arcsec |     1,1 |  57162.15420034722 |            1.0 |     0.0 |      d55 |
+    | b15.fits.gz |                      bias | 172.41833333333332 |  36.94444444444444 |       Bias | 600/4310 | 2.0 arcsec |     1,1 |  57162.15440162037 |            1.0 |     0.0 |      d55 |
+    | b16.fits.gz |                      bias | 172.49124999999995 |  36.97833333333333 |       Bias | 600/4310 | 2.0 arcsec |     1,1 |    57162.154603125 |            1.0 |     0.0 |      d55 |
[…]
