### 2020-08-12 pypeit/PypeIt#994 @profxj
Thanks for all of this @astrom-tom !

To address your comments/suggestions in this Issue, we have done the following:

- Updated the `joss-paper-branch` to the bleeding edge (this contains new docs, among other things)
- Entirely revamped the front page:  https://github.com/pypeit/PypeIt/blob/joss-paper-branch/doc/index.rst
- Reduced the doctree to a set of *user* focused items.  Unfortunately you can't see them on GitHub and we don't know how to push anything but `master` to readthedocs.  So I attach below a screen shot of what I see when building locally
- Added text to the Installation page to (hopefully) clarify your query
- Added a new usage page: https://github.com/pypeit/PypeIt/blob/joss-paper-branch/doc/step-by-step.rst

Hopefully the last bit is enough to get you commenting on software functionality.
And, of course, additional comments are welcome on the other items you raise.

Thanks!

![doctree](https://user-images.githubusercontent.com/3998093/90052910-a7637000-dc8e-11ea-89b6-835836696b6c.png)
