### 2020-08-12 pypeit/PypeIt#995 @profxj
The notes on testing have advanced considerably.  Please review the latest:
https://github.com/pypeit/PypeIt/blob/joss-paper-branch/doc/installing.rst

where you will see the test suite is only for Developers.
We prefer not to provide additional information at this stage,
in part because it is an evolving system.

If you think it is best (i.e. less confusing), we could remove those notes altogether.
