+++ doc/spectrographs.rst
@@ -23,7 +23,7 @@ keck_lris_red_orig      Keck        :doc:`lris` spectrometer; red camera + origi
 keck_mosfire            Keck        MOSFIRE spectrometer; J and Y gratings tested
 keck_nires              Keck        NIRES spectrometer
 keck_nirspec_low        Keck        NIRSPEC spectrometer; low-dispersion
-keck_deimos             Keck        DEIMOS spectrometer (600ZD, 830G, 1200G)
+keck_deimos             Keck        :doc:`deimos` spectrometer (600ZD, 830G, 1200G)
 lbt_luci1               LBT         LUCI-I spectrometer
 lbt_luci2               LBT         LUCI-II spectrometer
 magellan_fire           Magellan    FIRE spectrometer; Echelle mode
@@ -36,9 +36,9 @@ shane_kast_red          Lick 3m     Kast dual spectrometer; red camera
 shane_kast_red_ret      Lick 3m     Kast dual spectrometer; red reticon
 tng_dolores             TNG         DOLORES (LRS) spectrograph; LR-R
 vlt_fors2               VLT         FORS2 spectrometer; only a few gratings
-vlt_xshooter_uvb        VLT         :doc:`xshooter`; UVB camera
-vlt_xshooter_vis        VLT         :doc:`xshooter`; VIS camera
-vlt_xshooter_nir        VLT         :doc:`xshooter`; NIR camera
+vlt_xshooter_uvb        VLT         :doc:`xshooter` spectrometer; UVB camera
+vlt_xshooter_vis        VLT         :doc:`xshooter` spectrometer; VIS camera
+vlt_xshooter_nir        VLT         :doc:`xshooter` spectrometer; NIR camera
 wht_isis_blue           WHT         ISIS spectrometer; blue camera?
 p200_dbsp_blue          P200        DBSP spectrograph; blue camera
 p200_dbsp_red           P200        DBSP spectrograph; red camera
