### 2020-08-12 pypeit/PypeIt#993 @csnap001
I had this issue as well, but found that the package 'extension_helpers' comes with python3. Once I switched my default python distribution to python3 I was able install without issue. Not sure if there would be problems with trying to install this package into a python2 distribution?
