+++ pypeit/datamodel.py
@@ -929,6 +929,9 @@ class DataContainer:
                 #  Are we a DataContainer?
                 if obj_is_data_container(cls.datamodel[e]['otype']):
                     # Parse it!
+                    # TODO: This only works with single extension
+                    # DataContainers. Do we want this to be from_hdu
+                    # instead and add chk_version to _parse?
                     _d[e], p1, p2 = cls.datamodel[e]['otype']._parse(_hdu[hduindx])
                     dm_version_passed &= p1
                     dm_type_passed &= p2
@@ -1041,7 +1044,7 @@ class DataContainer:
             return
         # Check data type
         if not isinstance(value, self.datamodel[item]['otype']):
-            raise TypeError('Incorrect data type for {0}!'.format(item) +
+            raise TypeError('Incorrect data type for {0}!  '.format(item) +
                             'Allowed type(s) are: {0}'.format(self.datamodel[item]['otype']))
         # Array?
         if 'atype' in self.datamodel[item].keys():
@@ -1064,6 +1067,56 @@ class DataContainer:
             :obj:`dict_keys`: The iterable with the data model keys.
         """
         return self.datamodel.keys()
+
+    def _primary_header(self, hdr=None):
+        """
+        Construct a primary header that is included with the primary
+        HDU extension produced by :func:`to_hdu`.
+
+        Additional data can be added to the header for individual HDU
+        extensions in :func:`to_hdu` as desired, but this function
+        **should not** add any elements from the datamodel to the
+        header.
+
+        Args:
+            hdr (`astropy.io.fits.Header`_, optional):
+                Header for the primary extension. If None, set by
+                :func:`pypeit.io.initialize_header()`.
+
+        Returns:
+            `astropy.io.fits.Header`_: Header object to include in
+            the primary HDU.
+        """
+        return io.initialize_header() if hdr is None else hdr.copy()
+
+    def _base_header(self, hdr=None):
+        """
+        Construct a base header that is included with all HDU
+        extensions produced by :func:`to_hdu` (unless they are
+        overwritten by a nested :class:`DataContainer`).
+
+        Additional data can be added to the header for individual HDU
+        extensions in :func:`to_hdu` as desired, but this function
+        **should not** add any elements from the datamodel to the
+        header.
+
+        Args:
+            hdr (`astropy.io.fits.Header`_, optional):
+                Baseline header to add to all returned HDUs. If None,
+                set by :func:`pypeit.io.initialize_header()`.
+
+        Returns:
+            `astropy.io.fits.Header`_: Header object to include in
+            all HDU extensions.
+        """
+        # Copy primary header to all subsequent headers
+        _hdr = self._primary_header(hdr=hdr)
+        # Add DataContainer class name and datamodel version number.
+        # This is not added to the primary header because single output
+        # files can contain multiple DataContainer objects.
+        _hdr['DMODCLS'] = (self.__class__.__name__, 'Datamodel class')
+        _hdr['DMODVER'] = (self.version, 'Datamodel version')
+        return _hdr
 
     # TODO: Always have this return an HDUList instead of either that
     # or a normal list?
@@ -1106,7 +1159,8 @@ class DataContainer:
             limit_hdus (list, optional):
                 Limit the HDUs that can be written to the items in this list
             force_to_bintbl (bool, optional):
-                Force any dict into a BinTableHDU (e.g. for SpecObj)
+                Force any dict into a BinTableHDU (e.g. for
+                :class:`pypeit.specobj.SpecObj`)
 
         Returns:
             :obj:`list`, `astropy.io.fits.HDUList`_: A list of HDUs,
[…]

[…]
