### 2020-08-13 #2308 @Gabriel-p
@astrom-tom I see you marked **Community guidelines** as checked. Where did you find this in the docs? I was not able to find such requirements.
