### 2020-08-13 pypeit/PypeIt#995 @astrom-tom
Hi @profxj ,

Thanks for the update!

If the [Run the PypeIt unit tests] is enough to check that the functionalities are working correctly then I would remove the [Try running PypeIt on our extensive development test suite] section. Maybe just mention it in once sentence like 'A test suite for development purposes is available at [link]' at the end of [Run the PypeIt unit tests].
