### 2020-08-13 pypeit/PypeIt#993 @astrom-tom
My main point is that the package dependencies should be in the setup.py so there are installed when you install pypeit. For the moment  you are using the find_packages from setuptools. It seems it does not find some of the packages. Since your list of dependencies seems to be well defined, I would just write it down directly in the setup.py.
