+++ pypeit/flatfield.py
@@ -116,76 +116,55 @@ class FlatImages(datamodel.DataContainer):
             above.
         """
         d = []
-
-        # Rest of the datamodel
         for key in self.keys():
             # Skip None
             if self[key] is None:
                 continue
-            # Array?
-            if self.datamodel[key]['otype'] == np.ndarray and \
-                    key != 'pixelflat_spat_bsplines' and \
-                    key != 'illumflat_spat_bsplines':
-                d.append({key: self[key]})
-            elif key == 'pixelflat_spat_bsplines':
-                for ss, spat_bspl in enumerate(self[key]):
-                    # Naming
-                    spat_bspl.hdu_prefix = 'PIXELFLAT_SPAT_ID-{}_'.format(self.spat_id[ss])
-                    dkey = 'bspline-{}'.format(self.spat_id[ss])
-                    # Save
-                    d.append(copy.deepcopy({dkey: spat_bspl}))
-            elif key == 'illumflat_spat_bsplines':
-                for ss, spat_bspl in enumerate(self[key]):
-                    # Naming
-                    spat_bspl.hdu_prefix = 'ILLUMFLAT_SPAT_ID-{}_'.format(self.spat_id[ss])
-                    dkey = 'bspline-{}'.format(self.spat_id[ss])
-                    # Save
-                    d.append(copy.deepcopy({dkey: spat_bspl}))
-            else: # Add to header of the primary image
-                d[0][key] = self[key]
+            if self.datamodel[key]['otype'] == np.ndarray and 'bsplines' not in key:
+                d += [{key: self[key]}]
+            elif 'bsplines' in key:
+                flattype = 'pixelflat' if 'pixelflat' in key else 'illumflat'
+                d += [{'{0}_spat_id-{1}_bspline'.format(flattype, self.spat_id[ss]): self[key][ss]}
+                            for ss in range(len(self[key]))]
+            else:
+                if len(d) > 0:
+                    d[0][key] = self[key]
+                else:
+                    d += [{key: self[key]}]
         # Return
         return d
 
     @classmethod
-    def _parse(cls, hdu, ext=None, transpose_table_arrays=False, debug=False, hdu_prefix=None):
-        # Grab everything but the bspline's
-        _d, dm_version_passed, dm_type_passed = super(FlatImages, cls)._parse(hdu)
-        # Now the bsplines
-        list_of_pixbsplines, list_of_illbsplines = [], []
-        pixspat_ids, illspat_ids = [], []
-        has_pixel, has_illum = False, False   # Boolean flags to test if the input file has pixelflat and/or illumflat
-        for ihdu in hdu:
-            if 'PIXELFLAT' in ihdu.name and 'BSPLINE' in ihdu.name:
-                has_pixel = True
-                ibspl = bspline.bspline.from_hdu(ihdu)
-                if ibspl.version != bspline.bspline.version:
-                    msgs.warn("Your bspline is out of date!!")
-                list_of_pixbsplines.append(ibspl)
-                # Grab SPAT_ID for checking
-                i0 = ihdu.name.find('ID-')
-                i1 = ihdu.name.find('_BSP')
-                pixspat_ids.append(int(ihdu.name[i0+3:i1]))
-            if 'ILLUMFLAT' in ihdu.name and 'BSPLINE' in ihdu.name:
-                has_illum = True
-                ibspl = bspline.bspline.from_hdu(ihdu)
-                if ibspl.version != bspline.bspline.version:
-                    msgs.warn("Your bspline is out of date!!")
-                list_of_illbsplines.append(ibspl)
-                # Grab SPAT_ID for checking
-                i0 = ihdu.name.find('ID-')
-                i1 = ihdu.name.find('_BSP')
[…]
