+++ pypeit/edgetrace.py
@@ -447,7 +447,7 @@ class EdgeTraceSet(DataContainer):
 
         # Check input types
         if not isinstance(traceimg, TraceImage):
-            msgs.error('Input trace_mg must be a TraceImage object.')
+            msgs.error('Input traceimg must be a TraceImage object.')
         if not isinstance(spectrograph, Spectrograph):
             msgs.error('Input spectrograph must be a Spectrograph object.')
         if not isinstance(par, EdgeTracePar):
@@ -520,83 +520,84 @@ class EdgeTraceSet(DataContainer):
     def nslits(self):
         if self.is_synced:
             return self.ntrace//2
-        # TODO: Maybe this should only throw a warning
         msgs.error('Number of slits undefined because edges are not left-right synchronized.')
 
-#    @staticmethod
-#    def empty_design_table(rows=None):
-#        """
-#        Construct an empty `design` table.
-#
-#        Args:
-#            rows (:obj:`int`, optional):
-#                Number of table rows for each table column, expected
-#                to be the number of matched slits. If None, the table
-#                has empty columns.
-#
-#        Returns:
-#            `astropy.table.Table`_: Instance of the empty design
-#            table.
-#        """
-#        length = 0 if rows is None else rows
-#        return table.Table([
-#                    table.Column(name='TRACEID', dtype=int, length=length,
-#                                 description='Trace ID Number'),
-#                    table.Column(name='TRACESROW', dtype=int, length=length,
-#                                 description='Spectral row for provided left and right edges.'),
-#                    table.Column(name='TRACELPIX', dtype=float, length=length,
-#                                 description='Spatial pixel coordinate for left edge'),
-#                    table.Column(name='TRACERPIX', dtype=float, length=length,
-#                                 description='Spatial pixel coordinate for right edge'),
-#                    table.Column(name='SLITID', dtype=int, length=length,
-#                                 description='Slit ID Number'),
-#                    table.Column(name='SLITLFOC', dtype=float, length=length,
-#                                 description='Left edge of the slit in mm at the focal plane'),
-#                    table.Column(name='SLITRFOC', dtype=float, length=length,
-#                                 description='Right edge of the slit in mm at the focal plane'),
-#                    table.Column(name='SLITRA', dtype=float, length=length,
-#                                 description='Right ascension of the slit center (deg)'),
-#                    table.Column(name='SLITDEC', dtype=float, length=length,
-#                                 description='Declination of the slit center (deg)'),
-#                    table.Column(name='SLITLEN', dtype=float, length=length,
-#                                 description='Slit length (arcsec)'),
-#                    table.Column(name='SLITWID', dtype=float, length=length,
-#                                 description='Slit width (arcsec)'),
-#                    table.Column(name='SLITPA', dtype=float, length=length,
-#                                 description='Slit position angle onsky (deg from N through E)'),
-#                    table.Column(name='ALIGN', dtype=np.int16, length=length,
-#                                 description='Slit used for alignment (1-yes; 0-no), not target '
-#                                             'observations.')
-#                           ])
-#
-#    @staticmethod
-#    def empty_objects_table(rows=None):
-#        """
-#        Construct an empty `objects` table.
-#
-#        Args:
-#            rows (:obj:`int`, optional):
-#                Number of table rows for each table column, expected
-#                to be the number of objects. If None, the table has
-#                empty columns.
-#
-#        Returns:
-#            `astropy.table.Table`_: Instance of the empty object
-#            table.
-#        """
[…]
