### 2020-08-13 pypeit/PypeIt#995 @profxj
Good call @astrom-tom . Here it is updated:

https://github.com/pypeit/PypeIt/blob/joss-paper-branch/doc/installing.rst

The unit tests do test enough for a user to go forth.
