### 2020-08-14 e247310
Fixing bugs in coadd2d.
