+++ pypeit/bspline/__init__.py
@@ -1,3 +1,4 @@
 
 from pypeit.bspline.bspline import bspline
+from pypeit.bspline.bspline import bspline_profile
 
