+++ pypeit/core/skysub.py
@@ -3,9 +3,6 @@
 .. include common links, assuming primary doc root is up one directory
 .. include:: ../links.rst
 """
-import sys
-import os
-
 import numpy as np
 
 from scipy import ndimage
@@ -20,7 +17,6 @@ from pypeit.core import basis, pixels, extract
 from pypeit.core import fitting
 from pypeit import msgs, utils, bspline, slittrace
 from pypeit.display import display
-from pypeit.core.moment import moment1d
 
 def skysub_npoly(thismask):
     """
@@ -142,7 +138,7 @@ def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask=Non
             #lsky_ivar = np.full(lsky.shape, 0.1)
             # Init bspline to get the sky breakpoints (kludgy)
             lskyset, outmask, lsky_fit, red_chi, exit_status \
-                    = fitting.bspline_profile(pix[pos_sky], lsky, lsky_ivar, np.ones_like(lsky),
+                    = bspline.bspline_profile(pix[pos_sky], lsky, lsky_ivar, np.ones_like(lsky),
                                             ingpm=inmask_fit[pos_sky], upper=sigrej, lower=sigrej,
                                             kwargs_bspline={'bkspace':bsp},
                                             kwargs_reject={'groupbadpix': True, 'maxrej': 10})
@@ -162,7 +158,7 @@ def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask=Non
     # Perform the full fit now
     msgs.info("Full fit in global sky sub.")
     skyset, outmask, yfit, _, exit_status \
-            = fitting.bspline_profile(pix, sky, sky_ivar, poly_basis, ingpm=inmask_fit, nord=4,
+            = bspline.bspline_profile(pix, sky, sky_ivar, poly_basis, ingpm=inmask_fit, nord=4,
                                     upper=sigrej, lower=sigrej, maxiter=maxiter,
                                     kwargs_bspline={'bkspace':bsp},
                                     kwargs_reject={'groupbadpix':True, 'maxrej': 10})
@@ -177,7 +173,7 @@ def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask=Non
         poly_basis = np.ones_like(sky)
         # Perform the full fit now
         skyset, outmask, yfit, _, exit_status \
-                = fitting.bspline_profile(pix, sky, sky_ivar, poly_basis, ingpm=inmask_fit, nord=4,
+                = bspline.bspline_profile(pix, sky, sky_ivar, poly_basis, ingpm=inmask_fit, nord=4,
                                         upper=sigrej, lower=sigrej, maxiter=maxiter,
                                         kwargs_bspline={'bkspace': bsp},
                                         kwargs_reject={'groupbadpix': False, 'maxrej': 10})
@@ -273,7 +269,7 @@ def skyoptimal(wave, data, ivar, oprof, sortpix, sigrej=3.0, npoly=1, spatial=No
 
     if ngood > 0:
         sset1, outmask_good1, yfit1, red_chi1, exit_status \
-                = fitting.bspline_profile(wave[good], data[good], ivar[good], profile_basis[good, :],
+                = bspline.bspline_profile(wave[good], data[good], ivar[good], profile_basis[good, :],
                                         fullbkpt=fullbkpt, upper=sigrej, lower=sigrej,
                                         relative=relative,
                                         kwargs_reject={'groupbadpix': True, 'maxrej': 5})
@@ -292,7 +288,7 @@ def skyoptimal(wave, data, ivar, oprof, sortpix, sigrej=3.0, npoly=1, spatial=No
     msgs.info('Iter     Chi^2     Rejected Pts')
     if np.any(mask1):
[…]
