+++ pypeit/bspline/bspline.py
@@ -14,8 +14,10 @@ import warnings
 from IPython import embed
 
 import numpy as np
+from matplotlib import pyplot as plt
 
 from pypeit.core import basis
+from pypeit.core import pydl
 from pypeit import datamodel
 from pypeit import msgs
 
@@ -715,7 +717,7 @@ def uniq(x, index=None):
     Examples
     --------
     >>> import numpy as np
-    >>> from pydl import uniq
+    >>> from pypeit.core.pydl import uniq
     >>> data = np.array([ 1, 2, 3, 1, 5, 6, 1, 7, 3, 2, 5, 9, 11, 1 ])
     >>> print(uniq(np.sort(data)))
     [ 3  5  7  9 10 11 12 13]
@@ -726,4 +728,452 @@ def uniq(x, index=None):
         return np.flatnonzero(np.concatenate(([True], x[1:] != x[:-1], [True])))[1:]-1
     _x = x[index]
     return np.flatnonzero(np.concatenate(([True], _x[1:] != _x[:-1], [True])))[1:]-1
+
+
+def iterfit(xdata, ydata, invvar=None, inmask=None, upper=5, lower=5, x2=None,
+            maxiter=10, nord=4, bkpt=None, fullbkpt=None, kwargs_bspline={}, kwargs_reject={}):
+        """Iteratively fit a b-spline set to data, with rejection. This is a utility function that allows
+        the bspline to use via a direct function call.
+
+        Parameters
+        ----------
+        xdata : :class:`numpy.ndarray`
+            Independent variable.
+        ydata : :class:`numpy.ndarray`
+            Dependent variable.
+        invvar : :class:`numpy.ndarray`
+            Inverse variance of `ydata`.  If not set, it will be calculated based
+            on the standard deviation.
+        upper : :class:`int` or :class:`float`
+            Upper rejection threshold in units of sigma, defaults to 5 sigma.
+        lower : :class:`int` or :class:`float`
+            Lower rejection threshold in units of sigma, defaults to 5 sigma.
+        x2 : :class:`numpy.ndarray`, optional
+            Orthogonal dependent variable for 2d fits.
+        maxiter : :class:`int`, optional
+            Maximum number of rejection iterations, default 10.  Set this to
+            zero to disable rejection.
+
+        Returns
+        -------
+        :func:`tuple`
+            A tuple containing the fitted bspline object and an output mask.
+        """
+        # from .math import djs_reject
+        nx = xdata.size
+        if ydata.size != nx:
+            raise ValueError('Dimensions of xdata and ydata do not agree.')
+        if invvar is not None:
+            if invvar.size != nx:
+                raise ValueError('Dimensions of xdata and invvar do not agree.')
+        else:
+            #
+            # This correction to the variance makes it the same
+            # as IDL's variance()
+            #
+            var = ydata.var() * (float(nx) / float(nx - 1))
+            if var == 0:
+                var = 1.0
+            invvar = np.ones(ydata.shape, dtype=ydata.dtype) / var
+
+        if inmask is None:
+            inmask = invvar > 0.0
+
+        if x2 is not None:
+            if x2.size != nx:
+                raise ValueError('Dimensions of xdata and x2 do not agree.')
+        yfit = np.zeros(ydata.shape)
+        if invvar.size == 1:
+            outmask = True
+        else:
+            outmask = np.ones(invvar.shape, dtype='bool')
+        xsort = xdata.argsort()
+        maskwork = (outmask & inmask & (invvar > 0.0))[xsort]
+        if 'oldset' in kwargs_bspline:
+            sset = kwargs_bspline['oldset']
+            sset.mask = True
+            sset.coeff = 0
+        else:
+            if not maskwork.any():
+                raise ValueError('No valid data points.')
[…]

[…]
