+++ pypeit/bspline/__init__.py
@@ -1,4 +1,5 @@
 
 from pypeit.bspline.bspline import bspline
 from pypeit.bspline.bspline import bspline_profile
+from pypeit.bspline.bspline import iterfit
 
