+++ doc/scripts/build_specobj_rst.py
@@ -9,7 +9,7 @@ import time
 import numpy
 
 from pkg_resources import resource_filename
-from pypeit.par.parset import ParSet
+from pypeit.utils import to_string, string_table
 from pypeit import specobj
 from pypeit import spec2dobj
 from pypeit import coadd1d
@@ -24,13 +24,15 @@ def link_string(p):
 if __name__ == '__main__':
     t = time.perf_counter()
 
-    # Read the baseline file that is not changed and must be edited by
-    # the person building the documentation as necessary.
-    pypeit_root = os.path.dirname(resource_filename('pypeit', ''))
+    # Set the output directory
+    output_root = os.path.join(os.path.split(os.path.abspath(resource_filename('pypeit', '')))[0],
+                               'doc', 'include')
 
-    for obj, basefile in zip([specobj.SpecObj, spec2dobj.Spec2DObj, coadd1d.OneSpec],
-                             ['datamodel_specobj.rst', 'datamodel_spec2dobj.rst',
-                              'datamodel_onespec.rst']):
+    # Iterate through all the specobj classes
+    for obj in [specobj.SpecObj, spec2dobj.Spec2DObj, coadd1d.OneSpec]:
+
+        ofile = os.path.join(output_root, 'datamodel_{0}.rst'.format(obj.__name__.lower()))
+
         lines = []
         lines += ['']
 
@@ -39,40 +41,37 @@ if __name__ == '__main__':
         lines += ['Version: {:s}'.format(obj.version)]
         lines += ['']
 
-        data_model = obj.datamodel
-
-        keys = list(data_model.keys())
+        keys = list(obj.datamodel.keys())
         keys.sort()
 
-        data_table = numpy.empty((len(data_model)+1, 4), dtype=object)
-        data_table[0,:] = ['Obj', 'Obj Type', 'Array Type', 'Description']
+        data_table = numpy.empty((len(obj.datamodel)+1, 4), dtype=object)
+        data_table[0,:] = ['Obj Key', 'Obj Type', 'Array Type', 'Description']
         for i,k in enumerate(keys):
             # Key
-            data_table[i+1,0] = ParSet._data_string(k, use_repr=False, verbatum=True)
+            data_table[i+1,0] = to_string(k, use_repr=False, verbatim=True)
             # Object Type
-            if isinstance(data_model[k]['otype'], (list,tuple)):
-                data_table[i+1,1] = ', '.join([t.__name__ for t in data_model[k]['otype']])
+            if isinstance(obj.datamodel[k]['otype'], (list,tuple)):
+                data_table[i+1,1] = ', '.join([t.__name__ for t in obj.datamodel[k]['otype']])
             else:
-                data_table[i+1,1] = data_model[k]['otype'].__name__
+                data_table[i+1,1] = obj.datamodel[k]['otype'].__name__
             # Array type
-            if 'atype' in data_model[k].keys():
-                data_table[i+1,2] = data_model[k]['atype'].__name__
+            if 'atype' in obj.datamodel[k].keys():
+                data_table[i+1,2] = obj.datamodel[k]['atype'].__name__
             else:
                 data_table[i+1,2] = ' '
             if data_table[i+1,2][-1] == '_':
                 data_table[i+1,2] = data_table[i+1,2][:-1]
 
             # Description
-            data_table[i+1,3] = ParSet._data_string(data_model[k]['descr'])
+            data_table[i+1,3] = to_string(obj.datamodel[k]['descr'])
 
-        lines += [ParSet._data_table_string(data_table, delimeter='rst')]
+        lines += [string_table(data_table, delimeter='rst')]
 
         # Finish
[…]
