### 2020-08-14 d3940c9
updates to the deimos optical model
