+++ pypeit/spectrographs/opticalmodel.py
@@ -3,6 +3,9 @@ Module to generate an optical model for a spectrograph.
 """
 import warnings
 import numpy
+import scipy
+import pdb
+
 
 # ----------------------------------------------------------------------
 # General class for a reflection grating
@@ -11,12 +14,12 @@ class ReflectionGrating:
     Doc this!
     """
     def __init__(self, ruling, tilt, roll, yaw, central_wave=None):
-        self.ruling = ruling            # Grating groove spacing in lines per mm
-        self.tilt = tilt                # Effective grating tilt (deg)
-        self.roll = roll                # Grating roll (deg)
-        self.yaw = yaw                  # Grating yaw (deg)
+        self.ruling = ruling  # Grating groove spacing in lines per mm
+        self.tilt = tilt  # Effective grating tilt (deg)
+        self.roll = roll  # Grating roll (deg)
+        self.yaw = yaw  # Grating yaw (deg)
 
-        self.central_wave = central_wave    # Central wavelength (angstroms)
+        self.central_wave = central_wave  # Central wavelength (angstroms)
 
         self.transform = self._get_grating_transform()
 
@@ -54,7 +57,7 @@ class ReflectionGrating:
                             [     -sinx,                 cost*cosx,                 sint*cosx],
                             [ cosx*sinr, -sint*cosr+cost*sinx*sinr,  cost*cosr+sint*sinx*sinr]])
 
-    def reflect(self, r, wave=None, order=1):
+    def reflect(self, r, nslits, wave=None, order=1):
         """
         Propagate an input ray for a given wavelength and order.
 
@@ -75,6 +78,8 @@ class ReflectionGrating:
             raise NotImplementedError('Input wavelength must be one number or a vector.')
         nwave = _wave.size
 
+        _wave_arr = numpy.tile(_wave, (nslits, 1))
+
         _r = numpy.atleast_2d(r)
         if _r.ndim > 2:
             raise NotImplementedError('Rays must be 1D for a single ray, or 2D for multiple.')
@@ -83,24 +88,23 @@ class ReflectionGrating:
         _r = OpticalModel.conjugate_surface_transform(_r, self.transform, forward=True)
 
         # Get the grating input angles
-        alpha = -numpy.arctan2(-_r[:,1],-_r[:,2])
-        gamma = numpy.arctan2(_r[:,0],numpy.sqrt(numpy.square(_r[:,1])+numpy.square(_r[:,2])))
+        alpha = -numpy.arctan2(-_r[:, 1], -_r[:, 2])
+        gamma = numpy.arctan2(_r[:, 0], numpy.sqrt(numpy.square(_r[:, 1]) + numpy.square(_r[:, 2])))
 
         # Use the grating equation to get the output angle (minus sign
         # in front of sin(alpha) is for reflection)
-        beta = numpy.arcsin((order * 1e-7 * self.ruling * _wave[None,:] / numpy.cos(gamma)[:,None])
-                            - numpy.sin(alpha)[:,None])
+        beta = (numpy.arcsin((order * 1e-7 * self.ruling * _wave_arr.ravel() / numpy.cos(gamma))
+                             - numpy.sin(alpha))).reshape(_wave_arr.shape).T
 
         # Revert to ray vectors
-        wavesign = 1-2*(_wave < 0)
-        cosg = numpy.cos(gamma)
-        _r = numpy.array([numpy.repeat(numpy.sin(gamma), nwave).reshape(-1,nwave),
-                          numpy.sin(-beta*wavesign[None,:])*cosg[:,None],
-                          numpy.cos(-beta*wavesign[None,:])*cosg[:,None] ]).T
+        wavesign = 1 - 2 * (_wave_arr.T < 0)
+        _r = numpy.array([numpy.sin(gamma),
[…]
