+++ pypeit/spectrographs/keck_deimos.py
@@ -8,8 +8,11 @@ import numpy as np
 import warnings
 
 from scipy import interpolate
+from scipy.io import readsav
 
 from astropy.io import fits
+
+from pkg_resources import resource_filename
 
 from pypeit import msgs
 from pypeit import telescopes
@@ -703,23 +706,63 @@ class KeckDEIMOSSpectrograph(spectrograph.Spectrograph):
             raise ValueError('Ruling should be 0 if slider in position 2.')
 
         # Use the calibrated coefficients
+        # These orientation coefficients are the newest ones and are meant for
+        # observations obtained Post-2016 Servicing.
+        # TODO: Figure out the impact of these coefficients on the slits identification.
+        # We may not need to change them according to when the observations were taken
         _ruling = int(ruling) if int(ruling) in [600, 831, 900, 1200] else 'other'
-        orientation_coeffs = {3: {    600: [ 0.145, -0.008, 5.6e-4, -0.182],
-                                      831: [ 0.143,  0.000, 5.6e-4, -0.182],
-                                      900: [ 0.141,  0.000, 5.6e-4, -0.134],
-                                     1200: [ 0.145,  0.055, 5.6e-4, -0.181],
-                                  'other': [ 0.145,  0.000, 5.6e-4, -0.182] },
-                              4: {    600: [-0.065,  0.063, 6.9e-4, -0.298],
-                                      831: [-0.034,  0.060, 6.9e-4, -0.196],
-                                      900: [-0.064,  0.083, 6.9e-4, -0.277],
-                                     1200: [-0.052,  0.122, 6.9e-4, -0.294],
-                                  'other': [-0.050,  0.080, 6.9e-4, -0.250] } }
+        orientation_coeffs = {3: {    600: [ 0.145, -0.008, 5.6e-4, -0.146],
+                                      831: [ 0.143,  0.000, 5.6e-4, -0.018],
+                                      900: [ 0.141,  0.000, 5.6e-4, -0.118],
+                                     1200: [ 0.145,  0.055, 5.6e-4, -0.141],
+                                  'other': [ 0.145,  0.000, 5.6e-4, -0.141] },
+                              4: {    600: [-0.065,  0.063, 6.9e-4, -0.108],
+                                      831: [-0.034,  0.060, 6.9e-4, -0.038],
+                                      900: [-0.064,  0.083, 6.9e-4, -0.060],
+                                     1200: [-0.052,  0.122, 6.9e-4, -0.110],
+                                  'other': [-0.050,  0.080, 6.9e-4, -0.110] } }
+
+        # Orientation coefficients meant for observations taken Pre-2016 Servicing
+        # orientation_coeffs = {3: {    600: [ 0.145, -0.008, 5.6e-4, -0.182],
+        #                               831: [ 0.143,  0.000, 5.6e-4, -0.182],
+        #                               900: [ 0.141,  0.000, 5.6e-4, -0.134],
+        #                              1200: [ 0.145,  0.055, 5.6e-4, -0.181],
+        #                           'other': [ 0.145,  0.000, 5.6e-4, -0.182] },
+        #                       4: {    600: [-0.065,  0.063, 6.9e-4, -0.298],
+        #                               831: [-0.034,  0.060, 6.9e-4, -0.196],
+        #                               900: [-0.064,  0.083, 6.9e-4, -0.277],
+        #                              1200: [-0.052,  0.122, 6.9e-4, -0.294],
+        #                           'other': [-0.050,  0.080, 6.9e-4, -0.250] } }
 
         # Return calbirated roll, yaw, and tilt
         return orientation_coeffs[slider][_ruling][0], \
                 orientation_coeffs[slider][_ruling][1], \
                 tilt*(1-orientation_coeffs[slider][_ruling][2]) \
[…]
