+++ pypeit/tests/test_bspline.py
@@ -9,6 +9,7 @@ import numpy as np
 
 from pypeit import bspline
 from pypeit.tests.tstutils import bspline_ext_required, data_path
+from pypeit.core import fitting
 
 @bspline_ext_required
 def test_model_versions():
@@ -150,7 +151,7 @@ def test_profile_spec():
     for f in files:
         d = np.load(f)
         spec_bspl, spec_gpm_fit, spec_flat_fit, _, exit_status \
-                = bspline.bspline_profile(d['spec_coo_data'], d['spec_flat_data'], d['spec_ivar_data'],
+                = fitting.bspline_profile(d['spec_coo_data'], d['spec_flat_data'], d['spec_ivar_data'],
                                   np.ones_like(d['spec_coo_data']), ingpm=d['spec_gpm_data'],
                                   nord=4, upper=logrej, lower=logrej,
                                   kwargs_bspline={'bkspace': spec_samp_fine},
@@ -169,7 +170,7 @@ def test_io():
     for f in files:
         d = np.load(f)
         spec_bspl, spec_gpm_fit, spec_flat_fit, _, exit_status \
-            = bspline.bspline_profile(d['spec_coo_data'], d['spec_flat_data'], d['spec_ivar_data'],
+            = fitting.bspline_profile(d['spec_coo_data'], d['spec_flat_data'], d['spec_ivar_data'],
                               np.ones_like(d['spec_coo_data']), ingpm=d['spec_gpm_data'],
                               nord=4, upper=logrej, lower=logrej,
                               kwargs_bspline={'bkspace': spec_samp_fine},
@@ -212,7 +213,7 @@ def test_profile_spat():
                                     bkspace=np.fmax(1.0/d['median_slit_width']/10.0,
                                                     1.2*np.median(np.diff(d['spat_coo_data']))))
         spat_bspl, spat_gpm_fit, spat_flat_fit, _, exit_status \
-                = bspline.bspline_profile(d['spat_coo_data'], d['spat_flat_data'],
+                = fitting.bspline_profile(d['spat_coo_data'], d['spat_flat_data'],
                                   np.ones_like(d['spat_flat_data']),
                                   np.ones_like(d['spat_flat_data']), nord=4, upper=5.0, lower=5.0,
                                   fullbkpt=spat_bspl.breakpoints, quiet=True)
@@ -231,7 +232,7 @@ def test_profile_twod():
     for f in files:
         d = np.load(f)
         twod_bspl, twod_gpm_fit, twod_flat_fit, _ , exit_status \
-                = bspline.bspline_profile(d['twod_spec_coo_data'], d['twod_flat_data'],
+                = fitting.bspline_profile(d['twod_spec_coo_data'], d['twod_flat_data'],
                                   d['twod_ivar_data'], d['poly_basis'], ingpm=d['twod_gpm_data'],
                                   nord=4, upper=twod_sigrej, lower=twod_sigrej,
                                   kwargs_bspline={'bkspace': spec_samp_coarse},
