+++ pypeit/bspline/__init__.py
@@ -1,5 +1,3 @@
 
 from pypeit.bspline.bspline import bspline
-from pypeit.bspline.bspline import bspline_profile
-from pypeit.bspline.bspline import iterfit
 
