### 2020-08-14 69f45f8
Fix docstring on spec_flex_shift
