+++ pypeit/tracepca.py
@@ -17,6 +17,7 @@ from pypeit import utils
 from pypeit.io import hdu_iter_by_ext
 from pypeit.core import trace
 from pypeit.core import pca
+from pypeit.core.fitting import PypeItFit
 from pypeit.datamodel import DataContainer
 
 # TODO: This is even more general than a "trace" PCA. Rename to
@@ -59,7 +60,7 @@ class TracePCA(DataContainer):
             argument of :func:`pypeit.core.pca.pca_decomposition`.
     """
 
-    version = '1.0.0'
+    version = '1.1.0'
     """Datamodel version."""
 
     datamodel = {'reference_row': dict(otype=(int, np.integer),
@@ -89,43 +90,49 @@ class TracePCA(DataContainer):
                  'pca_mean': dict(otype=np.ndarray, atype=(float,np.floating),
                                   descr=r'The mean offset of the PCA decomposotion for each '
                                         r' spectral pixel. Shape is :math:`(N_{\rm spec},)`.'),
-                 'pca_bpm': dict(otype=np.ndarray, atype=(bool, np.bool_),
-                                 descr='A boolean numpy.ndarray masking data PCA component '
-                                       'coefficients that were rejected during the polynomial '
-                                       'fitting. Shape is the same as ``pca_coeff``.'),
-                 'function': dict(otype=str,
-                                  descr='Function type used to model the PCA coefficients as a '
-                                        'function of spatial position.'),
-                 'fit_coeff': dict(otype=list,
-                                   descr=r'A list of numpy.ndarray objects (or a single '
-                                         r'numpy.ndarray), one per PCA component where the length'
-                                         r'of the 1D array is the number of coefficients fit to '
-                                         r'the PCA-component coefficients. The number of function '
-                                         r'coefficients is typically :math:`N_{\rm coeff} = o+1`, '
-                                         r'where :math:`o` is the order of the fit.'),
-                 'minx': dict(otype=(float,np.floating),
-                              descr='The minimum spatial coordiante used to rescale the trace '
-                                    'coordinates.'),
-                 'maxx': dict(otype=(float,np.floating),
-                              descr='The maximum spatial coordiante used to rescale the trace '
-                                    'coordinates.'),
-                 'lower': dict(otype=(float,np.floating),
-                               descr='Number of standard deviations used for rejecting data '
-                                     '**below** the mean residual. If None, no rejection was '
-                                     'performed. See :func:`utils.robust_polyfit_djs`.'),
-                 'upper': dict(otype=(float,np.floating),
-                               descr='Number of standard deviations used for rejecting data '
-                                     '**above** the mean residual. If None, no rejection was '
-                                     'performed. See :func:`utils.robust_polyfit_djs`.'),
-                 'maxrej': dict(otype=int,
-                                descr='Maximum number of points to reject during fit iterations. '
-                                      'See :func:`utils.robust_polyfit_djs`.'),
-                 'maxiter': dict(otype=int,
-                                 descr='Maximum number of rejection iterations allows; will be 0 '
-                                       ' if no rejectsion were performed.')}
-    """Object datamodel."""
+                 'pca_coeffs_model': dict(otype=np.ndarray, atype=PypeItFit,
+                                          descr='An array of PypeItFit objects, one per PCA '
+                                                'component, that models the trend of the PCA '
+                                                'component coefficients with the reference '
[…]
