+++ pypeit/datamodel.py
@@ -815,7 +815,7 @@ class DataContainer:
     @classmethod
     def _parse(cls, hdu, ext=None, transpose_table_arrays=False, hdu_prefix=None):
         """
-        Parse data read from a set of HDUs.
+        Parse data read from one or more HDUs.
 
         This method is the counter-part to :func:`_bundle`, and
         parses data from the HDUs into a dictionary that can be used
@@ -836,7 +836,7 @@ class DataContainer:
 
             - Because the `astropy.table.Table`_ methods are used
               directly, any metadata associated with the Table will
-              also be included in the HDUs construced by
+              also be included in the HDUs constructed by
               :func:`to_hdu`. However, the
               `astropy.table.Table.read`_ method always returns the
               metadata with capitalized keys. This means that,
@@ -868,8 +868,10 @@ class DataContainer:
                 tables. This is meant to invert the use of
                 ``transpose_arrays`` in :func:`_bound`.
             hdu_prefix (:obj:`str`, optional):
-                Decorate the HDUs with this prefix
-                Note, this over-rides any internally set hdu_prefix value
+                Only parse HDUs with extension names matched to this
+                prefix. If None, :attr:`hdu_prefix` is used. If the
+                latter is also None, all HDUs are parsed. See
+                :func:`pypeit.io.hdu_iter_by_ext`.
 
         Returns:
             :obj:`tuple`: Return three objects
@@ -887,7 +889,7 @@ class DataContainer:
         dm_type_passed = True
 
         # Setup to iterate through the provided HDUs
-        _ext, _hdu = io.hdu_iter_by_ext(hdu, ext=ext)
+        _ext, _hdu = io.hdu_iter_by_ext(hdu, ext=ext, hdu_prefix=hdu_prefix)
         _ext = np.atleast_1d(np.array(_ext, dtype=object))
         str_ext = np.logical_not([isinstance(e, (int, np.integer)) for e in _ext])
 
@@ -915,6 +917,9 @@ class DataContainer:
         else:
             prefix = '' if cls.hdu_prefix is None else cls.hdu_prefix
 
+        # Save the list of hdus that have been parsed
+        parsed_hdus = []
+
         # HDUs can have dictionary elements directly.
         keys = np.array(list(_d.keys()))
         prefkeys = np.array([prefix+key.upper() for key in keys])
@@ -923,17 +928,18 @@ class DataContainer:
             found_data = True
             for e in keys[indx]:
                 hduindx = prefix+e.upper()
-                # Check version (and type)?
-                #  Are we a DataContainer?
+                # Add it to the list of parsed HDUs
+                parsed_hdus += [hduindx]
                 if obj_is_data_container(cls.datamodel[e]['otype']):
-                    # Parse it!
+                    # Parse the DataContainer
                     # TODO: This only works with single extension
                     # DataContainers. Do we want this to be from_hdu
                     # instead and add chk_version to _parse?
-                    _d[e], p1, p2 = cls.datamodel[e]['otype']._parse(_hdu[hduindx])
+                    _d[e], p1, p2, _ = cls.datamodel[e]['otype']._parse(_hdu[hduindx])
                     dm_version_passed &= p1
                     dm_type_passed &= p2
                 else:
+                    # Parse the Image or Table data
                     dm_type_passed &= hdu[hduindx].header['DMODCLS'] == cls.__name__
                     dm_version_passed &= hdu[hduindx].header['DMODVER'] == cls.version
                     # Grab it
@@ -949,6 +955,7 @@ class DataContainer:
[…]

[…]
