### 2020-08-15 f135233
doc fixes
