### 2020-08-15 9e64921
minor PR changes
