### 2020-08-16 96ad37d
Fixed bug associated with nans in profile for 2d coadds.
