+++ pypeit/core/extract.py
@@ -978,36 +978,6 @@ def fit_profile(image, ivar, waveimg, thismask, spat_img, trace_in, wave, flux, 
     return (profile_model, xnew, fwhmfit, med_sn2)
 
 
-def parse_manual(manual_par):
-    """
-    Parse the rather klunky ManualExtractionPar parameters into more useful items
-
-    Args:
-        manual_par (:class:`pypeit.par.pypeitpar.ManualExtractionPar`):
-            Manual extract parameter object
-
-    Returns:
-        tuple: spats, specs, det, fwhm
-
-    """
-    if isinstance(manual_par['det'], list):
-        spat_spec = manual_par['spat_spec']#.split(',')
-        det = [int(obj) for obj in manual_par['det']] #.split(',')]
-        fwhm = [float(obj) for obj in manual_par['fwhm']]#.split(',')]
-    else:
-        spat_spec = [manual_par['spat_spec']]
-        det = [manual_par['det']]
-        fwhm = [manual_par['fwhm']]
-    # Deal with spat_spec
-    spats, specs = [], []
-    for ispat_spec in spat_spec:
-        ps = ispat_spec.split(':')
-        spats.append(float(ps[0]))
-        specs.append(float(ps[1]))
-    # Return
-    return spats, specs, det, fwhm
-
-
 def parse_hand_dict(hand_extract_dict):
     """ Utility routine for objfind to parse the hand_extract_dict dictionary for hand selected apertures
 
@@ -1936,8 +1906,7 @@ def ech_objfind(image, ivar, slitmask, slit_left, slit_righ, order_vec, maskslit
     gdorders = np.arange(norders)[np.invert(maskslits)]
     for iord in gdorders: #range(norders):
         msgs.info('Finding objects on order # {:d}'.format(order_vec[iord]))
-        thisslit_gpm = slitmask == gdslit_spat[iord]
-        inmask_iord = inmask & thisslit_gpm
+        inmask_iord = inmask & (slitmask == gdslit_spat[iord])
         specobj_dict['SLITID'] = iord
         specobj_dict['ECH_ORDERINDX'] = iord
         specobj_dict['ECH_ORDER'] = order_vec[iord]
