### 2020-08-16 0d88449
painful fussing for naming of WaveFit HDUs
