+++ pypeit/core/wavecal/wv_fitting.py
@@ -7,6 +7,8 @@
 import numpy as np
 import inspect
 
+from astropy.io import fits
+
 from pypeit.core.wavecal import autoid
 from pypeit.core.wavecal import defs
 from pypeit.core import fitting
@@ -27,7 +29,8 @@ class WaveFit(datamodel.DataContainer):
     """
     version = '1.0.0'
 
-    datamodel = {'pypeitfit': dict(otype=fitting.PypeItFit,
+    datamodel = {'spat_id': dict(otype=(int,np.integer), descr='Spatial position of slit/order for this fit. Required for I/O'),
+                 'pypeitfit': dict(otype=fitting.PypeItFit,
                                    descr='Fit to 1D wavelength solutions'),
                  'pixel_fit': dict(otype=np.ndarray, atype=np.floating,
                                    descr='Pixel values of arc lines'),
@@ -49,7 +52,12 @@ class WaveFit(datamodel.DataContainer):
 
     bitmask = defs.LinesBitMask()
 
-    def __init__(self, pypeitfit=None, pixel_fit=None, wave_fit=None, ion_bits=None,
+    @staticmethod
+    def hduext_prefix_from_spatid(spat_id):
+        """ Naming for HDU extensions"""
+        return 'SPAT_ID-{}_'.format(spat_id)
+
+    def __init__(self, spat_id, pypeitfit=None, pixel_fit=None, wave_fit=None, ion_bits=None,
                  cen_wave=None, cen_disp=None, spec=None, wave_soln=None,
                  sigrej=None, shift=None, tcent=None, rms=None, xnorm=None):
         # Parse
@@ -70,10 +78,14 @@ class WaveFit(datamodel.DataContainer):
             list:
 
         """
-        _d = super(WaveFit, self)._bundle(ext='WAVEFIT', **kwargs)
-        # Pull the fit out
-        if _d[0]['WAVEFIT']['pypeitfit'] is not None:
-            _d.append(dict(PYPEITFIT=_d[0]['WAVEFIT'].pop('pypeitfit')))
+        # Extension prefix (for being unique with slits)
+        hdue_pref = self.hduext_prefix_from_spatid(self.spat_id)
+        # Without PypeItFit
+        _d = super(WaveFit, self)._bundle(
+            ext=hdue_pref+'WAVEFIT', **kwargs)
+        # Deal with PypeItFit
+        if _d[0][hdue_pref+'WAVEFIT']['pypeitfit'] is not None:
+            _d.append({hdue_pref+'PYPEITFIT': _d[0][hdue_pref + 'WAVEFIT'].pop('pypeitfit')})
         # Return
         return _d
 
@@ -88,6 +100,22 @@ class WaveFit(datamodel.DataContainer):
         """
         return super(WaveFit, self).to_hdu(hdr=hdr, add_primary=add_primary, primary_hdr=primary_hdr,
                                            limit_hdus=limit_hdus, force_to_bintbl=True)
+
+    @classmethod
+    def from_hdu(cls, hdu, chk_version=True):
+        """
+        Parse the data from the provided HDU.
+
+        See :func:`pypeit.datamodel.DataContainer._parse` for the
+        argument descriptions.
+        """
+        # Set hdu_prefix
+        if isinstance(hdu, fits.HDUList):
+            hdu_prefix = cls.hduext_prefix_from_spatid(hdu[1].header['SPAT_ID'])
+        else:
+            hdu_prefix = cls.hduext_prefix_from_spatid(hdu.header['SPAT_ID'])
+        # Run the default parser to get the data
+        return super(WaveFit, cls).from_hdu(hdu, hdu_prefix=hdu_prefix)
 
[…]
