+++ pypeit/spectrographs/mmt_mmirs.py
@@ -33,14 +33,6 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
         self.telescope = telescopes.MMTTelescopePar()
         self.camera = 'MMIRS'
 
-    @staticmethod
-    def default_pypeit_par():
-        """
-        Set default parameters for the reductions.
-        """
-        par = pypeitpar.PypeItPar()
-        return par
-
     def init_meta(self):
         """
         Generate the meta data dict
@@ -112,6 +104,7 @@ class MMTMMIRSSpectrograph(spectrograph.Spectrograph):
             )
         return detector_container.DetectorContainer(**detector_dict)
 
+    @staticmethod
     def default_pypeit_par(self):
         """
         Set default parameters for the reductions.
