### 2020-08-17 1761784
Fixed flexure QA.
