### 2020-08-18 5916b9c
Fixed bitmask failures for wavelengths.
