### 2020-08-18 f5fdddc
add a wvcalib mask test
