### 2020-08-18 953dd36
Create environment.yml
