### 2020-08-19 cabfee5
long-slit readout mode
