### 2020-08-21 fed21f4
Dealing with new output format in cooked.
