### 2020-08-25 1b319ca
adding a new spectrograph
