### 2020-08-25 pypeit/PypeIt#994 @profxj
Any additional review @astrom-tom ?

We are pushing to a v1.1 and would love to include
modifications related to anything else you suggest.
Your input has been very valuable already.  Thanks.
