### 2020-08-25 cb507bd
more lamps + other meta
