+++ CHANGES.rst
@@ -43,6 +43,7 @@
  - Changed numpy (> 1.18.0) and scipy (> 1.4.0) version requirements
  - Correct det bug in keck_lris
  - Modifications to allow for flailing LRISr detector
+ - Modifications for parse LRIS LAMPS prior to 2010 upgrade
 
 1.0.6 (22 Jul 2020)
 -------------------
