+++ doc/A-B_differencing.rst
@@ -8,79 +8,40 @@ Overview
 PypeIt can reduce observations obtained using two or more nod
 positions in an alternating pattern. In this case, the user needs to edit
 the :ref:`pypeit_file` according to the desired reduction.
+The good news is that the code has substantial flexibility,
+the bad news is that setup is somewhat complex.
 
 PypeIt file
 ===========
 
-When running :ref:`pypeit_setup` for most near-IR spectrographs the 
+When running :ref:`pypeit_setup` for most near-IR spectrographs, the
 :ref:`pypeit_file:Data Block` in the PypeIt file will include three extra 
 parameters (``calib``, ``comb_id``, and ``bkg_id``), which should be edited 
-according to the desired reduction. This can be also obtained running :ref:`pypeit_setup`
-with `-b` option.
+according to the desired reduction. This can also
+be obtained running :ref:`pypeit_setup` by adding the `-b` option.
 
 
 
-Example
--------
+Parameters
+==========
 
-Here is an example of the :ref:`pypeit_file:Data Block` for the science data::
+The three additional columns (``calib``, ``comb_id``, and ``bkg_id``)
+have the following meanings/definitions:
 
-    |          filename |        frametype | ... | calib | comb_id | bkg_id |
-    | m190627_0001.fits | tilt,arc,science | ... |     1 |       1 |      2 |
-    | m190627_0002.fits | tilt,arc,science | ... |     2 |       2 |      1 |
-    | m190627_0003.fits | tilt,arc,science | ... |     3 |       3 |      4 |
-    | m190627_0004.fits | tilt,arc,science | ... |     4 |       4 |      3 |
-
-
-Editing the PypeIt file
------------------------
-
-The user need to edit ``calib``, ``comb_id``, and ``bkg_id`` in order to
-control how PypeIt subtracts, combines, and calibrates the spectroscopic data.
-
-``comb_id`` assigns a group ID to each frame, while ``bkg_id`` indicates the ID of the
-frame that is used as background image to be subtracted from the current frame. Both 
-``comb_id`` and ``bkg_id`` should be integer numbers.
-
-PypeIt always produces two frames (A-B, B-A) for each set of ``comb_id`` and ``bkg_id``.
-For the example above, PypeIt produces four 2D spectral images::
-
-    Science/spec2d-m190627_0001...fits      # m190627_0001.fits - m190627_0002.fits (A-B)
-    Science/spec2d-m190627_0002...fits      # m190627_0002.fits - m190627_0001.fits (B-A)
-    Science/spec2d-m190627_0003...fits      # m190627_0003.fits - m190627_0004.fits (A-B)
-    Science/spec2d-m190627_0004...fits      # m190627_0004.fits - m190627_0003.fits (B-A)
+* ``comb_id`` assigns a group ID to each frame
+* ``bkg_id`` indicates the ID of the frame that is used as background image to be subtracted from the current frame.
+* ``calib`` represents a calibration ID assigned to each science frame.
 
 
-If each frame has a unique ``comb_id`` (as the example above) the frames will not be coadded
-before the 1D extraction. Alternatively, frames with common values of ``comb_id`` will be
-coadded. In this case, common ``bkg_id`` should be used for all frames to be subtracted
-from frames with common ``comb_id``.
-
-Here is an example of the PypeIt file for combining frames::
-
-    |          filename |        frametype | ... | calib | comb_id | bkg_id |
-    | m190627_0001.fits | tilt,arc,science | ... |     1 |       1 |      2 |
-    | m190627_0002.fits | tilt,arc,science | ... |     2 |       2 |      1 |
-    | m190627_0003.fits | tilt,arc,science | ... |     1 |       1 |      2 |
[…]
