### 2020-08-26 pypeit/PypeIt#994 @astrom-tom
I think the documentation is fine now! thanks for implementing the comments!
