### 2020-08-27 eac832d
platescale for old detector
