+++ doc/api/pypeit.spectrographs.rst
@@ -23,8 +23,10 @@ Submodules
    pypeit.spectrographs.magellan_mage
    pypeit.spectrographs.mdm_osmos
    pypeit.spectrographs.mmt_binospec
+   pypeit.spectrographs.mmt_mmirs
    pypeit.spectrographs.not_alfosc
    pypeit.spectrographs.opticalmodel
+   pypeit.spectrographs.p200_dbsp
    pypeit.spectrographs.shane_kast
    pypeit.spectrographs.slitmask
    pypeit.spectrographs.spectrograph
