+++ environment.yml
@@ -10,7 +10,7 @@ dependencies:
   - scipy>=1.4
   - scikit-learn>=0.20
   - astropy>=4.0
-  - pyqt5>=5.14
+  - pyqt>=5.12
   - setuptools>=46.4
   - numba>=0.39
   - matplotlib>=3.1
