+++ doc/installing.rst
@@ -4,77 +4,83 @@ Installation
 
 This document describes how to install ``PypeIt`` for both users and developers.
 
+----
+
+.. _dependencies:
+
 Installing Dependencies
 =======================
 
-Installing pypeit will not automatically ensure that all the
+Installing ``PypeIt`` will not automatically ensure that all the
 dependencies (and their appropriate versions) are installed and
-up-to-date.  This must be done manually.
+up-to-date. Below we provide two ways of ensuring that the relevant
+dependencies are available.
 
-The required packages for pypeit are listed in the `requirements.txt
-file
-<https://github.com/pypeit/PypeIt/blob/master/pypeit/requirements.txt>`__.
-If you download the file, you can install/update all the dependencies
-by executing::
-
-    pip install -r requirements.txt
-
-However, we highly recommend that you use `Anaconda
-<https://www.anaconda.com/>`_, as described below. The
-package and version requirements for pypeit are:
+The package and version requirements for ``PypeIt`` are:
 
 * `python <http://www.python.org/>`_ version 3.7 or later
 * `numpy <http://www.numpy.org/>`_ version 1.18.0 or later
 * `astropy <http://www.astropy.org/>`_ version 4.0 or later
 * `scipy <http://www.scipy.org/>`_ version 1.4 or later
 * `matplotlib <http://matplotlib.org/>`_  version 3.1 or later
-* `numba <https://numba.pydata.org/>`_ version 0.39.0 or later (optional - some speed ups, inc. wavecal)
-* `PyQT5 <https://wiki.python.org/moin/PyQt/>`_ version 5 (needed for linetools)
+* `numba <https://numba.pydata.org/>`_ version 0.39.0 or later
+* `PyQT5 <https://wiki.python.org/moin/PyQt/>`_ version 5
 * `pyyaml <https://pyyaml.org/>`_ -- version 5.1
 * `configobj <https://pypi.org/project/configobj/>`_ -- version 5.0.6 or later
 * `scikit-learn <https://scikit-learn.org/stable/>`_ -- version 0.20 or later
 * `IPython <https://ipython.org>`_ -- version 7.2.0 or later
 * `ginga <https://pypi.org/project/ginga/>`_ -- version 3.0 or later
+* `requests <https://requests.readthedocs.io/en/master/>`_ -- version 2.23 or later
 * `packaging <https://pypi.org/project/packaging/>`_ -- version 19.0 or later
-* `linetools <https://pypi.org/project/linetools/>`_ -- version 0.2 or later (see also below)
+* `linetools <https://pypi.org/project/linetools/>`_ -- version 0.2 or later
 * `extension_helpers <https://pypi.org/project/extension-helpers/>`_ -- version 0.1 or later
 * `pytest <https://pypi.org/project/pytest/>`_ -- version 3.0.7 or later
 
-With Anaconda, you can check the presence of these packages with::
+Create a conda environment (recommended)
+----------------------------------------
 
-	conda list "^python$|numpy|astropy$|scipy$|matplotlib|numba|PyQT|ginga|pyyaml"
+We highly recommend using `Anaconda <https://www.anaconda.com/>`_ as
+a package and environment manager. We provide a yaml file that can be
+used to setup a conda environment called ``PypeIt``.  To use this:
 
-If the packages have been installed, this command should print out
-all the packages and their version numbers.
+ #. Download `environment.yml <https://github.com/pypeit/PypeIt/blob/master/environment.yml>`__.
 
-If any of the packages are out-of-date, they can be updated with a
-command like::
+ #. Create the conda environment::
 
-	conda update scipy
+        conda env create -f environment.yml
 
+ #. Activate it::
 
[…]
