+++ pypeit/requirements.txt
@@ -6,7 +6,7 @@ numpy>=1.18.0
 scipy>=1.4
 matplotlib>=3.1
 astropy>=4.0
-extension_helpers>=0.1
+extension-helpers>=0.1
 PyYAML>=5.1
 configobj>=5.0.6
 scikit-learn>=0.20
