### 2020-08-28 pypeit/PypeIt#993 @kbwestfall
Hi @astrom-tom .

For `extension-helpers`, I think this was a typo in the requirements file.  This is now fixed, although weirdly both `pip3 install extension_helpers` and `pip3 install extension-helpers` works for me.

I'm not sure what you mean about `find_packages`, though.  This method finds the packages within pypeit itself.  I.e.:
```
$ cd PypeIt
$ python3
Python 3.7.6 (v3.7.6:43364a7ae0, Dec 18 2019, 14:18:50)
[Clang 6.0 (clang-600.0.57)] on darwin
>>> from setuptools import find_packages
>>> find_packages()
['pypeit', 'pypeit.core', 'pypeit.images', 'pypeit.tests', 'pypeit.bspline', 'pypeit.scripts', 'pypeit.display', 'pypeit.spectrographs', 'pypeit.par', 'pypeit.core.gui', 'pypeit.core.wavecal']
```
 Do you instead mean the `install_requires` keyword for `setuptools.setup`?

We previously had the dependencies listed by the `install_requires` keyword.  See the `setup.py` file.  It's been a while, but I think we removed this because we didn't want users to unknowingly alter their current environment when they installed pypeit.  We instead moved the requirements file into the `pypeit` subdirectory and then use it to check that the dependencies are met; see `pypeit/check_requirements.py`.  It may seem round-about, but this way users know for sure that they're altering their packages when they install the dependencies instead of having `pip` or `setup.py` do it without their permission.  I'm interested to know what you think and if you have a specific recommendation for how to restructure this.
