### 2020-08-28 pypeit/PypeIt#995 @kbwestfall
Hi @astrom-tom .  Just checking in on this.  Can we close this part of the review, or do you have additional requests/comments?
