### 2020-08-28 pypeit/PypeIt#949 @profxj
Hi @Gabriel-p,

We've cleaned up the installation instructions.  Here are some direct responses to your comments/questions:

>> First you recommend to use conda, but give no one-line command to install all the dependencies (which is possible) or a yaml file with all the dependencies. Why is this? Could you provide one?

We've added the `environment.yml` file to the top-level directory and provided a description how this is used in the installation page.

>> The possibility of using pip install -r path/requirements.txt is mentioned after the instructions to install pypeit. Shouldn't this be an alternative to the conda installation? Why is it located here?

Yes, we've cleaned up the organization of the installation page to hopefully make it clear that these are two separate paths to installation of the dependencies.  See further discussion with the other reviewer in #993

>> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?  I guess this is covered by "Python based data reduction pipeline (DRP) written oringinally for echelle spectroscopy and since expanded to low-resolution spectrometers", but it could be expanded a bit with some more details. Also, notice the typo in "oringinally".

This is now described in the opening clause of the documentation:

```
``PypeIt`` is a Python package for semi-automated reduction of
astronomical, spectroscopic data. Its algorithms build on
decades-long development of previous data reduction pipelines by the
developers. The reduction procedure - including a complete list of
the input parameters and available functionality - is provided by
this online documentation.``PypeIt`` is a set of commands designed to perform the reduction
without any additional coding.This v1.1 release of ``PypeIt`` is designed to be used by both
advanced spectroscopists with prior data reduction expertise and
astronomers with no prior experience of data reduction. It is highly
configurable and designed to be applied to any standard slit-imaging
spectrograph, and can accommodate long-slit, multi-slit, as well as
cross-dispersed echelle spectra.
```

>> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>> The docs could be better organized. There are currently 54 entries at the main level, which makes navigating the docs somewhat of a hassle. Could this be better organized?

We have streamlined the menu options in the navigation bar.  We continue to improve the documentation and its readability guided by our expanded user base.  See #994 for further details.

>> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)
>> See point above. I can't tell if there is a clear example somewhere.

The front page of the new documentation provides a series of "Getting Started" links that point to examples for using the code.  In particular, see the PypeIt Cookbook and the "Step-by-step example" (https://github.com/pypeit/PypeIt/blob/joss-paper-branch/doc/step-by-step.rst).  Other sections below provide more detail on how to manipulate the control flow of the code.

>> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>> Is this covered anywhere?

Yes, see the "Contributors" section of the front page.  We've also
improved the text that points out that we have a user-based slack
workspace and provide a link for submitting a GitHub issue.
