+++ pypeit/core/wavecal/templates.py
@@ -4,6 +4,8 @@ import os
 import numpy as np
 from IPython import embed
 
+from matplotlib import pyplot as plt
+
 from pkg_resources import resource_filename
 from scipy.io import readsav
 
@@ -40,7 +42,7 @@ outpath = resource_filename('pypeit', 'data/arc_lines/reid_arxiv')
 def build_template(in_files, slits, wv_cuts, binspec, outroot, outdir=None,
                    normalize=False, subtract_conti=False, wvspec=None,
                    lowredux=True, ifiles=None, det_cut=None, chk=False,
-                   miny=None, overwrite=True):
+                   miny=None, overwrite=True, ascii_tbl=False, in_vac=True):
     """
     Generate a full_template for a given instrument
 
@@ -75,6 +77,10 @@ def build_template(in_files, slits, wv_cuts, binspec, outroot, outdir=None,
             snippet to have the same maximum amplitude.
         subtract_conti (bool, optional):
             Subtract the continuum for the final archive
+        ascii_tbl (bool, optional):
+            Table is a simple ASCII 2 column wave,flux table
+        in_vac (bool, optional):
+            True if input wavelengths are in vacuum
     """
     if outdir is None:
         outdir = outpath
@@ -94,6 +100,8 @@ def build_template(in_files, slits, wv_cuts, binspec, outroot, outdir=None,
             in_file = in_files[ifiles[kk]]
             if lowredux:
                 wv_vac, spec = xidl_arcspec(in_file, slit)
+            elif ascii_tbl:
+                wv_vac, spec = read_ascii(in_file, in_vac=in_vac)
             else:
                 wv_vac, spec = pypeit_arcspec(in_file, slit)
         else:
@@ -138,8 +146,11 @@ def build_template(in_files, slits, wv_cuts, binspec, outroot, outdir=None,
         nwspec = np.maximum(nwspec, miny)
     # Check
     if chk:
-#        debugger.plot1d(nwwv, nwspec)
-        embed(header='123')
+        plt.clf()
+        ax = plt.gca()
+        ax.plot(nwwv, nwspec)
+        plt.show()
+        embed(header='152')
     # Generate the table
     write_template(nwwv, nwspec, binspec, outdir, outroot, det_cut=det_cut, overwrite=overwrite)
 
@@ -290,7 +301,43 @@ def poly_val(coeff, x, nrm):
     return y
 
 
+def read_ascii(tbl_file, in_vac=True):
+    """
+
+    The columns need to be wave, flux
+    And the data should be monoonically increasing in wavelength
+
+    Args:
+        tbl_file:
+        in_vac:
+
+    Returns:
+
+    """
+    arc_spec = Table.read(tbl_file, format='ascii')
+
+    # Wavelengths
+    wv_vac = arc_spec['wave']
+    if not in_vac:
+        wv_vac = airtovac(wv_vac * units.AA)
+
+    # Return
+    return wv_vac.value, arc_spec['flux']
+
+
 def xidl_arcspec(xidl_file, slit):
+    """
+    Read an XIDL format solution
+
+    Note:  These are in air
+
+    Args:
+        xidl_file (str):
+        slit (int):
+
+    Returns:
+
+    """
     xidl_dict = readsav(xidl_file)
     if xidl_dict['archive_arc'].ndim == 2:
         nspec = xidl_dict['archive_arc'].shape[0]
@@ -402,7 +449,7 @@ def main(flg):
         build_template([wfile1,wfile2], slits, lcut, binspec, outroot, lowredux=False,
[…]
