### 2020-08-31 pypeit/PypeIt#949 @Gabriel-p
Thank you for the detailed answer @profxj. I'll resume the reviewing as soon as possible
