### 2020-08-31 e75bef6
trace add rm test
