### 2020-09-05 pypeit/PypeIt#1019 @Gabriel-p
Installation and testing issues
I'm following the installation instructions from the `develop` branch (https://github.com/pypeit/PypeIt/blob/develop/doc/installing.rst)

1. Created a `conda` environment using the `environment.yml` file [here](https://raw.githubusercontent.com/pypeit/PypeIt/develop/environment.yml) (notice that `h5py` is not present)
2. Then tried `pip install pypeit`, but it failed with:
```
$ pip install pypeit
Collecting pypeit
  Downloading pypeit-1.0.6.tar.gz (845 kB)
     |████████████████████████████████| 845 kB 2.4 MB/s
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
   command: /home/gabriel/miniconda3/envs/pypeit/bin/python3.7 /home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /tmp/tmpjouppvac
       cwd: /tmp/pip-install-q0uj36cv/pypeit
  Complete output (20 lines):
  Traceback (most recent call last):
    File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pip/_vendor/pep517/_in_process.py", line 280, in <module>
      main()
    File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pip/_vendor/pep517/_in_process.py", line 263, in main
      json_out['return_val'] = hook(**hook_input['kwargs'])
    File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pip/_vendor/pep517/_in_process.py", line 114, in get_requires_for_build_wheel
      return hook(config_settings)
    File "/tmp/pip-build-env-lj5u2clw/overlay/lib/python3.7/site-packages/setuptools/build_meta.py", line 150, in get_requires_for_build_wheel
      config_settings, requirements=['wheel'])
    File "/tmp/pip-build-env-lj5u2clw/overlay/lib/python3.7/site-packages/setuptools/build_meta.py", line 130, in _get_build_requires
      self.run_setup()
    File "/tmp/pip-build-env-lj5u2clw/overlay/lib/python3.7/site-packages/setuptools/build_meta.py", line 254, in run_setup
      self).run_setup(setup_script=setup_script)
    File "/tmp/pip-build-env-lj5u2clw/overlay/lib/python3.7/site-packages/setuptools/build_meta.py", line 145, in run_setup
      exec(compile(code, __file__, 'exec'), locals())
    File "setup.py", line 108, in <module>
      install_requires = get_requirements()
    File "setup.py", line 47, in get_requirements
      install_requires = [line.strip().replace('==', '>=') for line in open(requirements_file)
  FileNotFoundError: [Errno 2] No such file or directory: 'pypeit/requirements.txt'
ERROR: Command errored out with exit status 1: /home/gabriel/miniconda3/envs/pypeit/bin/python3.7 /home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /tmp/tmpjouppvac Check the logs for full command output.
```
3. Tried `pip install git+https://github.com/pypeit/PypeIt.git` and it worked
4. `$ pypeit_c_enabled` shows no message, so no C code
5. ginga Plugins checking fails:
```
$ pypeit_chk_plugins
zsh: command not found: pypeit_chk_plugins`
```
6. Test Your Installation fails with:
```
$ run_pypeit -h
Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/check_requirements.py", line 13, in <module>
    pv = pkg_resources.get_distribution(pkg).version
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 480, in get_distribution
    dist = get_provider(dist)
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 356, in get_provider
    return working_set.find(moduleOrReq) or require(str(moduleOrReq))[0]
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 899, in require
    needed = self.resolve(parse_requirements(requirements))
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 785, in resolve
    raise DistributionNotFound(req, requirers)
pkg_resources.DistributionNotFound: The 'h5py' distribution was not found and is required by the application

Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/pypeit/bin/run_pypeit", line 12, in <module>
    import pypeit.scripts.run_pypeit as run_pypeit
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/__init__.py", line 33, in <module>
    from pypeit import check_requirements  # THIS IMPORT DOES THE CHECKING.  KEEP IT
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/check_requirements.py", line 15, in <module>
    raise ImportError("Package: {:s} not installed!".format(pkg))
ImportError: Package: h5py not installed!
```
7. Installed with `conda install h5py`
8. Run test again:
```
$ run_pypeit -h
Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/check_requirements.py", line 13, in <module>
    pv = pkg_resources.get_distribution(pkg).version
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 480, in get_distribution
    dist = get_provider(dist)
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 356, in get_provider
    return working_set.find(moduleOrReq) or require(str(moduleOrReq))[0]
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 899, in require
    needed = self.resolve(parse_requirements(requirements))
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pkg_resources/__init__.py", line 785, in resolve
    raise DistributionNotFound(req, requirers)
pkg_resources.DistributionNotFound: The 'pytest' distribution was not found and is required by the application

Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/pypeit/bin/run_pypeit", line 12, in <module>
    import pypeit.scripts.run_pypeit as run_pypeit
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/__init__.py", line 33, in <module>
    from pypeit import check_requirements  # THIS IMPORT DOES THE CHECKING.  KEEP IT
  File "/home/gabriel/miniconda3/envs/pypeit/lib/python3.7/site-packages/pypeit/check_requirements.py", line 15, in <module>
    raise ImportError("Package: {:s} not installed!".format(pkg))
ImportError: Package: pytest not installed!
```
9. Installed with `conda install pytest` (notice that `pytest` is also not included in the `environment.yml` file)
10. `run_pypeit -h` is now successful
11. Run the PypeIt unit tests. I did not clone the repo so I'm not sure how to access the tests
