### 2020-09-06 pypeit/PypeIt#1019 @profxj
@Gabriel-p -- Thanks for continuing.  `h5py` is no longer in the list of requirements
on the `develop` branch.  I suspect you were running on `master`.  But `pytest` is required
and I have now added that to the `environment.yml` on the `joss-paper-branch`.

For sanity sake, I suggest we stay on `joss-paper-branch` for the remainder of this review.

As for tests, please see #995 by the other reviewer and see if that suffices.
