### 2020-09-06 pypeit/PypeIt#1019 @Gabriel-p
OK, I'll follow the instructions on the `joss-paper-branch` then. I'll start the installation from scratch. It is still not clear to me how to "*Run the PypeIt unit tests*". The instructions only address the case where the user cloned the repo. What if you installed without having to clone the repo? How do you proceed then?
