### 2020-09-06 pypeit/PypeIt#1019 @profxj
Can't run the tests if you don't clone.
