### 2020-09-06 pypeit/PypeIt#1019 @Gabriel-p
Ok, I'll clone the repo then. Probably should make this clear in the docs.
