### 2020-09-23 #2308 @profxj
Got it.  @harpolea -- @Gabriel-p and @astrom-tom kindly broke their
review into a series of separate Issues on our GitHub:

https://github.com/pypeit/PypeIt/issues/993
https://github.com/pypeit/PypeIt/issues/994
https://github.com/pypeit/PypeIt/issues/995
https://github.com/pypeit/PypeIt/issues/1019
