### 2020-09-23 pypeit/PypeIt#1019 @profxj
This Issue is related to JOSS review: https://github.com/openjournals/joss-reviews/issues/2308
