### 2020-09-23 pypeit/PypeIt#994 @profxj
This Issue is related to JOSS review: openjournals/joss-reviews#2308
