### 2020-09-23 #2308 @Gabriel-p
I'm almost done with the review. The biggest hurdle now is that I do not have any input data files to test the package's functionality. I'm afraid that even if I did, I do not have the resources to do it:

> ## Grab a buff computer
> We recommend one with at least 32Gb of RAM and this might not be enough for many-detector instruments (e.g. Keck DEIMOS).
>
> Multi-processors are good too, although only a portion of the code runs in parallel.

This prevents me from checking the **Functionality** point. @harpolea what should be done in this case?

@profxj  some minor issues with the article:

1. It mentions release 1.0 of the package, and v1.1 is already out.
2. There is I believe a missing reference (page 2):
    > PypeIt jointly performs the object extraction and b-spline @kelson2003 sky subtraction.

   Either that or I do not understand what that sentence means.

3. Perhaps a link to the "SPectral Image Typing (SPIT)" tool could be added?
4. Regarding the point "State of the field: Do the authors describe how this software compares to other commonly-used packages?", there are no other packages mentioned but perhaps this is because there are no similar packages available?
