### 2020-09-24 #2308 @profxj
@Gabriel-p -- Thanks for the suggestions to the paper.
I will make those edits.

As for a test run, here are a set of files for Shane Kastb
from our Development Suite:  https://drive.google.com/drive/folders/11ItdTKbY5c7Hfx-AnhSCvKcLW1g_mLAx?usp=sharing

These are small enough, 16Gb of RAM should be plenty.
