### 2020-09-24 #2308 @profxj
Ok, addressed except the last one.  For SPIT, we reference it.
I attach the currently generated PDF in case the reviewers
are unable to make it themselves.

[klidmv2olhyhnshtsbjw.pdf](https://github.com/openjournals/joss-reviews/files/5277260/klidmv2olhyhnshtsbjw.pdf)
