### 2020-09-24 #2308 @Gabriel-p
> 16Gb of RAM should be plenty

Still a no go (I have an old 8 Gb system, sorry). Perhaps @astrom-tom is able to perform a test run using these files?

> PypeIt jointly performs the object extraction and b-spline @kelson2003 sky subtraction

This is still in the PDF. Is it a missing reference or is 'b-spline @kelson2003 sky subtraction' a method? If so, please re-phrase or expand.
