### 2020-10-17 #2308 @Gabriel-p
I'm pretty much done with the review (could I have a link to the latest PDF? @profxj), except that I can not check the **Functionality** item as the requirements are too much for my system to handle. @harpolea not sure what to do in these cases.
