### 2020-10-30 #2308 @dfm
:wave: @astrom-tom - it looks like there are a few issues on https://github.com/pypeit/PypeIt waiting for responses from you. Can you give us an idea of when you'll get a chance to get to those? Let me know if you have any questions!
