### 2020-11-09 #2308 @dfm
@Gabriel-p: I hadn't seen your remaining issue before - sorry about that!

For the functionality point: I agree that it would be good for the introductory tutorials to be self-contained and executable with moderate resources. @profxj: do you think that it would be feasible to provide a sample data set that could be tested with fewer resources? If that's not possible, let me know and we can figure out what to do. Thanks!
