### 2020-11-09 #2308 @profxj
These Shane Kast blue data in the Drive below should work
with modest resources (<16Gb RAM).  But since I don't have
any such machine, I can't guarantee it.

https://drive.google.com/drive/folders/11ItdTKbY5c7Hfx-AnhSCvKcLW1g_mLAx?usp=sharing
