### 2020-11-09 #2308 @profxj
These are the data from the StepbyStep Example:

https://pypeit.readthedocs.io/en/latest/step-by-step.html
