### 2020-11-12 #2308 @Gabriel-p
I was able to successfully run the Step by step example with the files provided by @profxj on a system with 8 Gb of memory.

Small fix, in the **Spec2d** sections the commands:

    pypeit_show_2dspec Science/spec2d_b27-J1217p3905_KASTb_2015May20T045733.560.fits
    pypeit_show_1dspec Science/spec1d_b27-J1217p3905_KASTb_2015May20T045733.560.fits

should be written as `may` instead of `May`.

I'm thus checking the **Functionality** box, which completes my review.
