### 2020-11-16 #2308 @profxj
ps.  note that the `joss-paper-branch` has evolved considerably
since you last touched it..
