### 2020-11-16 #2308 @profxj
also, I just turned on that branch on `readthedocs`

should show up here soon:  https://pypeit.readthedocs.io/en/joss-paper-branch/

in case you want to see the bleeding edge.
