### 2020-11-16 pypeit/PypeIt#993 @astrom-tom
Ok, once again sorry for the delay. It seems that now the installation works without problem! I will close this ticket.
