### 2020-11-16 pypeit/PypeIt#995 @astrom-tom
Yes, I think this part can be closed.
I reran all the tests with the latest version (with python setup.py test):
================== 153 passed, 86 skipped, 204 warnings in 24.10s ==============

Just as a side note (I guess you are aware of it) I got some warnings:

_Most of them are related to FITS header keywords, an example:_
   VerifyWarning: Keyword name 'FLEX_SHIFT_TOTAL' is greater than 8 characters or contains characters not allowed by the FITS standard; a HIERARCH card will be created. (card.py:263)

_A few like this as well:_
pypeit/tests/test_wvcalib.py: 24 warnings
   DeprecationWarning: tostring() is deprecated. Use tobytes() instead. (base.py:1555)

_One like this_:
pypeit/tests/test_arc.py::test_detect_lines
   OptimizeWarning: Covariance of the parameters could not be estimated (minpack.py:828)

pypeit/sampling.py:131
   DeprecationWarning: invalid escape sequence \l (sampling.py:131)
