### 2020-11-16 pypeit/PypeIt#1082 @astrom-tom
JOSS review: Step-by-step example
Dear @profxj ,

I just went through the step by step example. It works very well!
I have a couple of suggestions about the step-by-step example page:

- While I could find the data in the google drive with the sentence ' Shane Kast*b* datasets in the Development Suite.' it might by useful to add the link [https://github.com/pypeit/PypeIt-development-suite#data-access](url) so the user has an easier access to the data. Something like:
 'This doc goes through a full run of PypeIt on one of the Shane Kast*b* datasets in the Development Suite. (see, [https://github.com/pypeit/PypeIt-development-suite#data-access](url) )'

- I don't know if this is an error, but running  (Main run section)
```
cd shane_kast_blue_A
run_pypeit shane_kast_blue_A.pypeit -o
```
did not work for me. It gave the error : `pypeit.pypmsgs.PypeItError: File does not exist: 600_4310_d55/b1.fits.gz`. I used the following setup function: `pypeit_setup -t 600_4310_d55/b -s shane_kast_blue -c A`. But then the *.pypeit configuration file contain the following path `path 600_4310_d55`. So if I enter the shane_kast_blue subdirectory, and use the `run_pypeit` command, the relative path is not correct.
I could make it work going next to the directory 600_4310_d55 and run
```
cd shane_kast_blue_A
run_pypeit shane_kast_blue_A/shane_kast_blue_A.pypeit -o
```
I believe you can just adjust the command in the documentation and this will be fine.

- Finally (and this is absolutely not mandatory for the paper), the 1D spec GUI appearing with `pypeit_show_1dspec` prints out some data in the terminal (ex: `button=1, x=809.000000, y=374.000000, xdata=5223.354768, ydata=387.372`). I believe it could be useful to display these printouts directly in the GUI (so all the informations are there).

These will be the last comments before I accept the paper (You can safely ignore the last one!).
