### 2020-11-16 #2308 @astrom-tom
Dear @profxj , @dfm

I created a last issue [https://github.com/pypeit/PypeIt/issues/1082](url) with a couple of minor points. All the rest is fine for me.
