### 2020-11-16 pypeit/PypeIt#1082 @dfm
Ref: https://github.com/openjournals/joss-reviews/issues/2308
