### 2020-11-17 pypeit/PypeIt#1082 @profxj
Thanks @astrom-tom

I have:

- Added the pointer to the data as you suggested
- Modified the notes to emphasize that RAW_DATA needs to be the *full* path
- And as for the GUI suggestion, we are using a 3rd party (`linetools`) for that.  ie. rather beyond our control.  So we are ignoring it.  :)
