### 2020-11-18 pypeit/PypeIt#1082 @astrom-tom
Awesome! That's all for me then! Thanks for implementing the comments!
