### 2020-11-18 #2308 @astrom-tom
Dear @dfm
All the comments have been implemented, so for me the paper is good to go!
(sorry again for the delay).
