### 2020-11-18 #2308 @dfm
First, @Gabriel-p, @astrom-tom: thanks a lot for your constructive reviews!!

Next, @profxj: Things are looking good here! Here are the steps I need you to do before we head to final processing:

1. Are you planning on merging the `joss-paper-branch` into `master`? That's not required, but it might be a bit confusing if there is an out of date copy of the manuscript on `master`. Let me know how you want to handle this!
2. Once you've decided on the above, bump the version number and make a tagged release. That release should be archived somewhere like Zenodo and **the title and author list of the archive must _exactly_ match the paper**. Then, report the new version number and the DOI for the code archive back to this thread.
3. Finally, comment `@whedon generate pdf from joss-paper-branch` (or just `@whedon generate pdf` depending on what you do for step 1 above) and double check that you're happy with the article proofs. In particular, please check that author names and affiliations because that can be harder to edit later.

After those steps I think we're set!
