### 2020-11-18 #2308 @profxj
Hi @dfm

1.  Yes.  I foresee that happening in ~2-3 weeks when we release v1.3
2.  I presume then I should wait until we release v1.3 for this.  Right?
