### 2020-11-18 #2308 @dfm
@profxj: Sorry - I forgot the word "branch" in the whedon command. It should be:

`@whedon generate pdf from branch joss-paper-branch`
