### 2020-11-18 #2308 @dfm
@profxj: Yes - the release should include all the changes associated with the review so getting those merged and included first would be good. So just ping me when you have that 1.3 release archived and I'll take it from there. Thanks!
