### 2020-12-15 #2308 @profxj
@dfm -- v1.3 is on `master` now.
Please proceed with publication.

Thanks!
