### 2020-12-15 #2308 @dfm
The metadata *must be the same as the manuscript*:

- The title should be `PypeIt: The Python Spectroscopic Data Reduction Pipeline`
- And the author list (names and order) must also match

You can edit this fields on Zenodo.
