### 2020-12-16 #2308 @kthyng
Hi @profxj, I am taking over to finish publishing your work. I read through the paper and I found a number of edits to make. I edited the paper and created a [PR](https://github.com/pypeit/PypeIt/pull/1111) you could consider merging for the corrections. Let me know what you decide.
