### 2020-12-16 #2308 @kthyng
Turns out there was another reference to the old version. Here is a new PR: https://github.com/pypeit/PypeIt/pull/1112
